### 2026-02-08 #282 @joshuabmoore
pyhctsa: A Python package for highly-comparative time-series analysis
Submitting Author: (@joshuabmoore)
All current maintainers: (@joshuabmoore, @benfulcher)
Package Name: pyhctsa
One-Line Description of Package: Python package for highly comparative time-series analysis
Repository Link: https://github.com/DynamicsAndNeuralSystems/pyhctsa
Version submitted: 0.1.2
Reviewer 1: N/A
Reviewer 2: N/A
Archive: 10.5281/zenodo.20820138
JOSS DOI: 10.21105/joss.10581
Version accepted: v1.0.0

---

- [x] I agree to abide by [pyOpenSci's Code of Conduct][PyOpenSciCodeOfConduct] during the review process and in future interactions in spaces supported by pyOpenSci should it be accepted.

As time-series data grows increasingly massive and complex, researchers rely on extracting "features", statistical summaries of data patterns, to distill meaningful patterns and insights from these datasets.`pyhctsa` is a native Python library designed to bridge the gap between advanced time-series analysis and modern open-source workflows. It ports the massive library of the original and highly successful MATLAB-based "Highly Comparative Time-Series Analysis" (HCTSA) framework into Python, enabling users to automatically extract and compare thousands of diverse time-series features from their data.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

The target audience of `pyhctsa` encompasses a broad spectrum of the global research community, ranging from data scientists and physicists to biologists and engineers. The impact of `pyhctsa` is expected to be as far-reaching as its predecessor, the original MATLAB `HCTSA` framework, for which scientific utility has been proven through a myriad of publications across diverse disciplines including neuroscience, medicine, geoscience, engineering, and others. By bringing the framework to the Python open-source ecosystem, the package will facilitate seamless integration with modern machine learning workflows and pipelines.

`tsfresh` - Limited set of time-series features, most of which comprise simple distributional, spectral and a few hand-selected information-theoretic measures.
`tsfel` - Limited in scope, with time-series feature set restricted to only a handful of basic distributional, spectral and only a handful of nonlinear analysis-based features.
`Kats` - Limited scope as far as breadth of time-series features. Package seems to not be maintained in the past 3-4 years. Many open issues which have not been resolved.

Compared to the above packages that rely on a narrow, a priori selection of descriptive statistics, `pyhctsa` provides a truly comprehensive, interdisciplinary library of over 4000 features, ranging from nonlinear dynamics and scaling laws, to information theory and model-based quantities. By casting a wider net across time-series analysis methods, `pyhctsa` shifts the focus from simple feature extraction to scientific discovery, enabling the identification of the dynamical "fingerprints" that underlie time-series data across diverse problems. As the Python successor to the MATLAB-based `HCTSA`, long considered the gold standard in highly comparative time-series analysis, `pychtsa` intends to uphold the same level of depth and rigour.

n/a

- [x] The package is deposited in a long-term repository with the DOI: https://doi.org/10.5281/zenodo.18529935
