+++ tutorials/custom.yaml
@@ -1,0 +1,121 @@
+medical:
+
+  hrv_classic:
+    base_name: hrv_classic
+    dependencies:
+    configs:
+      - {zscore: True}
+    legacy_name: MD_hrv_classic
+
+spectral:
+  spectral_summaries:
+    base_name: spectral_summaries
+    depedencies:
+    configs:
+      - {psd_meth: 'welch', window_type: 'rect', zscore: True}
+      - {psd_meth: 'fft', window_type: 'none', zscore: True}
+    legacy_name: SP_Summaries
+    ordered_args: ['psd_meth', 'window_type']
+
+correlation:
+
+  autocorr:
+    base_name: ac
+    dependencies:
+    configs:
+      - {tau: !range [1, 40], zscore: True}
+    legacy_name: CO_AutoCorr
+    ordered_args: ['tau']
+  
+  nonlinear_autocorr:
+    base_name: ac_nl
+    dependencies:
+    configs:
+      - {taus: [[1, 2]], absval: False, zscore: True}
+      - {taus: [[1, 2]], absval: True, zscore: True}
+      - {taus: [[1, 2, 3]], absval: False, zscore: True}
+      - {taus: [[1, 2, 3, 4]], absval: False, zscore: True}
+      - {taus: [[1, 2, 3, 4]], absval: True, zscore: True}
+      - {taus: [[1, 2, 3, 4, 5]], absval: False, zscore: True}
+      - {taus: [[1, 2, 3, 4, 5, 6]], absval: False, zscore: True}
+      - {taus: [[1, 2, 3, 4, 5, 6]], absval: True, zscore: True}
+      - {taus: [[1, 2, 3, 4, 5, 6, 7]], absval: False, zscore: True}
+      - {taus: [[1, 2, 3, 4, 5, 6, 7, 8]], absval: False, zscore: True}
+      - {taus: [[1, 2, 3, 4, 5, 6, 7, 8]], absval: True, zscore: True}
+
+      - {taus: [[1, 3]], absval: False, zscore: True}
+      - {taus: [[1, 3]], absval: True, zscore: True}
+      - {taus: [[2, 4]], absval: False, zscore: True}
+      - {taus: [[2, 4]], absval: True, zscore: True}
+
+      - {taus: [[1, 3, 5]], absval: False, zscore: True}
+      - {taus: [[2, 4, 6]], absval: False, zscore: True}
+
+      - {taus: [[1, 3, 5, 7]], absval: False, zscore: True}
+      - {taus: [[1, 3, 5, 7]], absval: True, zscore: True}
+
+      - {taus: [[2, 4, 6, 8]], absval: False, zscore: True}
+      - {taus: [[2, 4, 6, 8]], absval: True, zscore: True}
+      
+      - {taus: [[1, 2, 4]], absval: False, zscore: True}
+      - {taus: [[1, 3, 4]], absval: False, zscore: True}
+
+      - {taus: [[1, 4]], absval: False, zscore: True}
+      - {taus: [[1, 4]], absval: True, zscore: True}
+
+      - {taus: [[1, 1]], absval: False, zscore: True}
+      - {taus: [[1, 1]], absval: True, zscore: True}
+
+      - {taus: [[2, 2]], absval: False, zscore: True}
+      - {taus: [[2, 2]], absval: True, zscore: True}
+
+      - {taus: [[3, 3]], absval: False, zscore: True}
+      - {taus: [[3, 3]], absval: True, zscore: True}
+
+      - {taus: [[4, 4]], absval: False, zscore: True}
+      - {taus: [[4, 4]], absval: True, zscore: True}
+
+      - {taus: [[5, 5]], absval: False, zscore: True}
[…]
