### 2026-02-09 f2a0eeb
add zenodo badge
