### 2026-02-09 e684d70
add bonn eeg tutorial to readme
