### 2026-02-09 a7adec3
add bonn eeg tutorial to readme
