### 2026-02-09 3f940b4
add bonn eeg tutorial to readme
