### 2026-02-09 20cdbe1
add bonn eeg tutorial to readme
