### 2026-02-09 e3c99a2
add badges to readme
