### 2026-02-09 653ec1a
add badges to readme
