### 2026-02-09 c54c4c8
add badges to readme
