### 2026-02-09 9bb1713
add citation CITATION.cff file
