### 2026-02-11 #282 @crhea93
- [x] **Installation** The package can be installed from a community repository such as PyPI (preferred), and/or a community channel on conda (for example, conda-forge, bioconda).
- [x] **Documentation** The package has sufficient online documentation to allow us to evaluate the package's function and scope *without installing the package*. This includes:
  - [x] Short, quickstart tutorials that help a user understand how to use the package and what it can do for them.
  - [x] API documentation (documentation for your code's functions, classes, methods, and attributes): this includes clearly written docstrings with variables defined using a standard docstring format.
  - [x] **README** The package has a `README.md` file with a clear explanation of what the package does, instructions on how to install it, and a link to development instructions.
- [ ] The package authors have disclosed the use of Generative AI tools in the development and/or maintenance of their package.

**Optional but recommended:**

- [ ] **Linting** The package uses automated linting (e.g., [ruff](https://docs.astral.sh/ruff/)) to enforce code style and quality standards, ideally integrated with CI.
- [ ] **Type checking** The package includes type hints and uses a type checker (e.g., [mypy](https://mypy-lang.org/)) to catch type-related issues, ideally integrated with CI.
- [ ] **Modern tooling** The package uses modern Python development tools (e.g., [uv](https://docs.astral.sh/uv/) for dependency management) and [Hatch](https://www.pyopensci.org/python-package-guide/package-structure-code/python-package-build-tools.html#about-hatch) for building publishing and running development tasks to streamline development workflows. See the pyOpenSci packaging guide for more on [modern packaging tools](https://www.pyopensci.org/python-package-guide/package-structure-code/python-package-build-tools.html).

**Optional:** Let projects know that it's a great idea for projects to have a .github repository for the project organization where they can host a commonly used LICENSE, Code of Conduct, and even a YAML file with label definitions. These items will then be automatically applied to every repository in the organization to ensure consistency (but can be customized within repos too). The [SunPy project](https://github.com/sunpy/.github/) has a great example of this.

---
A big thank you in advance to authors for setting aside five to ten minutes to do this. It truly helps our organization! :raised_hands:
---

Thank you for this submission. I've broken my comments into two main categories below:

### Major
- While the tutorial notebooks are very useful, this package is missing a well structured and detailed documentation. I suggest that the authors use a tool like [sphinx](https://www.sphinx-doc.org/en/master/). Please note that we require API documentation as well. The exact requirements can be found earlier in this comment.

- PyOpenSci requires a robust AI disclosure section in the README file. Although this is a constantly changing policy as we deal with an ever-changing agentic LLM landscape, we have crafted this policy outline https://www.pyopensci.org/blog/generative-ai-peer-review-policy.html

### Minor
- The automated tests are very useful. However, it would be very helpful to have a coverage report so that we can ensure enough code is being covered by the tests.

I’ve observed that the majority of the codebase was initially developed in September, with a number of more recent, sporadic contributions over the past month. This pattern may be worth noting in the project documentation or release notes, as it could provide useful context for contributors or users regarding the project’s development timeline and current focus areas.
**Question for Clarification:**
Could you share any background on why the development activity was concentrated in September, followed by a period of less frequent updates? Understanding the project’s development rhythm can help ensure that the documentation and community expectations are aligned with the actual workflow.
