### 2026-02-14 63debf9
paper draft V2.0
