+++ paper/paper.md
@@ -9,10 +9,10 @@ tags:
 authors:
   - name: Joshua B. Moore
     orcid: 0000-0002-5237-1087
-    equal-contrib: true
+    equal-contrib: false
     affiliation: 1
   - name: Ben D. Fulcher
-    equal-contrib: true
+    equal-contrib: false
     orcid: 0000-0002-3003-4055
     affiliation: 1
 affiliations:
@@ -24,66 +24,56 @@ bibliography: paper.bib
 ---
 
 # Summary
-\begin{center}
-\includegraphics[width=0.25\linewidth]{pyhctsa_logo-02.pdf}
-\end{center}
-
-Across diverse fields of science, finance, and industry, the generation of time-series data has reached an unprecedented scale. To navigate this abundance, there is an increasing need for tools and methods that distill complex, high-dimensional datasets down into interpretable summaries or `features' that capture the varied statistical and dynamical properties underlying the time-series data. From uncovering empirical structure in large datasets [@Fulcher:2013] to seizure and stellar light curve classification [@Barbara:2022], feature-based approaches play an integral role in tasks that would otherwise struggle with raw time-series values alone.
-
-Drawing on a vast interdisciplinary literature of quantitative methods, highly comparative time-series analysis (HCTSA) provides a unified framework for automatically computing and comparing thousands of time-series features. Although `HCTSA` has supported over XXX publications, its original MATLAB-based implementation presents significant accessibility and scalability barriers for many researchers. `pyhctsa` addresses these limitations by offering a native Python implementation, bringing the analytical power of the `HCTSA` library to the open-source ecosystem. By integrating with the modern Python data stack, `pyhctsa` will enable a broader community of researchers to perform large-scale, automated feature extraction in a transparent and reproducible environment. 
+<!-- Check validity of supporting > 500 publications claim - based on google scholar citations -->
+Across diverse fields of science, finance, and industry, the generation of time-series data has reached an unprecedented scale. Accordingly, there is an increasing need for tools and methods to distill complex, high-dimensional datasets down into interpretable summaries or "features" that capture the varied statistical and dynamical properties underlying the time-series data. From uncovering empirical structure in large datasets [@Fulcher:2013] to seizure and stellar light curve classification [@Barbara:2022], feature-based approaches facilitate a wide range of statistical learning tasks involving time series, wherein the features are derived from interpretable theory and can thus yield insights that can guide understanding and help motivate further experiments. Among these approaches, the MATLAB-based `hctsa` package (which stands for "highly comparative time-series analysis") is the most comprehensive, containing implementations of over 7000 time-series features that can quantify a wide range of distributional and correlation-based statistical properties of time series. Although `hctsa` has supported over 500 publications, its MATLAB-based implementation has presented accessibility issues that have prevented its incorporation into analysis pipelines for open science and across industry applications. Our software package `pyhctsa` addresses these limitations by implementing the majority of the `hctsa` feature library in native Python, bringing a uniquely comprehensive resource of scientific time-series analysis methods to the free and open-source software (FOSS) ecosystem for the first time. The design philosophy of `pyhctsa` is centred around extensibility, with scope to become a living library of time-series analysis methods through further user contributions. Thus far, `pyhctsa` has implemented over 4500 time-series features, making it the most comprehensive Python package for time-series feature extraction by both total volume of features and conceptual coverage.
 
 
 # Statement of need
[…]
