+++ pyhctsa/calculator.py
@@ -21,22 +21,22 @@ def range_constructor(loader, node) -> list:
 yaml.SafeLoader.add_constructor("!range", range_constructor)
 
 def classify_output(res) -> int:
-    # classify the type of output
+    """classify the type of output"""
+    out = 0
     if isinstance(res, str) and res.startswith("Error:"):
-        return 1
+        out = 1
     elif res is None:
-        return 6
+        out = 6
     elif np.iscomplexobj(res):
         # non-zero imaginary component
-        return 5
+        out = 5
     elif np.isnan(res):
-        return 2
+        out = 2
     elif np.isposinf(res):
-        return 3
+        out = 3
     elif np.isneginf(res):
-        return 4
-    else:
-        return 0
+        out = 4
+    return out
 
 def apply_selection_wrapper(func, keep_keys):
     """
@@ -47,8 +47,6 @@ def apply_selection_wrapper(func, keep_keys):
     :return: Wrapped function.
     :rtype: Any
     """
-    keys_list = [keep_keys] if isinstance(keep_keys, str) else keep_keys
-    
     def wrapper(*args, **kwargs):
         result = func(*args, **kwargs)
         # if result is a dict, then filter it according to the keys
@@ -65,7 +63,8 @@ def _standardise_inputs(data) -> list[np.ndarray]:
         elif data.ndim == 2:
             if data.shape[0] > data.shape[1]:
                 # notify the user to check that the shapes make sense
-                print(f"Check that the shape of the 2D input is such that (n_series, n_samples). Got shape: {data.shape}")
+                print(f"Check that the shape of the 2D input is such "
+                      f"that (n_series, n_samples). Got shape: {data.shape}")
             return [np.asarray(row, dtype=float) for row in data]
         else:
             raise ValueError("NumPy array must be 1D or 2D.")
@@ -84,8 +83,8 @@ def _standardise_inputs(data) -> list[np.ndarray]:
         raise ValueError(
         "Input must be a 1D series, a list of 1D series, or a 2D array "
         "with shape (n_series, n_samples)")
-      
-def _format_param_value(val, key=None) -> str: 
+   
+def _format_param_value(val, key=None) -> str:
     """ 
     Format parameter value for label: 
     - For bools: if True, return the key name.
@@ -96,51 +95,73 @@ def _format_param_value(val, key=None) -> str:
     if isinstance(val, bool):
         return key if val and key else ""
 
-    if isinstance(val, list): 
-        # Check if it's a contiguous range 
-        if len(val) > 1 and all(isinstance(x, (int, float)) for x in val): 
-            diffs = [val[i+1] - val[i] for i in range(len(val)-1)] 
-            if all(d == 1 for d in diffs): 
+    if isinstance(val, list):
+        # Check if it's a contiguous range
+        if len(val) > 1 and all(isinstance(x, (int, float)) for x in val):
+            diffs = [val[i+1] - val[i] for i in range(len(val)-1)]
+            if all(d == 1 for d in diffs):
                 # Pass key down for recursion if needed, though usually lists aren't bools
-                return f"{_format_param_value(val[0])}_{_format_param_value(val[-1])}" 
-        return "_".join(_format_param_value(x) for x in val) 
+                return f"{_format_param_value(val[0])}_{_format_param_value(val[-1])}"
+        return "_".join(_format_param_value(x) for x in val)
 
-    if isinstance(val, (float, int)): 
-        if val < 0: 
-            return 'm' + _format_param_value(-val) 
-        elif val == int(val): 
-            return str(int(val)) 
-        elif 0 < val < 1: 
-            return '0p' + str(val).split(".")[1].rstrip('0') 
-        else: 
[…]
