+++ pyhctsa/operations/hypothesis_tests.py
@@ -8,7 +8,8 @@ from statsmodels.sandbox.stats.runs import runstest_1samp
 from statsmodels.stats.descriptivestats import sign_test
 from statsmodels.stats.diagnostic import acorr_ljungbox
 
-def variance_ratio_test(y : ArrayLike, periods : Union[int, list[int], float] = 2, iids : Union[int, list[int]] = 0) -> dict:
+def variance_ratio_test(y : ArrayLike, periods : Union[int, list[int], float] = 2,
+                        iids : Union[int, list[int]] = 0) -> dict:
     """
     Variance ratio test for random walk.
 
@@ -34,14 +35,16 @@ def variance_ratio_test(y : ArrayLike, periods : Union[int, list[int], float] = 
     """
     y = np.asarray(y)
     out = {}
-    logical_check = lambda lst: all(x in (0, 1) for x in lst)
+    def logical_check(lst):
+        return all(x in (0, 1) for x in lst)
     if isinstance(periods, list):
         # Return statistics on multiple outputs for multiple periods/IIDs
         # check that IIDS is also a list
         if isinstance(iids, list):
             # some checks on the data types...
             if len(iids) != len(periods):
-                raise ValueError(f"Length of IIDs list ({len(iids)}) does not match the list of periods ({len(periods)}).")
+                raise ValueError(f"Length of IIDs list ({len(iids)}) does not match "
+                    "the list of periods ({len(periods)}).")
             if not logical_check(iids):
                 raise ValueError("List of IIDs must only be logicals (0 or 1).")
 
@@ -67,7 +70,8 @@ def variance_ratio_test(y : ArrayLike, periods : Union[int, list[int], float] = 
             out['maxstat'] = np.max(all_stats)
             out['minstat'] = np.min(all_stats)
         else:
-            raise ValueError(f"Expected iids to be a list of bools, since periods are also a list. Got data type: {type(iids)} instead.")
+            raise ValueError(f"Expected iids to be a list of bools, since periods "
+                             f"are also a list. Got data type: {type(iids)} instead.")
     elif isinstance(periods, (int, float, np.number)):
         periods = int(periods)
         robust = True if iids == 0 else False 
@@ -77,7 +81,8 @@ def variance_ratio_test(y : ArrayLike, periods : Union[int, list[int], float] = 
         out['stat'] = vr.stat
         out['ratio'] = vr.vr
     else:
-        raise ValueError(f"Unknown data type for periods: {type(periods)}, select either integer or list of integers.")
+        raise ValueError(f"Unknown data type for periods: {type(periods)}, "
+                         "select either integer or list of integers.")
 
     return out
 
@@ -127,10 +132,9 @@ def hypothesis_test(x : ArrayLike, the_test: str = "signtest") -> float:
         _, p = wilcoxon(x)
     elif the_test == "lbq":
         # Ljung-Box Q-test for residual autocorrelation
-        T = np.sum(~np.isnan(x)) # get the effective sample size
-        n_lags = min(20, T-1)
+        t = np.sum(~np.isnan(x)) # get the effective sample size
+        n_lags = min(20, t-1)
         p = acorr_ljungbox(x, lags=[n_lags])['lb_pvalue'].to_numpy()[0]
     else:
         raise ValueError(f"Unknown test: {the_test}.")
-    
     return p
