+++ pyhctsa/utils.py
@@ -173,12 +173,14 @@ def z_score(x : ArrayLike) -> np.ndarray:
 
     # Check for NaNs, infs, etc. i.e., check data is finite
     if not np.isfinite(x).all():
-        raise ValueError(f'data contains non-finite values (NaN/inf) at idxs: {np.argwhere(np.isfinite(x) == False)}')
+        raise ValueError(f"data contains non-finite values (NaN/inf) at "
+                         f"idxs: {np.argwhere(np.isfinite(x) == False)}")
     
     # robust checks for constant values
     var_x = np.var(x, ddof=1)
     if var_x < 1e-10:
-        raise ValueError(f"Data has sample variance {var_x:.2e} < 1e-10. Values appear to be constant.")
+        raise ValueError(f"Data has sample variance {var_x:.2e} < 1e-10. "
+                         "Values appear to be constant.")
     data_range = np.ptp(x)  # peak-to-peak (max - min)
     if data_range < 1e-10:
         raise ValueError(f"Data range {data_range:.2e} < 1e-10. Values appear to be constant.")
@@ -191,13 +193,15 @@ def z_score(x : ArrayLike) -> np.ndarray:
 
 def histc(x : ArrayLike, bins : ArrayLike) -> int:
     """Counts the number of values in x that are within each specified bin."""
-    map_to_bins = np.digitize(x, bins) # Get indices of the bins to which each value in input array belongs.
+    # Get indices of the bins to which each value in input array belongs.
+    map_to_bins = np.digitize(x, bins)
     res = np.zeros(bins.shape)
     for el in map_to_bins:
         res[el-1] += 1 # Increment appropriate bin.
     return res
 
-def bin_picker(x_min : float, x_max : float, n_bins : Union[None, int], bin_width_est : Union[None, float] = None) -> np.ndarray:
+def bin_picker(x_min : float, x_max : float, n_bins : Union[None, int],
+               bin_width_est : Union[None, float] = None) -> np.ndarray:
     """
     Choose histogram bins. 
 
@@ -217,10 +221,10 @@ def bin_picker(x_min : float, x_max : float, n_bins : Union[None, int], bin_widt
     edges : numpy.ndarray
         Array of bin edges.
     """
-    if bin_width_est == None:
-        rawBinWidth = abs(x_max - x_min)/n_bins
+    if bin_width_est is None:
+        raw_bin_width = abs(x_max - x_min)/n_bins
     else:
-        rawBinWidth = bin_width_est
+        raw_bin_width = bin_width_est
 
     if x_min is not None:
         if not np.issubdtype(type(x_min), np.floating):
@@ -230,13 +234,13 @@ def bin_picker(x_min : float, x_max : float, n_bins : Union[None, int], bin_widt
         xrange = x_max - x_min
 
         # Make sure the bin width is not effectively zero
-        rawBinWidth = max(rawBinWidth, np.spacing(xscale))
+        raw_bin_width = max(raw_bin_width, np.spacing(xscale))
 
         # If the data are not constant, place the bins at "nice" locations
         if xrange > max(np.sqrt(np.spacing(xscale)), np.finfo(xscale).tiny):
             # Choose the bin width as a "nice" value
-            pow_of_ten = 10 ** np.floor(np.log10(rawBinWidth))
-            rel_size = rawBinWidth / pow_of_ten  # guaranteed in [1, 10)
+            pow_of_ten = 10 ** np.floor(np.log10(raw_bin_width))
+            rel_size = raw_bin_width / pow_of_ten  # guaranteed in [1, 10)
 
             # Automatic rule specified
[…]
