+++ pyhctsa/operations/correlation.py
@@ -13,20 +13,23 @@ from ..operations.information import automutual_info, first_min
 from ..toolboxes.c22 import periodicity_wang_wrapper
 from ..utils import bin_picker, make_mat_buffer, point_of_crossing, sign_change, z_score, histc
 
-def add_noise(y : ArrayLike, tau : Union[int, str] = 1, ami_method : str = 'even', 
+def add_noise(y : ArrayLike, tau : Union[int, str] = 1, ami_method : str = 'even',
               extra_param : Union[int, None] = None, random_seed = None) -> dict:
     """
     Changes in the automutual information with the addition of noise.
 
-    Adds Gaussian-distributed noise to the time series with increasing standard deviation, eta, across the range eta = 0, 0.1, ..., 2, and measures the
-    mutual information at each point. Can be measured using histograms with extra_param bins or using the Information Dynamics Toolkit.
-    The output is a set of statistics on the resulting set of automutual information estimates, including a fit to an exponential decay, since the
-    automutual information decreases with the added white noise.
-    This algorithm is quite different, but was based on the idea in [1].
+    Adds Gaussian-distributed noise to the time series with increasing standard deviation, eta, 
+    across the range eta = 0, 0.1, ..., 2, and measures the mutual information at each point. 
+    Can be measured using histograms with extra_param bins or using the Information Dynamics 
+    Toolkit. The output is a set of statistics on the resulting set of automutual information
+    estimates, including a fit to an exponential decay, since the automutual information 
+    decreases with the added white noise. This algorithm is quite different, but was based 
+    on the idea in [1].
 
     References
     ----------
-    .. [1] "Titration of chaos with added noise", Chi-Sang Poon and Mauricio Barahona P. Natl. Acad. Sci. USA, 98(13) 7107 (2001)
+    .. [1] "Titration of chaos with added noise", Chi-Sang Poon and Mauricio Barahona 
+        P. Natl. Acad. Sci. USA, 98(13) 7107 (2001)
 
     Parameters
     -----------
@@ -60,7 +63,6 @@ def add_noise(y : ArrayLike, tau : Union[int, str] = 1, ami_method : str = 'even
     if extra_param is None:
         # JIDT expects empty string for no extra params
         extra_param = ''
-    
     # Generate noise
     if random_seed is not None:
         np.random.seed(random_seed)
@@ -85,7 +87,6 @@ def add_noise(y : ArrayLike, tau : Union[int, str] = 1, ami_method : str = 'even
             amis[i] = automutual_info(y + noise_range[i]*noise, tau, ami_method, str(extra_param))
             if np.isnan(amis[i]):
                 raise ValueError('Error computing AMI: Time series too short (?)')
-    
     # Output statistics
     out = {}
     # Proportion decreases
@@ -111,7 +112,7 @@ def add_noise(y : ArrayLike, tau : Union[int, str] = 1, ami_method : str = 'even
     # Count number of times the AMI function crosses its mean
     out['pcrossmean'] = np.sum(np.diff(np.sign(amis - np.mean(amis))) != 0) / (num_repeats - 1)
 
-    # Fit exponential decay model 
+    # Fit exponential decay model
     exp_func = lambda x, a, b : a * np.exp(b * x)
     popt, pcov = curve_fit(exp_func, noise_range, amis, p0=[amis[0], -1])
     out['fitexpa'], out['fitexpb'] = popt
[…]
