+++ pyhctsa/operations/criticality.py
@@ -1,3 +1,27 @@
+r"""
+          _____                    _____                    _____          
+         /\    \                  /\    \                  /\    \         
+        /::\    \                /::\    \                /::\    \        
+       /::::\    \              /::::\    \              /::::\    \       
+      /::::::\    \            /::::::\    \            /::::::\    \      
+     /:::/\:::\    \          /:::/\:::\    \          /:::/\:::\    \     
+    /:::/__\:::\    \        /:::/__\:::\    \        /:::/  \:::\    \    
+   /::::\   \:::\    \      /::::\   \:::\    \      /:::/    \:::\    \   
+  /::::::\   \:::\    \    /::::::\   \:::\    \    /:::/    / \:::\    \  
+ /:::/\:::\   \:::\____\  /:::/\:::\   \:::\    \  /:::/    /   \:::\ ___\ 
+/:::/  \:::\   \:::|    |/:::/  \:::\   \:::\____\/:::/____/     \:::|    |
+\::/   |::::\  /:::|____|\::/    \:::\  /:::/    /\:::\    \     /:::|____|
+ \/____|:::::\/:::/    /  \/____/ \:::\/:::/    /  \:::\    \   /:::/    / 
+       |:::::::::/    /            \::::::/    /    \:::\    \ /:::/    /  
+       |::|\::::/    /              \::::/    /      \:::\    /:::/    /   
+       |::| \::/____/               /:::/    /        \:::\  /:::/    /    
+       |::|  ~|                    /:::/    /          \:::\/:::/    /     
+       |::|   |                   /:::/    /            \::::::/    /      
+       \::|   |                  /:::/    /              \::::/    /       
+        \:|   |                  \::/    /                \::/____/        
+         \|___|                   \/____/                  ~~              
+                                                                           
+"""
 from typing import Union
 
 import numpy as np
