+++ pyhctsa/operations/distribution.py
@@ -35,7 +35,6 @@ def compare_ks_fit(x: ArrayLike, what_distn: str) -> dict:
         Includes the absolute area between the two distributions, the peak separation,
         overlap integral, and relative entropy.
     """
-    
     x = np.asarray(x)
     x_step = np.std(x, ddof=1) / 100  # set a step size
 
@@ -152,7 +151,7 @@ def compare_ks_fit(x: ArrayLike, what_distn: str) -> dict:
 
     return out
 
-def _find_bounds(pdf_func, start_left, start_right, xStep, thresh):
+def _find_bounds(pdf_func, start_left, start_right, x_step, thresh):
     """Expand left/right until pdf falls below threshold."""
     xf = [start_left, start_right]
 
@@ -160,13 +159,13 @@ def _find_bounds(pdf_func, start_left, start_right, xStep, thresh):
     if start_left is not None:
         ange = 10
         while ange > thresh:
-            xf[0] -= xStep
+            xf[0] -= x_step
             ange = pdf_func(xf[0])
 
     # Right search
     ange = 10
     while ange > thresh:
-        xf[1] += xStep
+        xf[1] += x_step
         ange = pdf_func(xf[1])
 
     return xf
@@ -182,7 +181,8 @@ def withinp(x : ArrayLike, p : float = 1.0, mean_or_median : str = 'mean') -> fl
     p : float
         The number (proportion) of standard deviations
     mean_or_median : str 
-        Whether to use units of 'mean' and standard deviation, or 'median' and rescaled interquartile range.
+        Whether to use units of 'mean' and standard deviation, or 'median' 
+        and rescaled interquartile range.
 
     Returns
     --------
@@ -581,11 +581,11 @@ def burstiness(y: ArrayLike) -> dict:
         'B_Kim': Improved burstiness for finite series
     """
     y = np.asarray(y)
-    mean = np.mean(y)
+    me = np.mean(y)
     std = np.std(y, ddof=1)
 
-    r = np.divide(std,mean) # coefficient of variation
-    B = np.divide((r - 1), (r + 1)) # Original Goh and Barabasi burstiness statistic, B
+    r = np.divide(std,me) # coefficient of variation
+    b = np.divide((r - 1), (r + 1)) # Original Goh and Barabasi burstiness statistic, B
 
     # improved burstiness statistic, accounting for scaling for finite time series
     # Kim and Jo, 2016, http://arxiv.org/pdf/1604.01125v1.pdf
@@ -593,9 +593,9 @@ def burstiness(y: ArrayLike) -> dict:
     p1 = np.sqrt(N+1)*r - np.sqrt(N-1)
     p2 = (np.sqrt(N+1)-2)*r + np.sqrt(N-1)
 
-    B_Kim = np.divide(p1, p2)
+    b_kim = np.divide(p1, p2)
 
-    out = {'B': B, 'B_Kim': B_Kim}
+    out = {'B': b, 'B_Kim': b_kim}
 
     return out
 
@@ -624,9 +624,12 @@ def outlier_include(y: ArrayLike, threshold_how: str = 'abs', inc: float = 0.01)
     """
     How statistics depend on distributional outliers.
 
-    Measures how various statistics of a time series change as more and more outliers are included in the calculation, according to a specified rule for defining outliers.
+    Measures how various statistics of a time series change as more and more outliers 
+    are included in the calculation, according to a specified rule for defining outliers.
 
[…]
