### 2026-02-14 a6e1aab
linting entropy
