+++ pyhctsa/operations/entropy.py
@@ -222,8 +222,8 @@ def multi_scale_entropy(
     m = int(m)
     if scale_range is None:
         scale_range = range(1, 10)
-    minTsLength = 20
-    numScales = len(scale_range)
+    min_ts_length = 20
+    num_scales = len(scale_range)
 
     if pre_process_how is not None:
         if pre_process_how == 'diff1':
@@ -238,15 +238,15 @@ def multi_scale_entropy(
     
     # Coarse-graining across scales
     y_cg = []
-    for i in range(numScales):
+    for i in range(num_scales):
         buffer_size = scale_range[i]
         y_buffer = make_buffer(y, buffer_size)
         y_cg.append(np.mean(y_buffer, 1))
     
     # Run sample entropy for each m and r value at each scale
-    samp_ens = np.zeros(numScales)
-    for si in range(numScales):
-        if len(y_cg[si]) >= minTsLength:
+    samp_ens = np.zeros(num_scales)
+    for si in range(num_scales):
+        if len(y_cg[si]) >= min_ts_length:
             samp_en_struct = sample_entropy(y_cg[si], m, r)
             samp_ens[si] = samp_en_struct[f'sampen{m}']
         else:
@@ -263,7 +263,7 @@ def multi_scale_entropy(
         return {'out': np.nan}
 
     # Output raw values
-    out = {f'sampen_s{scale_range[i]}': samp_ens[i] for i in range(numScales)}
+    out = {f'sampen_s{scale_range[i]}': samp_ens[i] for i in range(num_scales)}
 
      # Summary statistics of the variation
     max_samp_en = np.nanmax(samp_ens)
@@ -284,7 +284,7 @@ def multi_scale_entropy(
 
     return out
 
-def sample_entropy(y: ArrayLike, m: int = 2, r: Optional[float] = None, 
+def sample_entropy(y: ArrayLike, m: int = 2, r: Optional[float] = None,
                     pre_process_how: Optional[str] = None) -> dict:
     """
     Compute Sample Entropy (SampEn) of a time series.
@@ -296,7 +296,8 @@ def sample_entropy(y: ArrayLike, m: int = 2, r: Optional[float] = None,
 
     References
     ----------
-    .. [1] "Sample Entropy Estimation 1.0.0", https://physionet.org/content/sampen/1.0.0/c/sampen-1.1.c
+    .. [1] "Sample Entropy Estimation 1.0.0", 
+        https://physionet.org/content/sampen/1.0.0/c/sampen-1.1.c
     .. [2] "Control Entropy: A complexity measure for nonstationary signals"
         E. M. Bollt and J. Skufca, Math. Biosci. Eng., 6(1) 1 (2009).
 
@@ -328,7 +329,7 @@ def sample_entropy(y: ArrayLike, m: int = 2, r: Optional[float] = None,
         y = np.diff(y)
 
     samp_en = _sampen_c.calculate(y, m+1, r)
-    samp_en = samp_en[:-1] # always that extra one for the M = 0 
+    samp_en = samp_en[:-1] # always that extra one for the M = 0
     out = {}
     for mi in range(m + 1):
         out[f'sampen{mi}'] = samp_en[mi]
@@ -346,11 +347,13 @@ def permutation_entropy(y: ArrayLike, m: int = 2, tau: int = 1) -> dict:
     as described in [1]. 
 
     This implementation modifies code from the antropy package:
-    https://github.com/raphaelvallat/antropy to provide both raw and normalised permutation entropy values.
+    https://github.com/raphaelvallat/antropy to provide both raw and 
+    normalised permutation entropy values.
 
     References
     ----------
[…]
