### 2026-02-14 2730dce
linting for spectral + robust fit features
