### 2026-02-14 4d9ff17
stationarity linting
