+++ pyhctsa/operations/stationarity.py
@@ -14,7 +14,8 @@ from ..operations.distribution import moments
 from ..operations.entropy import approximate_entropy, distribution_entropy, sample_entropy
 from ..utils import make_mat_buffer, sign_change, z_score
 
-def local_distributions(y : ArrayLike, num_segs : int = 5, each_or_par : str = 'par', num_points : int = 200) -> dict:
+def local_distributions(y : ArrayLike, num_segs : int = 5, each_or_par : str = 'par',
+                        num_points : int = 200) -> dict:
     """
     Compares the distribution in consecutive time-series segments.
 
@@ -39,12 +40,13 @@ def local_distributions(y : ArrayLike, num_segs : int = 5, each_or_par : str = '
     dict
         Measures of the sum of absolute deviations between distributions across the different pairwise comparisons.
     """
-    # preliminaries 
+    # preliminaries
     y = np.asarray(y)
     N = len(y)
     lseg = int(np.floor(N / num_segs))
     dns = np.zeros((num_points, num_segs))
-    r = np.linspace(np.min(y), np.max(y), num_points) # Make range of ksdensity uniform across all subsegments
+    # Make range of ksdensity uniform across all subsegments
+    r = np.linspace(np.min(y), np.max(y), num_points)
     # Compute the kernel-smoothed distribution in all num_segs segments of the time series
     for i in range(num_segs):
         start_idx = i * lseg
@@ -52,7 +54,7 @@ def local_distributions(y : ArrayLike, num_segs : int = 5, each_or_par : str = '
         segment_data = y[start_idx:end_idx]
         #kde = KDEUnivariate(segment_data)
         kde = gaussian_kde(segment_data, bw_method="scott")
-        #kde.fit(bw="scott") # tune bw adjustment factor empiricially? 
+        #kde.fit(bw="scott") # tune bw adjustment factor empiricially?
         dns[:, i] = kde.evaluate(r)
     # Compare the local distributions
     if each_or_par in ["par", "parent"]:
@@ -152,7 +154,8 @@ def dyn_win(y: ArrayLike, max_num_segments: int = 10) -> dict:
         fs[i, :num_features] = np.std(qs, ddof=1, axis=0)
 
     # fs contains std of quantities at all different 'scales' (segment lengths)
-    fs = np.std(fs, ddof=1, axis=0)  # how much does the 'std stationarity' vary over different scales?
+    # how much does the 'std stationarity' vary over different scales?
+    fs = np.std(fs, ddof=1, axis=0)
 
     # Outputs
     out = {}
@@ -170,8 +173,9 @@ def dyn_win(y: ArrayLike, max_num_segments: int = 10) -> dict:
 
     return out
 
-def moment_corr(x : ArrayLike, window_length : Union[None, float] = None, w_overlap : Union[None, float] = None, 
-               mom_1 : str = 'mean', mom_2 : str = 'std', what_transform : str = 'none') -> dict:
+def moment_corr(x : ArrayLike, window_length : Union[None, float] = None, 
+                w_overlap : Union[None, float] = None, mom_1 : str = 'mean', 
+                mom_2 : str = 'std', what_transform : str = 'none') -> dict:
     """
     Correlations between simple statistics in local windows of a time series.
     The idea to implement this was that of Prof. Nick S. Jones (Imperial College London).
@@ -181,7 +185,8 @@ def moment_corr(x : ArrayLike, window_length : Union[None, float] = None, w_over
     x : array-like
[…]
