### 2026-02-14 68c1b94
linting for surrogates and symbolic
