### 2026-02-15 13fde66
add auto publish workflow
