+++ .github/workflows/wheels.yml
@@ -1,6 +1,10 @@
 name: Build Wheels
 
-on: workflow_dispatch
+on: 
+  workflow_dispatch:
+  push:
+    tags:
+      - "v*"
 
 jobs:
   build_wheels:
@@ -59,4 +63,26 @@ jobs:
         with:
           name: all-dist-files
           path: dist/
+  
+  publish_pypi:
+    name: Publish to PyPI
+    needs: [merge_artifacts]
+    runs-on: ubuntu-latest
+    if: startsWith(github.ref, 'refs/tags/v')  # publish only on version tags
+    runs-on: ubuntu-latest
+    environment: testpypi
+    permissions:
+      id-token: write   # required for PyPI trusted publishing (OIDC)
+      contents: read
+    steps:
+      - name: Download merged dist artifacts
+        uses: actions/download-artifact@v4
+        with:
+          name: all-dist-files
+          path: dist
+
+      - name: Publish to PyPI
+        uses: pypa/gh-action-pypi-publish@release/v1
+        with:
+          packages-dir: dist
           
