+++ .github/workflows/wheels.yml
@@ -65,9 +65,8 @@ jobs:
           path: dist/
   
   publish_pypi:
-    name: Publish to PyPI
+    name: Publish to TestPyPI
     needs: [merge_artifacts]
-    runs-on: ubuntu-latest
     if: startsWith(github.ref, 'refs/tags/v')  # publish only on version tags
     runs-on: ubuntu-latest
     environment: testpypi
@@ -81,8 +80,9 @@ jobs:
           name: all-dist-files
           path: dist
 
-      - name: Publish to PyPI
+      - name: Publish to TestPyPI
         uses: pypa/gh-action-pypi-publish@release/v1
         with:
+          repository-url: https://test.pypi.org/legacy/
           packages-dir: dist
           
