### 2026-02-15 d59a3a1
CI: build wheels and publish (TestPyPI manual, PyPI gated on tags)
