+++ .github/workflows/build_and_publish.yml
@@ -1,17 +1,23 @@
-name: Build Wheels
+name: Build & Publish (manual)
 
-on: 
+on:
   workflow_dispatch:
-  push:
-    tags:
-      - "v*"
+    inputs:
+      target:
+        description: "Where to publish"
+        type: choice
+        required: true
+        default: testpypi
+        options:
+          - testpypi
+          - pypi
 
 jobs:
   build_wheels:
     name: Build wheels on ${{ matrix.os }}
     runs-on: ${{ matrix.os }}
     strategy:
-      fail-fast: false 
+      fail-fast: false
       matrix:
         os: [macos-latest, ubuntu-latest, ubuntu-24.04-arm, windows-latest]
 
@@ -38,6 +44,8 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
 
       - name: Set up Python
         uses: actions/setup-python@v5
@@ -72,13 +80,12 @@ jobs:
         with:
           name: all-dist-files
           path: dist/
-  
-  publish_testpypi:
-    name: Publish to TestPyPI
+
+  publish:
+    name: Publish
     needs: [merge_artifacts]
-    if: github.event_name == 'workflow_dispatch'
     runs-on: ubuntu-latest
-    environment: testpypi
+    environment: ${{ inputs.target }}   # gate pypi if you set approvals on the pypi environment
     permissions:
       id-token: write
       contents: read
@@ -87,26 +94,17 @@ jobs:
         with:
           name: all-dist-files
           path: dist
-      - uses: pypa/gh-action-pypi-publish@release/v1
+
+      - name: Publish to TestPyPI
+        if: inputs.target == 'testpypi'
+        uses: pypa/gh-action-pypi-publish@release/v1
         with:
           repository-url: https://test.pypi.org/legacy/
           packages-dir: dist
 
-  publish_pypi:
-    name: Publish to PyPI
-    needs: [merge_artifacts]
-    if: startsWith(github.ref, 'refs/tags/v')
-    runs-on: ubuntu-latest
-    environment: pypi
-    permissions:
-      id-token: write
-      contents: read
-    steps:
-      - uses: actions/download-artifact@v4
-        with:
-          name: all-dist-files
-          path: dist
-      - uses: pypa/gh-action-pypi-publish@release/v1
+      - name: Publish to PyPI
+        if: inputs.target == 'pypi'
+        uses: pypa/gh-action-pypi-publish@release/v1
         with:
           packages-dir: dist
           
