### 2026-02-17 c9f5100
add SigmoidTransformer to utils for data normalisation
