### 2026-02-18 #282 @crhea93
I've added some comments to the original post @joshuabmoore
