+++ paper/paper.bib
@@ -158,3 +158,15 @@
   url = {http://feasts.tidyverts.org/},
 }
 
+@article{Henderson:2025,
+  author = {Henderson, Trent and Fulcher, Ben D.},
+  title = {Feature-Based Time-Series Analysis in R using the Theft Ecosystem},
+  journal = {The R Journal},
+  year = {2025},
+  note = {https://doi.org/10.32614/RJ-2025-023},
+  doi = {10.32614/RJ-2025-023},
+  volume = {17},
+  issue = {3},
+  issn = {2073-4859},
+  pages = {43-68}
+}
