+++ paper/paper.md
@@ -24,42 +24,45 @@ bibliography: paper.bib
 ---
 
 # Summary
-<!-- Check validity of supporting > 500 publications claim - based on google scholar citations -->
-Across diverse fields of science, finance, and industry, the generation of time-series data has reached an unprecedented scale. Accordingly, there is an increasing need for tools and methods to distill complex, high-dimensional datasets down into interpretable summaries or "features" that capture the varied statistical and dynamical properties underlying the time-series data. From uncovering empirical structure in large datasets [@Fulcher:2013] to seizure and stellar light curve classification [@Barbara:2022], feature-based approaches facilitate a wide range of statistical learning tasks involving time series, wherein the features are derived from interpretable theory and can thus yield insights that can guide understanding and help motivate further experiments. Among these approaches, the MATLAB-based `hctsa` package (which stands for "highly comparative time-series analysis") is the most comprehensive, containing implementations of over 7000 time-series features that can quantify a wide range of distributional and correlation-based statistical properties of time series. Although `hctsa` has supported over 500 publications, its MATLAB-based implementation has presented accessibility issues that have prevented its incorporation into analysis pipelines for open science and across industry applications. Our software package `pyhctsa` addresses these limitations by implementing the majority of the `hctsa` feature library in native Python, bringing a uniquely comprehensive resource of scientific time-series analysis methods to the free and open-source software (FOSS) ecosystem for the first time. The design philosophy of `pyhctsa` is centred around extensibility, with scope to become a living library of time-series analysis methods through further user contributions. Thus far, `pyhctsa` has implemented over 4500 time-series features, making it the most comprehensive Python package for time-series feature extraction by both total volume of features and conceptual coverage.
+Across diverse fields, the generation of time-series data has reached an unprecedented scale. Large feature sets have enabled researchers to distill these complex, high-dimensional datasets into interpretable summary statistics that yield new scientific insight and facilitate a wide range of statistical learning tasks involving time series. While the MATLAB-based `hctsa` (which stands for "highly comparative time-series analysis") offers the most comprehensive feature set to date -- containing over 7000 time-series features and supporting over 500 publications -- its proprietary implementation presents accessibility barriers for open science and industry applications. We address these limitations with `pyhctsa`, which implements the majority of the `hctsa` feature library in native Python, bringing a uniquely comprehensive resource of scientific time-series analysis methods to the free and open-source software (FOSS) ecosystem for the first time. Designed for extensibility, `pyhctsa` includes over 4500 features, making it the most comprehensive Python-based feature set to date, leading in both total volume of features and conceptual coverage.
 
 
 # Statement of need
-Numerous FOSS packages for time-series feature extraction exist across various programming languages such as `kats` [@Jiang:2022] (40 features; Python), `tsfresh` [@Christ:2018] (783 features; Python), `tsfel` [@Barandas:2020] (156 features; Python), and `feasts` (43 features; R) [@Feasts:2026]. However, because these libraries were developed in isolation for different specific purposes, such as tracking human movement (`tsfel`) or econometrics (`kats`), they often overlap in the algorithms they implement while missing many others from the broader scientific literature. This presents a challenge for new problems: there are no clear rules on which tool or feature set to use for the given data, and picking one often means inheriting the narrow focus of its original creators. As a result, these packages may not generalise to new domains where the most informative and useful representations of the data lie outside their scope of feature coverage. 
-
[…]
