+++ pyhctsa/operations/symbolic.py
@@ -27,9 +27,11 @@ def surprise(y : ArrayLike, what_prior : str = 'dist', memory : float = 0.2, num
         The input time series.
     what_prior : {'dist', 'T1', 'T2'}, optional
         The type of information to store in memory:
-            - 'dist': the values of the time series in the previous memory samples (default),
-            - 'T1': the one-point transition probabilities in the previous memory samples,
-            - 'T2': the two-point transition probabilities in the previous memory samples.
+
+        - 'dist': the values of the time series in the previous memory samples (default),
+        - 'T1': the one-point transition probabilities in the previous memory samples,
+        - 'T2': the two-point transition probabilities in the previous memory samples.
+
     memory : float, optional
         The memory length (either number of samples, or a proportion of the time-series length 
         if between 0 and 1). Default is 0.2.
@@ -37,10 +39,12 @@ def surprise(y : ArrayLike, what_prior : str = 'dist', memory : float = 0.2, num
         The number of groups to coarse-grain the time series into. Default is 3.
     coarse_grain_method : {'quantile', 'updown', 'embed2quadrants'}, optional
         The coarse-graining or symbolization method:
-            - 'quantile': equiprobable alphabet by value of each time-series datapoint (default),
-            - 'updown': equiprobable alphabet by incremental changes in the time-series values,
-            - 'embed2quadrants': 4-letter alphabet of the quadrant each data point resides in a 
+
+        - 'quantile': equiprobable alphabet by value of each time-series datapoint (default),
+        - 'updown': equiprobable alphabet by incremental changes in the time-series values,
+        - 'embed2quadrants': 4-letter alphabet of the quadrant each data point resides in a 
             2D embedding space.
+
     num_iters : int, optional
         The number of iterations to repeat the procedure for. Default is 500.
     random_seed : int, optional
@@ -152,18 +156,20 @@ def motif_two(y : ArrayLike, binarize_how : str = 'diff') -> dict:
 
     binarize_how : str, optional
         The method used for binary transformation. One of:
-            - 'diff': Encode increases in the time series as 1, and decreases as 0 (default).
-            - 'mean': Encode values above the mean as 1, and below as 0.
-            - 'median': Encode values above the median as 1, and below as 0.
+
+        - 'diff': Encode increases in the time series as 1, and decreases as 0 (default).
+        - 'mean': Encode values above the mean as 1, and below as 0.
+        - 'median': Encode values above the median as 1, and below as 0.
 
     Returns
     -------
     dict
         A dictionary containing:
-            - 'prob_len_1', 'prob_len_2', ..., 'prob_len_4': 
-              Lists of probabilities for each binary word of lengths 1 to 4.
-            - 'entropy_len_1', 'entropy_len_2', ..., 'entropy_len_4': 
-              Entropy values associated with the word distributions of lengths 1 to 4.
+
+        - 'prob_len_1', 'prob_len_2', ..., 'prob_len_4': 
+            Lists of probabilities for each binary word of lengths 1 to 4.
+        - 'entropy_len_1', 'entropy_len_2', ..., 'entropy_len_4': 
+            Entropy values associated with the word distributions of lengths 1 to 4.
 
     """
     # Generate a binarized version of the input time series
@@ -309,6 +315,7 @@ def motif_three(y : ArrayLike, cg_how : str = 'quantile') -> dict:
         Time series to analyze.
     cg_how : {'quantile', 'diffquant'}, optional
         The coarse-graining method to use:
+
         - 'quantile': equiprobable alphabet by time-series value (default)
[…]
