+++ pyhctsa/operations/distribution.py
@@ -170,7 +170,7 @@ def _find_bounds(pdf_func, start_left, start_right, x_step, thresh):
 
     return xf
 
-def withinp(x : ArrayLike, p : float = 1.0, mean_or_median : str = 'mean') -> float:
+def withinp(x: ArrayLike, p: float = 1.0, mean_or_median: str = 'mean') -> float:
     """
     Proportion of data points within p standard deviations of the mean or median.
 
@@ -234,6 +234,7 @@ def spread(y : ArrayLike, spread_measure : str = 'std') -> float:
         The input time series or data vector.
     spread_measure : str, optional
         The spread measure to use (default is 'std'):
+
         - 'std': standard deviation
         - 'iqr': interquartile range 
         - 'mad': mean absolute deviation
@@ -294,6 +295,7 @@ def proportion_values(x : ArrayLike, prop_what : str = 'positive') -> float:
         Input time series or data vector
     prop_what : str, optional (default is 'positive')
         Type of values to count:
+
         - 'zeros': values equal to zero
         - 'positive': values strictly greater than zero
         - 'geq0': values greater than or equal to zero
@@ -322,8 +324,9 @@ def pleft(y : ArrayLike, th : float = 0.1) -> float:
     """
     Distance from the mean at which a given proportion of data are more distant.
     
-    Measures the maximum distance from the mean at which a given fixed proportion, `th`, of the time-series data points are further.
-    Normalizes by the standard deviation of the time series.
+    Measures the maximum distance from the mean at which a given fixed proportion, `th`, of 
+    the time-series data points are further. Normalizes by the standard deviation of the time 
+    series.
     
     Parameters
     ----------
@@ -354,6 +357,7 @@ def min_max(y : ArrayLike, min_or_max : str = 'max') -> float:
         Input time series or data vector
     minOrMax : str, optional
         Return either the minimum or maximum of y. Default is 'max':
+
         - 'min': minimum of y
         - 'max': maximum of y
 
@@ -382,6 +386,7 @@ def mean(y : ArrayLike, mean_type : str = 'arithmetic') -> float:
         Input time series or data vector
     mean_type : str, optional
         Type of mean to calculate. Default is 'arithmtic':
+
         - 'norm' or 'arithmetic': standard arithmetic mean
         - 'median': middle value (50th percentile)
         - 'geom': geometric mean (nth root of product)
@@ -419,16 +424,16 @@ def mean(y : ArrayLike, mean_type : str = 'arithmetic') -> float:
 
     return float(out)
 
-def high_low_mu(y: ArrayLike) -> float:
+def high_low_mu(y : ArrayLike) -> float:
     """
     The high_low_mu statistic.
 
     The high_low_mu statistic is the ratio of the mean of the data that is above the
     (global) mean compared to the mean of the data that is below the global mean.
 
-    Paramters
+    Parameters
     ----------
-    y : array-like
+    y: array-like
         The input data vector
 
     Returns
@@ -457,6 +462,7 @@ def fit_mle(y : ArrayLike, fit_what : str = 'gaussian') -> Union[Dict[str, float
         Input time series or data vector
     fit_what : {'gaussian', 'uniform', 'geometric'}, optional
         Distribution type to fit:
+
         - 'gaussian': Normal distribution (returns mean and std)
         - 'uniform': Uniform distribution (returns bounds a and b)
         - 'geometric': Geometric distribution (returns p parameter)
@@ -499,20 +505,27 @@ def cv(x : ArrayLike, k : int = 1) -> float:
     """
     Calculate the coefficient of variation of order k.
 
-    The coefficient of variation of order k is (sigma/mu)^k, where sigma is the
[…]
