+++ docs/requirements.txt
@@ -1,1 +1,2 @@
 furo==2021.11.16
+sphinx-immaterial
