+++ docs/requirements.txt
@@ -1,2 +1,3 @@
 furo==2021.11.16
 sphinx-immaterial
+nbsphinx
