+++ docs/requirements.txt
@@ -1,3 +1,4 @@
-furo==2021.11.16
+sphinx
+furo
 sphinx-immaterial
 nbsphinx
