### 2026-02-20 3d9ec05
fix index for usage in docs
