### 2026-02-20 af35418
add development section to docs
