+++ docs/source/development/index.rst
@@ -1,0 +1,14 @@
+Development
+===========
+
+Interested in contributing to `pyhctsa`? Select from the following for more information:
+
+- :doc:`Contributing to pyhctsa <contributing>`
+- :doc:`Conventions <conventions>`
+
+.. toctree::
+   :maxdepth: 1
+   :hidden:
+
+   contributing
+   conventions
