### 2026-02-20 e698503
add coverage report + badge to README
