### 2026-02-20 #282 @joshuabmoore
Hi @crhea93,
Thank you for taking the time to perform the pre-review checks.
Regarding the revisions:

__Major__:
- Sphinx documentation (with API) for `pyhctsa` is now available at https://dynamicsandneuralsystems.github.io/pyhctsa/
- An AI disclosure has been added to the README [here](https://dynamicsandneuralsystems.github.io/pyhctsa/)

__Minor__:
- A coverage report has been added. Coverage is now summarised with the badge on the README.

__Development Activity__
The concentration of commits in September reflects a repository migration and consolidation phase rather than the start of the project's development. The core functionality and codebase was originally developed and iterated on over a much longer period (July 2024 - July 2025) in a separate repository available [here](https://github.com/joshuabmoore/py-hctsa-project). In September, the codebase was reorganised, refactored, and migrated into the current public repository to formalise the package structure, documentation, testing, and release pipeline.
