### 2026-02-22 7df6a2d
add function name mappings
