+++ paper/paper.md
@@ -24,42 +24,42 @@ bibliography: paper.bib
 ---
 
 # Summary
-Across diverse fields, the generation of time-series data has reached an unprecedented scale. Large feature sets have enabled researchers to distill these complex, high-dimensional datasets into interpretable summary statistics that yield new scientific insight and facilitate a wide range of statistical learning tasks involving time series. While the MATLAB-based `hctsa` (which stands for "highly comparative time-series analysis") offers the most comprehensive feature set to date -- containing over 7000 time-series features and supporting over 500 publications -- its proprietary implementation presents accessibility barriers for open science and industry applications. We address these limitations with `pyhctsa`, which implements the majority of the `hctsa` feature library in native Python, bringing a uniquely comprehensive resource of scientific time-series analysis methods to the free and open-source software (FOSS) ecosystem for the first time. Designed for extensibility, `pyhctsa` includes over 4500 features, making it the most comprehensive Python-based feature set to date, leading in both total volume of features and conceptual coverage.
+Across diverse fields, the generation of time-series data has reached an unprecedented scale. Large feature sets have emerged as a powerful solution to distill these datasets into interpretable summary statistics that can yield new insights and facilitate a wide range of statistical learning tasks. While the MATLAB-based `hctsa` (which stands for "highly comparative time-series analysis") offers the broadest coverage of features to date -- containing over 7000 time-series features and supporting over 500 publications -- its proprietary implementation presents accessibility barriers for open science and industry applications. We address these limitations with `pyhctsa`, which implements the majority of the `hctsa` feature set in native Python, bringing a uniquely comprehensive resource of scientific time-series analysis methods to the free and open-source software (FOSS) ecosystem for the first time. `pyhctsa` includes over 4500 time-series features, making it the most extensive Python-based feature set currently available, leading in both total volume of features and conceptual coverage.
 
 
 # Statement of need
-The highly comparative approach, formalised via the `hctsa` package in MATLAB, aims to implement and compare as many existing time-series analysis methods as possible. By simultaneously computing a large set of summary statistics from a diverse scientific literature of methods -- ranging from linear correlation to information theoretic quantities, model fits, scaling, stationarity, and more -- `hctsa` offers the broadest feature coverage of any existing feature set to date. The utility of `hctsa` is well-documented; for example, `hctsa` has identified EEG markers of seizure disorders [@Faiman:2023], and uncovered the dynamical structure of sleep EEG, including substructure in sleep states [@Decat:2022]. Despite these strengths, the reliance of `hctsa` on the proprietary MATLAB ecosystem presents several barriers to broader adoption:
+The highly comparative approach, formalised via the `hctsa` package in MATLAB, aims to implement and compare as many existing time-series analysis methods as possible. By simultaneously computing a large set of summary statistics from a diverse scientific literature of methods -- ranging from linear correlation to information theoretic quantities, model fits, scaling, stationarity, and more -- `hctsa` offers the broadest feature coverage of any existing feature set to date. The utility of `hctsa` is well-documented; for example, `hctsa` has identified EEG markers of seizure disorders [@Faiman:2023], uncovered the dynamical structure of sleep EEG, including substructure in sleep states [@Decat:2022], and characterised subtle behavioural differences in the movement patterns of _C._ _elegans_ [@Fulcher:2017]. Despite its many strengths, the reliance of `hctsa` on the proprietary MATLAB ecosystem presents several barriers to broader adoption:
 
-- __Financial barriers__: The high cost of commercial and academic MATLAB licenses, often compounded by the need for specialised, paid toolboxes of proprietary algorithms, restricts access for independent researchers, students, and even institutions in resource-limited settings.
[…]
