+++ paper/paper.md
@@ -24,59 +24,99 @@ bibliography: paper.bib
 ---
 
 # Summary
-Across diverse fields, the generation of time-series data has reached an unprecedented scale. Large feature sets have emerged as a powerful solution to distill these datasets into interpretable summary statistics that can yield new insights and facilitate a wide range of statistical learning tasks. While the MATLAB-based `hctsa` (which stands for "highly comparative time-series analysis") offers the broadest coverage of features to date -- containing over 7000 time-series features and supporting over 500 publications -- its proprietary implementation presents accessibility barriers for open science and industry applications. We address these limitations with `pyhctsa`, which implements the majority of the `hctsa` feature set in native Python, bringing a uniquely comprehensive resource of scientific time-series analysis methods to the free and open-source software (FOSS) ecosystem for the first time. `pyhctsa` includes over 4500 time-series features, making it the most extensive Python-based feature set currently available, leading in both total volume of features and conceptual coverage.
+Methods for quantifying interpretable dynamical structure in time-series data have been central to many applications across science and industry, from medical diagnosis from physiological signals to demand forecasting in complex supply chains. These methods often yield real-valued summary statistics, or features of the dynamics that, because they are derived from underlying theory, can facilitate an understanding of the structured processes underlying time-series data and guide decision-making. Large feature sets, which aggregate large numbers of diverse time-series analysis methods, provide a powerful way to simultaneously leverage diverse representations of dynamical structure. The MATLAB-based `hctsa` (which stands for “highly comparative time-series analysis”) has the broadest coverage of methods to date, but its proprietary implementation is a barrier to open science and industry applications. Here we introduce `pyhctsa`, which implements the majority of the hctsa feature set in native Python, bringing this comprehensive and unified resource of time-series analysis methods to the free and open-source software (FOSS) ecosystem for the first time. `pyhctsa` includes over 4500 time-series features, making it the most extensive Python-based feature set currently available in both number and conceptual coverage.
 
 
 # Statement of need
-The highly comparative approach, formalised via the `hctsa` package in MATLAB, aims to implement and compare as many existing time-series analysis methods as possible. By simultaneously computing a large set of summary statistics from a diverse scientific literature of methods -- ranging from linear correlation to information theoretic quantities, model fits, scaling, stationarity, and more -- `hctsa` offers the broadest feature coverage of any existing feature set to date. The utility of `hctsa` is well-documented; for example, `hctsa` has identified EEG markers of seizure disorders [@Faiman:2023], uncovered the dynamical structure of sleep EEG, including substructure in sleep states [@Decat:2022], and characterised subtle behavioural differences in the movement patterns of _C._ _elegans_ [@Fulcher:2017]. Despite its many strengths, the reliance of `hctsa` on the proprietary MATLAB ecosystem presents several barriers to broader adoption:
[…]
