+++ pyhctsa/configurations/hctsa.yaml
@@ -11,7 +11,7 @@ medical:
     dependencies:
     configs:
       - {zscore: False}
-    legacy_name: MD_pNN
+    legacy_name: MD_pNN 
 
   pol_var:
     base_name: pol_var
@@ -1583,8 +1583,7 @@ surrogates:
     configs:
       - {surr_meth: 'RP', num_surrs: 99, the_test_stat: 'ami1', zscore: True}
       - {surr_meth: 'RP', num_surrs: 99, the_test_stat: 'fmmi', zscore: True}
-      - {surr_meth: 'RP', num_surrs: 99, the_test_stat: 'o3', zscore: True}
-      - {surr_meth: 'RP', num_surrs: 99, the_test_stat: 'tc3', zscore: True}
+      - {surr_meth: 'RP', num_surrs: 99, the_test_stat: 'o3',   zscore: True}
+      - {surr_meth: 'RP', num_surrs: 99, the_test_stat: 'tc3',  zscore: True}
     legacy_name: SD_SurrogateTest
     ordered_args: ['surr_meth', 'num_surrs', 'the_test_stat']
-    
