### 2026-02-27 c4f06f5
drop nl from unit tests
