+++ paper/paper.md
@@ -36,87 +36,47 @@ Despite its many strengths, the reliance of `hctsa` on the proprietary MATLAB ec
 - __Workflow fragmentation__: As many data-analysis workflows are Python-based, `hctsa` users are often forced into inefficient and convoluted multi-language workflows.
 - __Transparency__: As a proprietary environment, the source code for many MATLAB algorithms are not directly accessible. This conflicts with the ethos of open science, which prioritizes algorithmic transparency and reproducibility.
 
-Given these limitations, there is a clear need for a FOSS implementation of `hctsa`. While several FOSS packages for extracting sets of general time-series features from data exist -- such as `Kats` [@Jiang:2022] (40 features; Python), `tsfresh` [@Christ:2018] (783 features; Python), `TSFEL` [@Barandas:2020] (156 features; Python), and `feasts` (43 features; R) [@Feasts:2026] -- they each have a different scope of inclusion [@Henderson:2025] and, crucially, none match the scale and interdisciplinary coverage of `hctsa`.
+Given these limitations, there is a clear need for a FOSS implementation of `hctsa`.
+
+# State of the field
+While several FOSS packages for extracting sets of general time-series features from data exist -- such as `Kats` [@Jiang:2022] (40 features; Python), `tsfresh` [@Christ:2018] (783 features; Python), `TSFEL` [@Barandas:2020] (156 features; Python), and `feasts` (43 features; R) [@Feasts:2026] -- they each have a different scope of inclusion [@Henderson:2025] and, crucially, none match the scale and interdisciplinary coverage of `hctsa`.
 
 Here we report on `pyhctsa`, which fills the gap in existing FOSS packages with an efficient and extendable architecture for computing a comprehensive set of time-series features from data in native Python. By porting the majority of the `hctsa` library, `pyhctsa` brings an implementation of over 4500 features while preserving the broad methodological reach of the original software. `pyhctsa` can be straightforwardly integrated into existing Python-based data science workflows for performing statistical learning on time series (including classification and regression problems) in ways that connect the analyst to interpretable algorithms derived from scientific theory. 
-
-In the following sections, we describe the architecture and workflow of `pyhctsa` (summarized by the schematic in \autoref{fig:schematic}), detail the methodology used to ensure algorithmic consistency with the original MATLAB implementations, and provide a code example demonstrating its use. 
-
 
 # Software design
 As a package for large-scale feature extraction, the aim of `pyhctsa` is to compute a comprehensive time-series feature set on a given set of univariate time-series data, as shown schematically in \autoref{fig:schematic}(i) and (ii). To achieve this, `pyhctsa` development was guided by several design considerations:
 
-1. __Extensibility__: A modular architecture that allows users to contribute new time-series analysis methods without modifying the core `pyhctsa` codebase. This ensures the software can scale to accommodate new time-series analysis methods.
-2. __Semantic consistency with hctsa__: To preserve the semantic meaning of algorithms across platforms, `pyhctsa` retains original `hctsa` analysis function names and parameter specifications, normalized to Pythonic case conventions. Original function prefixes (e.g., `DN_`) have been dropped from function names and refactored into dedicated Python modules to group conceptually similar algorithms (e.g., distribution for all functions previously prefixed with `DN_`). For example, the `hctsa` function `DN_Spread` is implemented in `pyhctsa` as `distribution.spread`.
+1. __Extensibility__: A modular architecture that can scale to accomodate new time-series anlaysis methods without requiring modifications to the core `pyhctsa` codebase.
+2. __Semantic consistency with hctsa__: To preserve the semantic meaning of algorithms across platforms, `pyhctsa` retains original `hctsa` analysis function names and parameter specifications, normalized to Pythonic case conventions.
[…]
