### 2026-03-30 d0bd3c5
Fix text displacement issue in paper.md
