### 2026-04-13 d69f08b
pin jpype version since no builds for arm64 on latest version
