+++ tests/test_core.py
@@ -7,32 +7,33 @@ import pytest
 from pyhctsa.calculator import FeatureCalculator, classify_output, _standardise_inputs
 from pyhctsa.utils import get_dataset
 
-#----------------- High-level module tests ------------------
-# does the calculator run on module functions (yes/no)?
-@pytest.mark.parametrize("x", [
-    "medical", "extreme_events", "criticality", "correlation", "information", "entropy",
-    "stationarity", "distribution", "scaling", "symbolic", "wavelet", 
-    "hypothesis", "spectral", "model_fit", "graph", "physics", "pre_process",
-    "surrogates"])
-def test_module_basic(x):
-    # basic checks on medical module
-    data = get_dataset(which="sinusoid")
-    config_path = os.path.join(
-        os.path.dirname(os.path.dirname(__file__)), "pyhctsa", "configurations", "module_configs", f"{x}.yaml")
-    assert os.path.exists(config_path), f"Config file not found: {config_path}"
-    calc = FeatureCalculator(config_path)
-    fvec = calc.extract(data)
-    # Check that something is returned and it's not empty
-    assert fvec is not None, "No output returned"
-    assert isinstance(fvec, pd.DataFrame),  "Output should be a dataframe."
+# 1. Time-series feature module tests
+class TestOperations:
+    @pytest.mark.parametrize("x", [
+        "medical", "extreme_events", "criticality", "correlation", "information", "entropy",
+        "stationarity", "distribution", "scaling", "symbolic", "wavelet", 
+        "hypothesis", "spectral", "model_fit", "graph", "physics", "pre_process",
+        "surrogates"])
+    def test_module_basic(self, x):
+        # basic checks on each module
+        data = get_dataset(which="sinusoid")
+        config_path = os.path.join(
+            os.path.dirname(os.path.dirname(__file__)), "pyhctsa", "configurations", "module_configs", f"{x}.yaml")
+        assert os.path.exists(config_path), f"Config file not found: {config_path}"
+        calc = FeatureCalculator(config_path)
+        fvec = calc.extract(data)
+        # Check that something is returned and it's not empty
+        assert fvec is not None, "No output returned"
+        assert isinstance(fvec, pd.DataFrame),  "Output should be a dataframe."
 
-    # check output for multiple time series
-    data2 = get_dataset("e1000")[0:3] # just the first 3 time-series instances
-    fvec2 = calc.extract(data2)
-    assert fvec2 is not None, "No output returned for multiple time-series input"
-    assert isinstance(fvec2, pd.DataFrame), "Output should be a list (of feature dicts)"
-    assert fvec2.shape[0] == 3, "Expected three rows of features for three time series."
+        # check output for multiple time series
+        data2 = get_dataset("e1000")[0:3] # just the first 3 time-series instances
+        fvec2 = calc.extract(data2)
+        assert fvec2 is not None, "No output returned for multiple time-series input"
+        assert isinstance(fvec2, pd.DataFrame), "Output should be a list (of feature dicts)"
+        assert fvec2.shape[0] == 3, "Expected three rows of features for three time series."
 
+# 2. Feature output classification tests
 def test_output_classification():
     test_res = [0.3, 12.3, np.nan, np.inf, -np.inf]
     expected_out = np.array([0, 0, 2, 3, 4])
@@ -43,6 +44,7 @@ def test_output_classification():
     actual_out = np.array(actual_out)
     assert (actual_out == expected_out).all()
 
+# 3. Input data type handling checks
 def test_standardise_inputs():
     # try each input type
     d = get_dataset()
