+++ tests/test_calculator.py
@@ -1,0 +1,203 @@
+import os
+import yaml
+
+import numpy as np
+import pandas as pd
+import pytest
+
+from pyhctsa.calculator import FeatureCalculator, classify_output, _standardise_inputs, _apply_selection_wrapper
+from pyhctsa.utils import get_dataset
+
+# 1. Time-series feature module tests
+class TestOperations:
+    @pytest.mark.parametrize("x", [
+        "medical", "extreme_events", "criticality", "correlation", "information", "entropy",
+        "stationarity", "distribution", "scaling", "symbolic", "wavelet", 
+        "hypothesis", "spectral", "model_fit", "graph", "physics", "pre_process",
+        "surrogates"])
+    def test_module_basic(self, x):
+        # basic checks on each module
+        data = get_dataset(which="sinusoid")
+        config_path = os.path.join(
+            os.path.dirname(os.path.dirname(__file__)), "pyhctsa", "configurations", "module_configs", f"{x}.yaml")
+        assert os.path.exists(config_path), f"Config file not found: {config_path}"
+        calc = FeatureCalculator(config_path)
+        fvec = calc.extract(data)
+        # Check that something is returned and it's not empty
+        assert fvec is not None, "No output returned"
+        assert isinstance(fvec, pd.DataFrame),  "Output should be a dataframe."
+
+        # check output for multiple time series
+        data2 = get_dataset("e1000")[0:3] # just the first 3 time-series instances
+        fvec2 = calc.extract(data2)
+        assert fvec2 is not None, "No output returned for multiple time-series input"
+        assert isinstance(fvec2, pd.DataFrame), "Output should be a list (of feature dicts)"
+        assert fvec2.shape[0] == 3, "Expected three rows of features for three time series."
+
+# 2. Test the calculator
+class TestCalculator:
+    # test the default configuration
+    def test_calculator_defaults(self):
+        calc = FeatureCalculator()
+        # check that default feature set is loaded
+        assert len(calc.config) > 0, "Expected a non-zero number of modules loaded into the FeatureCalculator."
+        config_path = os.path.join(
+            os.path.dirname(os.path.dirname(__file__)), "pyhctsa", "configurations", "hctsa.yaml")
+        # load yaml for comparison
+        with open(config_path, 'rb') as f:
+            hctsa_yaml = yaml.safe_load(f)
+        assert len(hctsa_yaml) == len(calc.config), f"Expected {len(hctsa_yaml)} modules to be loaded, got {len(hctsa_yaml)} instead."
+    # test the loading of a single module e.g. correlation
+    def test_calculator_custom(self):
+        confpath = os.path.join(
+            os.path.dirname(os.path.dirname(__file__)), "pyhctsa", "configurations", "module_configs", "correlation.yaml")
+        calc = FeatureCalculator(config_path=confpath)
+        assert len(calc.config) == 1, f"Expected only a single module, got {len(calc.config)} instead."
+    # test sucessful instantiation of calculator
+    def test_instantiation(self):
+        calc = FeatureCalculator()
+        assert isinstance(calc, FeatureCalculator), "Failed to instantiate FeatureCalculator object."
+    # test basic feature extraction
+    def test_basic_feature_extraction(self):
+        data = get_dataset(which='sinusoid')
+        calc = FeatureCalculator()
+        result = calc.extract(data)
+        assert result is not None
+        assert isinstance(result, pd.DataFrame), "Expected a DataFrame to be returned."
+        assert result.shape[0] == 1, "Expected a single row as output."
+        assert result.shape[1] > 0, "Expected multiple columns as output."
+    # test label functionality
+    def test_output_labelling(self):
[…]
