+++ tests/test_utils.py
@@ -3,7 +3,8 @@ import pytest
 import os
 from pathlib import Path
 
-from pyhctsa.utils import get_dataset, z_score, check_optional_deps, validate_data, _load_csv
+from pyhctsa.utils import get_dataset, z_score, check_optional_deps, validate_data, sign_change
+import pyhctsa.utils
 
 # 1. Dataset loading tests
 class TestDataLoader:
@@ -48,6 +49,13 @@ class TestZScoring:
         x = [7, 7, 7, 7]
         with pytest.raises(ValueError):
             z_score(x)
+    
+    def test_zscore_nonumeric(self):
+        # check handling of non-numeric values
+        x = ['val1', 'val2', 'val3', 'val4', 'val5', 'val6', 'val7']
+        with pytest.raises(TypeError) as excinfo:
+            z_score(x)
+        assert "Input cannot be converted to numeric array" in str(excinfo.value)
 
     @pytest.mark.parametrize("x", [
         [1, 2, np.nan, 4],
@@ -89,3 +97,98 @@ class TestValidateData:
         dat = np.random.randn(1000)
         dat[5] = np.inf
         assert validate_data(dat) is False, "Expected False to be returned"
+
+# 5. Periphery function checks
+class TestPeripheryFuncs:
+    def test_sign_change_default(self):
+        d1 = np.array([0.1, 2, 1, 5, -1, 2, 3, -1, -0.1, -2, 3])
+        expected1 = np.array([False, False, False,  True,  True, False,
+                              True, False, False,True])
+        assert np.array_equal(sign_change(d1), expected1)
+        d2 = np.array([1, 2, 3, 4])
+        expected2 = np.array([False, False, False])
+        assert np.array_equal(sign_change(d2), expected2)
+    def test_sign_change_idx(self):
+        d1 = np.array([0.1, 2, 1, 5, -1, 2, 3, -1, -0.1, -2, 3])
+        expected1 = np.array([3, 4, 6, 9])
+        assert np.array_equal(sign_change(d1, 1), expected1)
+        d2 = np.array([1, 2, 3, 4])
+        assert np.array_equal(sign_change(d2, 1), np.array([]))
+    def test_histc(self):
+        data = get_dataset(which='sinusoid')[0]
+        bin_edges10 = np.linspace(data.min(), data.max(), 10)
+        expected10 = np.array([218.,  94.,  79.,  72.,  70.,  74.,  79.,  97., 215.,   2.])
+        res10 = pyhctsa.utils.histc(data, bin_edges10)
+        assert np.array_equal(res10, expected10), "mismatch between expected counts and actual counts"
+        assert len(res10) == 10, "Expected 10 counts for 10 bins"
+        bin_edges20 = np.linspace(data.min(), data.max(), 20)
+        expected20 = np.array([148.,  63.,  50.,  41.,  41.,  36.,  34.,  37.,  31.,  33.,  36.,
+            31.,  37.,  37.,  38.,  45.,  51.,  63., 146.,   2.])
+        res20 = pyhctsa.utils.histc(data, bin_edges20)
+        assert np.array_equal(res20, expected20), "mismatch between expected counts and actual counts"
+        assert len(res20) == 20, "Expected 20 counts for 20 bins"
+    def test_preproc_decorator(self):
+        simple_test_func = lambda x: sum(x)
+        zs_true_abs_false = pyhctsa.utils.preprocess_decorator(True, False)
+        zs_true_abs_true = pyhctsa.utils.preprocess_decorator(True, True)
+        zs_false_abs_true = pyhctsa.utils.preprocess_decorator(False, True)
[…]
