+++ pyhctsa/configurations/catch22.yaml
@@ -1,124 +1,136 @@
-distribution:
-  histogram_mode:
-    base_name: histogram_mode
-    dependencies:
-    configs:
-      - {num_bins: 5, do_simple: True, zscore: False}
-      - {num_bins: 10, do_simple: True, zscore: False}
-    hctsa_name: DN_HistogramMode
-    ordered_args: ['num_bins']
+# distribution:
+#   histogram_mode:
+#     base_name: histogram_mode
+#     dependencies:
+#     configs:
+#       - {num_bins: 5, do_simple: True, zscore: False}
+#       - {num_bins: 10, do_simple: True, zscore: False}
+#     hctsa_name: DN_HistogramMode
+#     ordered_args: ['num_bins']
   
-  outlier_include:
-    base_name: outlier_include
-    dependencies:
-    configs:
-      - {threshold_how: 'pos', inc: 0.01, zscore: True}
-      - {threshold_how: 'neg', inc: 0.01, zscore: True}
-    legacy_name: DN_OutlierInclude_001_mdrmd
-    ordered_args: ['threshold_how', 'inc']
+#   outlier_include:
+#     base_name: outlier_include
+#     dependencies:
+#     configs:
+#       - {threshold_how: 'pos', inc: 0.01, zscore: True}
+#       - {threshold_how: 'neg', inc: 0.01, zscore: True}
+#     legacy_name: DN_OutlierInclude_001_mdrmd
+#     ordered_args: ['threshold_how', 'inc']
   
-  first_crossing:
-    base_name: first_crossing
-    dependencies:
-    configs:
-      - {corr_fun: 'ac', threshold: '1/e', what_out: 'discrete', zscore: True}
-    legacy_name: CO_f1ecac
-    ordered_args: ['corr_fun', 'threshold', 'what_out']
+#   first_crossing:
+#     base_name: first_crossing
+#     dependencies:
+#     configs:
+#       - {corr_fun: 'ac', threshold: '1/e', what_out: 'discrete', zscore: True}
+#     legacy_name: CO_f1ecac
+#     ordered_args: ['corr_fun', 'threshold', 'what_out']
 
-information:
-  first_min:
-    base_name: first_min
-    dependencies:
-    configs:
-      - {min_what: 'ac', zscore: True}
-    legacy_name: CO_FirstMin_ac
-    ordered_args: ['min_what', 'extra_param', 'min_not_max']
+# information:
+#   first_min:
+#     base_name: first_min
+#     dependencies:
+#     configs:
+#       - {min_what: 'ac', zscore: True}
+#     legacy_name: CO_FirstMin_ac
+#     ordered_args: ['min_what', 'extra_param', 'min_not_max']
   
-spectral:
-  spectral_summaries:
-    base_name: spectral_summaries
-    depedencies:
-    configs:
-      - {psd_meth: 'welch', window_type: 'rect', zscore: True}
-    legacy_name: SP_Summaries_welch_rect
-    ordered_args: ['psd_meth', 'window_type']
+# spectral:
+#   spectral_summaries:
+#     base_name: spectral_summaries
+#     depedencies:
+#     configs:
+#       - {psd_meth: 'welch', window_type: 'rect', zscore: True}
+#     legacy_name: SP_Summaries_welch_rect
+#     ordered_args: ['psd_meth', 'window_type']
 
-model_fit:
-  local_simple:
-    base_name: local_simple
-    dependencies:
-    configs:
-      - {forecast_meth: 'mean', train_length: 1, zscore: True}
-      - {forecast_meth: 'mean', train_length: 3, zscore: True}
-    legacy_name: FC_LocalSimple
[…]
