### 2026-04-16 76c3e6f
add auto-generated legacy function mapping table to docs
