+++ docs/source/mappings/index.rst
@@ -1,0 +1,10 @@
+HCTSA function mappings
+======================
+
+For existing `hctsa` users, the following table provides a mapping between existing `hctsa`
+functions and equivalent `pyhctsa` implementations for reference.
+
+.. csv-table:: 
+   :file: legacy_function_name_mappings.csv
+   :header-rows: 1
+   :widths: auto
