### 2026-04-16 853b685
add auto-generated legacy function mapping table to docs
