### 2026-04-16 c79cb1c
add auto-generated legacy function mapping table to docs
