### 2026-04-16 44585a5
update .toml file to modify the coverage tests (only cover relevant files)
