### 2026-04-16 db4bdc7
update .toml file to modify the coverage tests (only cover relevant files)
