### 2026-04-16 01405e9
update .toml file to modify the coverage tests (only cover relevant files)
