### 2026-04-18 67cd5c8
add new nonlinearity module and validated method nsamdf
