+++ pyhctsa/configurations/hctsa.yaml
@@ -1587,3 +1587,13 @@ surrogates:
       - {surr_meth: 'RP', num_surrs: 99, the_test_stat: 'tc3',  zscore: True}
     legacy_name: SD_SurrogateTest
     ordered_args: ['surr_meth', 'num_surrs', 'the_test_stat']
+
+nonlinearity:
+  nsamdf:
+    base_name: nsamdf
+    depedencies:
+    configs:
+      - {fs: 1, win_len_rel: 14, shift_len_rel: 0.5, lag_rel: 1, degree: 7, zscore: False}
+      - {fs: 1, win_len_rel: 14, shift_len_rel: 0.5, lag_rel: 1, degree: 7, zscore: True}
+    legacy_name: NL_nsamdf
+    ordered_args: ['fs', 'win_len_rel', 'shift_len_rel', 'lag_rel', 'degree']
