+++ pyhctsa/configurations/module_configs/nonlinearity.yaml
@@ -1,0 +1,10 @@
+nonlinearity:
+
+  nsamdf:
+    base_name: nsamdf
+    depedencies:
+    configs:
+      - {fs: 1, win_len_rel: 14, shift_len_rel: 0.5, lag_rel: 1, degree: 7, zscore: False}
+      - {fs: 1, win_len_rel: 14, shift_len_rel: 0.5, lag_rel: 1, degree: 7, zscore: True}
+    legacy_name: NL_nsamdf
+    ordered_args: ['fs', 'win_len_rel', 'shift_len_rel', 'lag_rel', 'degree']
