### 2026-04-18 ce8dbd5
add new validated nonlinear function nlpe
