### 2026-04-19 22f7159
update api to include nonlinearity funcs
