### 2026-04-19 84b8c24
update api to include nonlinearity funcs
