### 2026-04-19 40fc51c
fix correlation module type signatures
