+++ pyhctsa/operations/correlation.py
@@ -13,8 +13,8 @@ from ..operations.information import automutual_info, first_min
 from ..toolboxes.c22 import periodicity_wang_wrapper
 from ..utils import bin_picker, make_mat_buffer, point_of_crossing, sign_change, z_score, histc
 
-def add_noise(y : ArrayLike, tau : Union[int, str] = 1, ami_method : str = 'even',
-              extra_param : Union[int, None] = None, random_seed = None) -> dict:
+def add_noise(y: ArrayLike, tau: Union[int, str] = 1, ami_method: str = 'even',
+              extra_param: Union[int, None] = None, random_seed = None) -> dict:
     """
     Changes in the automutual information with the addition of noise.
 
@@ -167,7 +167,7 @@ def first_under_fn(x : ArrayLike, m : ArrayLike, p : ArrayLike):
     return first_i
 
 
-def theiler_q(y : ArrayLike) -> float:
+def theiler_q(y: ArrayLike) -> float:
     """
     Computes Theiler's Q statistic which quantifies asymmetry in time. 
 
@@ -195,7 +195,7 @@ def theiler_q(y : ArrayLike) -> float:
 
     return float(q)
 
-def crinkle_statistic(y : ArrayLike) -> float:
+def crinkle_statistic(y: ArrayLike) -> float:
     """
     Computes Theiler's crinkle statistic.
 
@@ -230,7 +230,7 @@ def crinkle_statistic(y : ArrayLike) -> float:
 
     return float(out)
 
-def time_rev_kaplan(y : ArrayLike, time_lag : int = 1) -> float:
+def time_rev_kaplan(y: ArrayLike, time_lag: int = 1) -> float:
     """
     Time reversal asymmetry statistic.
 
@@ -257,7 +257,7 @@ def time_rev_kaplan(y : ArrayLike, time_lag : int = 1) -> float:
 
     return float(res)
 
-def _lag_embed(x : ArrayLike, m : int, lag : int = 1) -> ArrayLike:
+def _lag_embed(x: ArrayLike, m: int, lag: int = 1) -> ArrayLike:
     """Constructs a time-delay embedding of a time series."""
     x = np.asarray(x).flatten()
     lx = len(x)
@@ -273,7 +273,7 @@ def _lag_embed(x : ArrayLike, m : int, lag : int = 1) -> ArrayLike:
 
     return y
 
-def embed2_angle_tau(y : ArrayLike, max_tau : int) -> dict:
+def embed2_angle_tau(y: ArrayLike, max_tau: int) -> dict:
     """
     Angle autocorrelation in a 2-dimensional embedding space.
 
@@ -472,7 +472,7 @@ def _histcounts(x: ArrayLike, bins: Union[int, None, str] = None,
 
     return n, edges
 
-def periodicity_wang(y : ArrayLike) -> dict:
+def periodicity_wang(y: ArrayLike) -> dict:
     """
     Periodicity extraction measure of Wang et al. (2007).
 
@@ -700,7 +700,7 @@ def histogram_ami(
     else:
         return {f'ami{i+1}': ami for i, ami in enumerate(amis)}
 
-def stick_angles(y : ArrayLike) -> dict:
+def stick_angles(y: ArrayLike) -> dict:
     """
     Analysis of the line-of-sight angles between time series data pts. 
 
@@ -915,7 +915,7 @@ def stick_angles(y : ArrayLike) -> dict:
 
     return out
 
-def _sub_statav(x : ArrayLike, n : int) -> tuple:
+def _sub_statav(x: ArrayLike, n : int) -> tuple:
     # helper function
     nn = len(x)
     if nn < 2 * n: # not long enough
[…]
