### 2026-04-19 12ace97
fix criticality module type signatures
