### 2026-04-19 159226b
fix distribution module type signatures
