+++ pyhctsa/operations/distribution.py
@@ -205,7 +205,7 @@ def withinp(x: ArrayLike, p: float = 1.0, mean_or_median: str = 'mean') -> float
     # The withinp statistic:
     return np.divide(np.sum((x >= mu - p * sig) & (x <= mu + p * sig)), N)
 
-def unique(y : ArrayLike) -> float:
+def unique(y: ArrayLike) -> float:
     """
     The proportion of the time series that are unique values.
 
@@ -222,7 +222,7 @@ def unique(y : ArrayLike) -> float:
     y = np.asarray(y)
     return np.divide(len(np.unique(y)), len(y))
 
-def spread(y : ArrayLike, spread_measure : str = 'std') -> float:
+def spread(y: ArrayLike, spread_measure: str = 'std') -> float:
     """
     Measure of spread of the input time series.
 
@@ -263,7 +263,7 @@ def spread(y : ArrayLike, spread_measure : str = 'std') -> float:
     
     return out
 
-def quantile(y : ArrayLike, p : float = 0.5) -> float:
+def quantile(y: ArrayLike, p: float = 0.5) -> float:
     """ 
     Calculates the quantile value at a specified proportion, p.
 
@@ -285,7 +285,7 @@ def quantile(y : ArrayLike, p : float = 0.5) -> float:
     
     return float(np.quantile(y, p, method = 'hazen'))
 
-def proportion_values(x : ArrayLike, prop_what : str = 'positive') -> float:
+def proportion_values(x: ArrayLike, prop_what: str = 'positive') -> float:
     """
     Calculate the proportion of values meeting specific conditions in a time series.
 
@@ -320,7 +320,7 @@ def proportion_values(x : ArrayLike, prop_what : str = 'positive') -> float:
 
     return out
 
-def pleft(y : ArrayLike, th : float = 0.1) -> float:
+def pleft(y: ArrayLike, th: float = 0.1) -> float:
     """
     Distance from the mean at which a given proportion of data are more distant.
     
@@ -347,7 +347,7 @@ def pleft(y : ArrayLike, th : float = 0.1) -> float:
 
     return float(out)
 
-def min_max(y : ArrayLike, min_or_max : str = 'max') -> float:
+def min_max(y: ArrayLike, min_or_max: str = 'max') -> float:
     """
     The maximum and minimum values of the input data vector.
 
@@ -376,7 +376,7 @@ def min_max(y : ArrayLike, min_or_max : str = 'max') -> float:
     
     return out
 
-def mean(y : ArrayLike, mean_type : str = 'arithmetic') -> float:
+def mean(y: ArrayLike, mean_type: str = 'arithmetic') -> float:
     """
     A given measure of location of a data vector.
 
@@ -424,7 +424,7 @@ def mean(y : ArrayLike, mean_type : str = 'arithmetic') -> float:
 
     return float(out)
 
-def high_low_mu(y : ArrayLike) -> float:
+def high_low_mu(y: ArrayLike) -> float:
     """
     The high_low_mu statistic.
 
@@ -449,7 +449,7 @@ def high_low_mu(y : ArrayLike) -> float:
 
     return out
 
-def fit_mle(y : ArrayLike, fit_what : str = 'gaussian') -> Union[Dict[str, float], float]:
+def fit_mle(y: ArrayLike, fit_what: str = 'gaussian') -> Union[Dict[str, float], float]:
     """
     Maximum likelihood distribution fit to data.
 
@@ -501,7 +501,7 @@ def fit_mle(y : ArrayLike, fit_what : str = 'gaussian') -> Union[Dict[str, float
 
     return out
 
-def cv(x : ArrayLike, k : int = 1) -> float:
+def cv(x: ArrayLike, k: int = 1) -> float:
     """
[…]
