### 2026-04-19 f187787
fix entropy module type signatures
