+++ docs/source/installation.rst
@@ -1,7 +1,7 @@
 Installation
 ============
 
-`pyhctsa` requires Python 3.10 or higher.
+`pyhctsa` requires `Python 3.10 or higher <https://pypi.org/project/pyhctsa/>`_.
 
 Using pip
 ---------
@@ -29,3 +29,13 @@ To verify that `pyhctsa` is installed correctly, you can run:
 .. code-block:: bash
 
    pip show pyhctsa
+
+If `pyhctsa` has been successfully installed, you will see information about the package version, license info, etc: ::
+
+   Name: pyhctsa
+   Version: 0.1.0
+   Summary: Highly comparative time-series analysis in Python.
+   Home-page: https://github.com/DynamicsAndNeuralSystems/pyhctsa
+   Author: 
+   Author-email: "Joshua B. Moore" <joshua.moore@sydney.edu.au>, "Ben D. Fulcher" <ben.fulcher@sydney.edu.au>
+   License: GNU GENERAL PUBLIC LICENSE
