### 2026-04-21 736412d
add _build_repr_html object instantiation printout like sklearn
