+++ pyhctsa/utils.py
@@ -1,11 +1,13 @@
 import csv
 import os
 from functools import wraps
+from itertools import product
 from importlib.metadata import PackageNotFoundError, version
 from importlib import resources
-from typing import Union, Callable
+from typing import Union, Callable, Any
 import yaml
 from pyhctsa import __version__
+from pathlib import Path
 
 import numpy as np
 from numpy.typing import ArrayLike
@@ -14,6 +16,76 @@ import pandas as pd
 from sklearn.base import BaseEstimator, TransformerMixin
 from sklearn.utils.validation import check_array, check_is_fitted
 
+def _build_repr_html(feature_funcs: dict[str, Callable],
+    skipped_functions: list[tuple[str, list[str]]],
+    config: dict[str, Any],
+    config_path: str) -> str:
+    """
+    Construct an sklearn-like information box when instantiating a FeatureCalculator object.
+
+    Contains information about the modules and feature functions loaded for a given
+    configuration YAML file.
+
+    """
+
+    config_name = Path(config_path).name # extract the name of the module
+    n_feature_funcs = len(feature_funcs)
+    n_skipped = len(skipped_functions)
+    module_counts = {}
+    for module_key, module_cfg in config.items():
+        count = 0
+        for feature_name, feature_config in module_cfg.items():
+            configs = feature_config.get("configs", [{}])
+            for config_item in configs:
+                clean = {k: v for k, v in config_item.items()
+                         if k not in ("zscore", "abs", "select", "exclude")}
+                keys = list(clean.keys())
+                values = [v if isinstance(v, list) else [v] for v in clean.values()]
+                combos = list(product(*values)) if keys else [()]
+                count += len(combos)
+        if count > 0:
+            module_counts[module_key] = count
+    module_counts = dict(sorted(module_counts.items(), key=lambda x: -x[1]))
+    total = sum(module_counts.values())
+    max_count = max(module_counts.values(), default=1)
+    rows = ""
+    skipped_color = "#A32D2D" if n_skipped > 0 else "#412402"
+    for mod, count in module_counts.items():
+        pct = count / max_count * 100
+        pct_of_total = count / total * 100
+        rows += (
+            f'<div style="display:flex;align-items:center;gap:8px;margin-bottom:4px;">'
+            f'<span style="font-size:12px;color:#633806;width:140px;'
+            f'white-space:nowrap;overflow:hidden;text-overflow:ellipsis;">{mod}</span>'
+            f'<div style="flex:1;background:#FAC775;border-radius:2px;height:6px;overflow:hidden;">'
+            f'<div style="width:{pct:.1f}%;height:100%;background:#BA7517;"></div>'
+            f'</div>'
+            f'<span style="font-size:12px;color:#412402;min-width:32px;text-align:right;">{count}</span>'
+            f'<span style="font-size:11px;color:#633806;min-width:36px;text-align:right;">({pct_of_total:.1f}%)</span>'
+            f'</div>'
+        )
+    return (
+        f'<div style="display:inline-block;border:0.5px solid #EF9F27;border-left:4px solid #BA7517;'
+        f'border-radius:8px;padding:1rem 1.25rem;background:#FAEEDA;'
+        f'font-family:monospace;min-width:340px;max-width:560px;">'
+        f'<div style="font-size:13px;font-weight:500;color:#412402;margin-bottom:12px;">FeatureCalculator</div>'
+        f'<div style="display:grid;grid-template-columns:1fr 1fr;gap:8px;margin-bottom:14px;">'
+        f'<div style="background:#FAC775;border-radius:8px;padding:10px 12px;">'
[…]
