### 2026-04-21 f7df8f6
add termnyl animation to docs
