+++ docs/source/development/index.rst
@@ -1,10 +1,23 @@
 Development
 ===========
 
-Interested in contributing to `pyhctsa`? Select from the following for more information:
+Interested in contributing to ``pyhctsa``? Select from the following for more information:
 
-- :doc:`Contributing to pyhctsa <contributing>`
-- :doc:`Conventions <conventions>`
+.. grid:: 3
+   :gutter: 2
+
+   .. grid-item-card:: :material-regular:`volunteer_activism;3em` Contributing
+      :link: contributing
+      :link-type: doc
+
+      General guide on contributing to ``pyhctsa``.
+   
+   .. grid-item-card:: :material-regular:`gavel;3em` Conventions
+      :link: conventions
+      :link-type: doc
+
+      Coding and naming conventions in ``pyhctsa``.
+
 
 .. toctree::
    :maxdepth: 1
