### 2026-04-21 a9782e4
Fix README title for clarity and consistency
