### 2026-04-22 e163a7d
update function documentation with default comments
