### 2026-04-22 722474a
add new validated method is_seasonal + config in hctsa.yaml
