### 2026-04-22 dede484
add lmfit to requirements
