### 2026-04-22 699c109
add is_seasonal to docs api
