+++ docs/source/api.rst
@@ -179,6 +179,7 @@ Model Fit
    pyhctsa.operations.model_fit.exp_smoothing
    pyhctsa.operations.model_fit.ar_cov
    pyhctsa.operations.model_fit.ar_fit
+   pyhctsa.operations.model_fit.is_seasonal
 
 .. _physicsmeths:
 
