### 2026-04-23 6f15e31
add new validated nonlinearity function embed_pca
