+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "pyhctsa"
-version = "0.1.1"
+version = "0.1.2"
 description = "Highly comparative time-series analysis in Python."
 readme = "README.md"
 license = {file = "LICENSE"}
