### 2026-04-29 7fe5805
Fix sliding stationarity (don't assign array to scalar)
