+++ pyhctsa/operations/stationarity.py
@@ -1056,7 +1056,7 @@ def sliding_window(y: ArrayLike, window_stat: str = 'mean', across_win_stat: str
             qs[i] = moments(y[_get_window(i, inc, win_length)], 5)
     elif window_stat == 'AC1':
         for i in range(num_steps):
-            qs[i] = autocorr(y[_get_window(i, inc, win_length)], 1, 'Fourier')
+            qs[i] = np.asarray(autocorr(y[_get_window(i, inc, win_length)], 1, 'Fourier')).item()
     else:
         raise ValueError(f"Unknown statistic '{window_stat}'")
     
