+++ pyhctsa/operations/stationarity.py
@@ -569,7 +569,7 @@ def range_evolve(y: ArrayLike) -> dict:
         if N >= n_val:
             out[f'nuql{n_val}'] = lunique(cums[:n_val])/out['totnuq']
         else:
-            out[f'nuql{N}'] = np.nan
+            out[f'nuql{n_val}'] = np.nan
     # (**2**) Actual proportion of full range captured at different points
     out['p1'] = cums[int(np.ceil(N * 0.01)) - 1]/fullr
     out['p10'] = cums[int(np.ceil(N * 0.1)) - 1]/fullr
