### 2026-04-29 4df3962
Update list formatting from 'start_end' to 'start_to_end'
E.g. this. Could be anything really as long as its not just '1_n'
