+++ paper/paper.md
@@ -58,7 +58,7 @@ In `pyhctsa` features are generated from a YAML file (\autoref{fig:schematic}(ii
 
 
 ## Validating function implementations
-To provide broad access to the time-series analysis methods implemented in `hctsa`, we ported the majority of its library to native Python, while remaining as consistent as possible with the original algorithms. Achieving strict numerical equivalence between Python and MATLAB was often unattainable due to differences in floating-point arithmetic and linear algebra routines. Consequently, to verify the capture of a common time-series property, we required that ported algorithms exhibit a similar variation across a wide range of data. To acheive this, 
+To provide broad access to the time-series analysis methods implemented in `hctsa`, we ported the majority of its library to native Python, while remaining as consistent as possible with the original algorithms. Achieving strict numerical equivalence between Python and MATLAB was often unattainable due to differences in floating-point arithmetic and linear algebra routines. Consequently, to verify the capture of a common time-series property, we required that ported algorithms exhibit a similar variation across a wide range of data. To achieve this, 
 we computed the Pearson correlation $r$ between the outputs of a given MATLAB and Python algorithm across a benchmark of 1000 diverse (simulated and empirical) time series [@Fulcher:2021].
 We then retained only those ported functions that demonstrated strong statistical agreement (defined as $r \geq 0.9$) with original MATLAB implementations. At release, `pyhctsa` includes 119 algorithms (73% of the original library) and maintains similar conceptual coverage to `hctsa`. Together, these algorithms yield over 4500 validated features. The remaining 44 algorithms from `hctsa` were excluded due to the absence of open-source equivalents or failure to meet the threshold for statistical agreement.
 
