### 2026-04-29 047f9a8
Fix scalar assignment in corr_fun autocorr branch
