### 2026-05-03 3bbce65
Update correlation.py
`pos_drown` was a recent change in the autocorr_shape and hadn't been updated in the config yaml file to reflect this. As with all other functions and parameter settings in pyhctsa, this is supposed to be camel case (to match the pythonic naming conventions.
