### 2026-05-03 a28a2fe
update docstring for format_param_value
