### 2026-05-05 8317721
replace print statements with logging messages
