### 2026-05-05 e463065
replace print statement with logging for multiscale entropy
