### 2026-05-05 e6fcac8
replace print statements with logging for autocorr
