### 2026-05-05 4a083d4
replace print statement with logging info msg for viz graph
