### 2026-05-05 e8e718f
replace print statements with logging msgs for nonlinearity funcs
