### 2026-05-05 aa3ccc5
replace print statements with logging msgs for model fit funcs
