### 2026-05-05 ae7be2e
replace print statements with logging msgs for stationarity funcs
