### 2026-05-05 0b3c77f
replace print statements with logging msgs for scaling funcs
