### 2026-05-05 196535e
replace print statements with logging msgs for wavelet funcs
