### 2026-05-05 74f592f
replace print statements with logging msgs for preproc funcs
