### 2026-05-05 d7f8bea
replace print statements with logging msgs for information theoretic funcs
