### 2026-05-05 e63cb23
replace print statements with logging msgs for information theoretic funcs
