### 2026-05-05 490fb8d
cleanup comments in various funcs
