### 2026-05-05 1d24ffe
fix various type signature formatting inconsistencies
