+++ pyhctsa/utils.py
@@ -107,7 +107,7 @@ def _check_optional_deps(dep: str) -> bool:
     except PackageNotFoundError:
         return False
 
-def _validate_data(ts : np.ndarray) -> bool:
+def _validate_data(ts: np.ndarray) -> bool:
     """validate a time series before computing features"""
     if len(ts) < 100:
         logging.warning("Time series is too short!")
@@ -194,7 +194,7 @@ def get_dataset(which: str = "e1000") -> list:
     print(f"Loaded dataset of {len(dataset)} time series.")
     return dataset
     
-def _preprocess_decorator(zscore : bool = False, absval : bool = False) -> Callable:
+def _preprocess_decorator(zscore: bool = False, absval: bool = False) -> Callable:
     """
     Decorator to preprocess time series data before feature computation.
     
@@ -231,7 +231,7 @@ def _preprocess_decorator(zscore : bool = False, absval : bool = False) -> Calla
         return wrapper
     return decorator
 
-def z_score(x : ArrayLike) -> np.ndarray:
+def z_score(x: ArrayLike) -> np.ndarray:
     """
     Z-score the input data vector.
 
@@ -278,7 +278,7 @@ def z_score(x : ArrayLike) -> np.ndarray:
 
     return zscored_data
 
-def histc(x : ArrayLike, bins : ArrayLike) -> int:
+def histc(x: ArrayLike, bins: ArrayLike) -> int:
     """Counts the number of values in x that are within each specified bin."""
     # Get indices of the bins to which each value in input array belongs.
     map_to_bins = np.digitize(x, bins)
@@ -287,8 +287,8 @@ def histc(x : ArrayLike, bins : ArrayLike) -> int:
         res[el-1] += 1 # Increment appropriate bin.
     return res
 
-def bin_picker(x_min : float, x_max : float, n_bins : Union[None, int],
-               bin_width_est : Union[None, float] = None) -> np.ndarray:
+def bin_picker(x_min: float, x_max: float, n_bins: Union[None, int],
+               bin_width_est: Union[None, float] = None) -> np.ndarray:
     """
     Choose histogram bins. 
 
@@ -387,7 +387,7 @@ def bin_picker(x_min : float, x_max : float, n_bins : Union[None, int],
 
     return edges
 
-def simple_binner(x_data : ArrayLike, num_bins : int) -> tuple:
+def simple_binner(x_data: ArrayLike, num_bins: int) -> tuple:
     """
     Generate a histogram from equally spaced bins.
    
@@ -460,7 +460,7 @@ def point_of_crossing(x: ArrayLike, threshold: float) -> tuple:
 
     return fc, poc
 
-def sign_change(y : Union[list, np.ndarray], do_find: int = 0) -> ArrayLike:
+def sign_change(y: Union[list, np.ndarray], do_find: int = 0) -> ArrayLike:
     """
     Where a data vector changes sign.
 
@@ -478,7 +478,7 @@ def sign_change(y : Union[list, np.ndarray], do_find: int = 0) -> ArrayLike:
 
     return indexs
 
-def make_buffer(y : ArrayLike, buffer_size : int) -> np.ndarray:
+def make_buffer(y: ArrayLike, buffer_size: int) -> np.ndarray:
     """
     Make a buffered version of a time series.
 
@@ -507,8 +507,8 @@ def make_buffer(y : ArrayLike, buffer_size : int) -> np.ndarray:
 
     return y_buffer
 
-def make_mat_buffer(x : ArrayLike, n : int, p : int = 0,
-                    opt : Union[str, None] = None) -> np.ndarray:
+def make_mat_buffer(x: ArrayLike, n: int, p: int = 0,
[…]
