### 2026-05-05 24336b5
fix type signatures in various function docstrings
