### 2026-05-15 85e5816
add JOSS under review badge to README
