### 2026-05-15 307d14c
fix docs methods index file
