### 2026-06-09 b0dea91
fix typo in information.yaml conf
