### 2026-06-09 f443b5a
updated handling of java and java depedencies
