+++ pyhctsa/calculator.py
@@ -6,6 +6,8 @@ from pathlib import Path
 from typing import Union, Any, Callable
 import logging
 
+logger = logging.getLogger('pyhctsa')
+
 import numpy as np
 import pandas as pd
 import yaml
@@ -70,7 +72,7 @@ def _apply_selection_wrapper(func: Callable, filter_keys: Union[str, list[str]],
         if isinstance(result, dict):
             missing = [k for k in keys if k not in result] # log all of the missing keys
             if missing:
-                logging.info(f'Warning: time-series features for func {func} not found {missing}')
+                logger.info(f'Warning: time-series features for func {func} not found {missing}')
             if keep:
                 return {k: result[k] for k in keys if k in result}
             else:
@@ -87,7 +89,7 @@ def _standardise_inputs(data) -> list[np.ndarray]:
         elif data.ndim == 2:
             if data.shape[0] > data.shape[1]:
                 # notify the user to check that the shapes make sense
-                logging.warning(f"Check that the shape of the 2D input is such "
+                logger.warning(f"Check that the shape of the 2D input is such "
                       f"that (n_series, n_samples). Got shape: {data.shape}")
             return [np.asarray(row, dtype=float) for row in data]
         else:
@@ -209,34 +211,36 @@ class FeatureCalculator:
         return _build_repr_html(self.feature_funcs, self._skipped_functions, self.config, self.config_path)
 
     def _check_deps(self, module_key, feature_name, config):
-        raw_deps = config.get("dependencies")
+        raw_deps = config.get("dependencies", None)
         if not raw_deps:
             return True
         deps_to_check = [raw_deps] if isinstance(raw_deps, str) else raw_deps
         missing = [dep for dep in deps_to_check if not _check_optional_deps(dep)]
         if missing:
             full_name = f"{module_key}.{feature_name}"
-            logging.info(f"Skipping function '{full_name}' - missing dependencies: {', '.join(missing)}")
+            logger.info(f"Skipping function '{full_name}' - missing dependencies: {', '.join(missing)}")
             self._skipped_functions.append((full_name, missing))
             return False
         return True
     
     def _build_feature_funcs(self):
         feature_funcs = {}
-        skipped_functions = []       
+        self._skipped_functions = []      
         for module_key in self.config.keys():
 
             try:
                 module = importlib.import_module(f"{self._operations_package}.{module_key}")
             except ImportError as e:
-                logging.warning(f"Failed to import module '{module_key}': {e}")
+                logger.warning(f"Failed to import module '{module_key}': {e}")
                 # Skip all functions in this module since we can't import it
                 for feature_name in self.config[module_key].keys():
-                    skipped_functions.append((f"{module_key}.{feature_name}", ["import_error"]))
+                    self._skipped_functions.append((f"{module_key}.{feature_name}", ["import_error"]))
                 continue
 
             # Process features from this module
             for feature_name, feature_config in self.config[module_key].items():
+                if not self._check_deps(module_key, feature_name, feature_config):
+                    continue
                 op_func = getattr(module, feature_name)
                 base_name = feature_config.get("base_name", feature_name)
                 ordered_args = feature_config.get("ordered_args", [])
@@ -270,11 +274,8 @@ class FeatureCalculator:
                         
                         feature_funcs[label] = final_func
         
-        # store information about skipped functions for later reference
-        self._skipped_functions = skipped_functions
-        if skipped_functions:
-            logging.info(f"Total functions skipped due to missing dependencies: {len(skipped_functions)}")
-        
+        if self._skipped_functions:
+            logger.info(f"Total functions skipped due to missing dependencies: {len(self._skipped_functions)}")
         return feature_funcs
 
     def extract(self, data: Union[ArrayLike, list[ArrayLike]],
