### 2026-06-09 2e53a0e
update README and docs to notify users of feature skipping when java is missing
