### 2026-06-10 48879cc
update return values for critical errors to be nan rather than string error
