### 2026-06-10 ef7463f
java and jidt deps have been removed and mutual info/ami features have been replaced by validated python drop-ins
