+++ pyhctsa/operations/information.py
@@ -8,6 +8,7 @@ from numpy.typing import ArrayLike
 from scipy import stats
 
 from ..utils import sign_change
+from ..toolboxes.infotheory.mutual_info import KraskovMI, GaussianMI
 
 def _get_corr_fn(y: np.ndarray, min_what: str, extra_param: Union[int, float, None]) -> Callable:
     """Helper to return the correct correlation function based on method type."""
@@ -22,8 +23,6 @@ def _get_corr_fn(y: np.ndarray, min_what: str, extra_param: Union[int, float, No
         return lambda x: automutual_info(y, x, 'kraskov2', extra_param)
     elif min_what == 'mi-kraskov1':
         return lambda x: automutual_info(y, x, 'kraskov1', extra_param)
-    elif min_what == 'mi-kernel':
-        return lambda x: automutual_info(y, x, 'kernel', extra_param)
     elif min_what in ['mi', 'mi-gaussian']:
         return lambda x: automutual_info(y, x, 'gaussian', extra_param)
     else:
@@ -51,10 +50,9 @@ def first_min(
 
         Automutual information (AMI):
 
-        - ``'mi'``: AMI using the JIDT Gaussian estimator (default for AMI).
-        - ``'mi-kernel'``: AMI using the JIDT kernel estimator.
-        - ``'mi-kraskov1'``: AMI using the JIDT Kraskov estimator (variant 1).
-        - ``'mi-kraskov2'``: AMI using the JIDT Kraskov estimator (variant 2).
+        - ``'mi'``: AMI using the Gaussian estimator (default for AMI).
+        - ``'mi-kraskov1'``: AMI using the Kraskov estimator (variant 1).
+        - ``'mi-kraskov2'``: AMI using the Kraskov estimator (variant 2).
         - ``'mi-hist'``: AMI using a histogram-based estimator.
 
         Default is ``'mi-gaussian'``.
@@ -109,10 +107,9 @@ def first_max(
 
         Automutual information (AMI):
 
-        - ``'mi'``: AMI using the JIDT Gaussian estimator (default for AMI).
-        - ``'mi-kernel'``: AMI using the JIDT kernel estimator.
-        - ``'mi-kraskov1'``: AMI using the JIDT Kraskov estimator (variant 1).
-        - ``'mi-kraskov2'``: AMI using the JIDT Kraskov estimator (variant 2).
+        - ``'mi'``: AMI using the Gaussian estimator (default for AMI).
+        - ``'mi-kraskov1'``: AMI using the Kraskov estimator (variant 1).
+        - ``'mi-kraskov2'``: AMI using the Kraskov estimator (variant 2).
         - ``'mi-hist'``: AMI using a histogram-based estimator.
 
         Default is ``'mi'``.
@@ -228,7 +225,7 @@ def automutual_info_stats(
     max_tau : int, optional
         Maximum time delay to investigate. If None, uses N/4 where N is the length
         of the time series, but won't exceed N/2. Default is `None`.
-    est_method : {'gaussian', 'kernel', 'kraskov1', 'kraskov2'}, optional
+    est_method : {'gaussian', 'kraskov1', 'kraskov2'}, optional
         Method for estimating mutual information (passed to automutual_info).
         Default is ``'kernel'``.
     extra_param : int or str, optional
@@ -321,18 +318,17 @@ def automutual_info_stats(
     out['amiac1'] = autocorr(ami, 1, 'Fourier')[0]
 
     return out
-
+    
 def automutual_info(
-    y: ArrayLike,
-    time_delay: Union[int, str, List[int]] = 1,
-    est_method: str = 'kernel',
-    extra_param: Optional[Union[int, str]] = None
-) -> Union[float, Dict[str, float]]:
[…]
