### 2026-06-10 4966700
remove jidt toolbox, add infotheory toolbox
