### 2026-06-13 08dc25d
fix transition matrix to use resample_poly instead of ssre
