### 2026-06-20 7547244
Vectorize first min ami
