### 2026-06-21 cf38c5d
Use vectorized autocorr in first_max also
