### 2026-06-21 f375ae8
Faster features with value-equivalent output
