### 2026-06-22 40a9b78
Add iteration count to hmm feature output
