### 2026-06-22 a7169ef
dropped medianf from hctsa.yaml due to reproducibility issues
