+++ pyhctsa/operations/surrogates.py
@@ -1,5 +1,5 @@
 import warnings
-from typing import Union, Optional
+from typing import Union
 import logging
 logger = logging.getLogger('pyhctsa')
 
@@ -12,118 +12,106 @@ from ..operations.information import automutual_info, first_min
 
 warnings.filterwarnings("ignore", category=RuntimeWarning)
 
-def _nan_stats() -> dict:
-    """Return a nan-filled stats dict matching the schema of `sd_give_me_stats`.
- 
-    Used as a soft-fail so downstream feature extraction still receives the
-    expected keys (rather than a bare scalar) when the surrogate distribution
-    is degenerate.
-    """
-    return {'p': np.nan, 'zscore': np.nan, 'f': np.nan,
-            'mediqr': np.nan, 'prank': np.nan}
-
 def sd_give_me_stats(stat_x: float, stat_surr: ArrayLike, left_right_both: str) -> dict:
-    """Compute statistics on the surrogate distribution."""
+    """Compute statistiscs on the surrogate distribution."""
     num_surrs = len(stat_surr)
     out = {}
     if np.isnan(stat_surr).any():
         logger.warning("SDgivemestats failed")
-        return _nan_stats()
- 
-    # Scope the numeric-warning suppression to this computation only, rather
-    # than silencing all RuntimeWarnings for every module that imports this one.
-    with warnings.catch_warnings(), np.errstate(divide='ignore', invalid='ignore', over='ignore'):
-        warnings.simplefilter("ignore", category=RuntimeWarning)
- 
-        #% ASSUME GAUSSIAN DISTRIBUTION:
-        #% so can use 1/2-sided z-statistic
-        z_stat = zmap(np.atleast_1d(stat_x), stat_surr, ddof=1)[0]
-        p = None
-        if left_right_both == 'both':
-            p = 2 * norm.sf(np.abs(z_stat))
-        elif left_right_both == 'right':
-            p = norm.sf(z_stat)
-        elif left_right_both == 'left':
-            p = norm.cdf(z_stat)
-        out['p'] = p
-        out['zscore'] = z_stat
- 
-        # fit a kernel distribution to zscored distributions
-        sigma = np.std(stat_surr, ddof=1)
-        mu = np.mean(stat_surr)
-        if sigma == 0 or not np.isfinite(sigma):
-            # all surrogates have same value of this statistic
-            # cannot do a meaningful zscore
-            c = float(stat_surr[0])
-            bw = 1.0  # bandwidth scale
-            xi = np.linspace(c - 3 * bw, c + 3 * bw, 100)
-            # tiny gaussian around c
-            f = norm.pdf(xi, loc=c, scale=bw)
-            if (stat_x < xi.min()) or (stat_x > xi.max()):
-                out["f"] = 0.0
-            else:
-                idx = int(np.argmin(np.abs(stat_x - xi)))
-                out["f"] = float(f[idx])
- 
+        return np.nan
+    #% ASSUME GAUSSIAN DISTRIBUTION:
+    #% so can use 1/2-sided z-statistic
+    z_stat = zmap(np.atleast_1d(stat_x), stat_surr, ddof=1)[0]
+    p = None
+    if left_right_both == 'both':
+        p = 2 * norm.sf(np.abs(z_stat))
+    elif left_right_both == 'right':
+        p = norm.sf(z_stat)
+    elif left_right_both == 'left':
+        p = norm.cdf(z_stat)
+    out['p'] = p
+    out['zscore'] = z_stat
+
+    # fit a kenerel distribution to zscored distributions
+    sigma = np.std(stat_surr, ddof=1)
+    mu = np.mean(stat_surr)
+    if sigma == 0 or not np.isfinite(sigma):
+        # all surrogates have same value of this statisitc
[…]
