### 2026-06-22 fa961ea
add validation harness
