### 2026-06-22 ceac2b1
zscore revert
