### 2026-06-22 dff1ab9
Update README.md for validation instructions
Removed notes about adjusting filenames and Python dependencies. Updated instructions for running the validation harness.
