### 2026-06-22 d6115cd
update validation README
