+++ tests/validation/run_validation.py
@@ -13,8 +13,9 @@ from sklearn.metrics import mean_absolute_percentage_error, median_absolute_erro
 import importlib
 import numpy as np
 from collections import defaultdict
-from scipy.stats import pearsonr, ks_2samp
+from scipy.stats import pearsonr
 import pandas as pd
+import argparse
  
 try:
     from tqdm import tqdm
@@ -41,20 +42,12 @@ def _distribution_parity(py_dist, mat_dist):
     """Compare two per-series seed distributions for one feature."""
     py_dist = np.asarray(py_dist, float); py_dist = py_dist[np.isfinite(py_dist)]
     mat_dist = np.asarray(mat_dist, float); mat_dist = mat_dist[np.isfinite(mat_dist)]
-    out = {
+    return {
         "py":  np.mean(py_dist)  if py_dist.size  else np.nan,
         "mat": np.mean(mat_dist) if mat_dist.size else np.nan,
         "std_py":  np.std(py_dist,  ddof=1) if py_dist.size  > 1 else np.nan,
         "std_mat": np.std(mat_dist, ddof=1) if mat_dist.size > 1 else np.nan,
     }
-    if py_dist.size >= 2 and mat_dist.size >= 2:
-        if np.ptp(py_dist) == 0 and np.ptp(mat_dist) == 0:
-            out["ks_p"] = 1.0 if np.isclose(py_dist[0], mat_dist[0]) else 0.0
-        else:
-            out["ks_p"] = ks_2samp(py_dist, mat_dist).pvalue
-    else:
-        out["ks_p"] = np.nan
-    return out
  
  
 def generate_test_cases(yaml_file):
@@ -240,7 +233,7 @@ def test_matlab_parity(matlab_engine, data, suite_name, settings, params):
         py_once = op_func(test_data, **run_args)
         py_once = py_once if isinstance(py_once, dict) else {"output": py_once}
         for k in _flatten(_normalize(py_once)):
-            results[f"{base_name}::{k}"] = {"py": np.nan, "mat": np.nan, "ks_p": np.nan}
+            results[f"{base_name}::{k}"] = {"py": np.nan, "mat": np.nan}
         return results
  
     all_keys = set().union(*(set(d) for d in seed_py), *(set(d) for d in seed_mat))
@@ -276,8 +269,6 @@ def _accumulate(aggregator, res):
             continue
         aggregator[full_key]["py"].append(values["py"])
         aggregator[full_key]["mat"].append(values["mat"])
-        if "ks_p" in values and np.isfinite(values["ks_p"]):
-            aggregator[full_key]["ks_p"].append(values["ks_p"])
  
  
 def _finalize_param(settings, aggregator):
@@ -299,9 +290,6 @@ def _finalize_param(settings, aggregator):
         else:
             mae = mape = np.nan
  
-        ks = lists["ks_p"]
-        ks_pass_frac = float(np.mean(np.asarray(ks) > 0.05)) if ks else np.nan
- 
         fname = feature_name.split(":")[2]
         if fname == 'output':
             fname = feature_name.split(":")[0]
@@ -312,12 +300,10 @@ def _finalize_param(settings, aggregator):
             "matlab_func": settings["matlab_func"],
             "python_module": settings["python_module"],
             "corr": r, "mae": mae, "mape": mape,
-            "ks_pass_frac": ks_pass_frac,
-            "n_stochastic_series": len(ks) if ks else 0,
             "nan_count_py": int(np.sum(~np.isfinite(py_vec))),
             "nan_count_ml": int(np.sum(~np.isfinite(mat_vec))),
         })
-        print(f"{settings['python_func']}-{feature_name}: corr={r:.4f} ks_pass={ks_pass_frac}")
[…]
