### 2026-06-23 fd1eb75
Replace acorr_ljunbox with faster manual test elsewhere
