+++ CITATION.cff
@@ -12,7 +12,7 @@ authors:
     orcid: "https://orcid.org/0000-0002-3003-4055"
 
 doi: 10.5281/zenodo.18529935
-version: 0.1.2
+version: 1.0.0
 date-released: 2026-02-09
 url: https://github.com/DynamicsAndNeuralSystems/pyhctsa
 repository-code: https://github.com/DynamicsAndNeuralSystems/pyhctsa
