### 2026-06-23 f0d5b84
Revise README badges and formatting
Updated badge links and improved formatting in README.
