+++ paper/paper.bib
@@ -99,16 +99,6 @@
   year = {2024}
 }
 
-@article{Gorgannejad:2023,
-  title = {Localized keyhole pore prediction during laser powder bed fusion via multimodal process monitoring and X-ray radiography},
-  journal = {Additive Manufacturing},
-  volume = {78},
-  pages = {103810},
-  year = {2023},
-  doi = {10.1016/j.addma.2023.103810},
-  author = {Gorgannejad, Sanam and Martin, Aiden A. and Nicolino, Jenny W. and Strantza, Maria and Guss, Gabriel M. and Khairallah, Saad and Forien, Jean-Baptiste and Thampy, Vivek and Liu, Sen and Quan, Peiyu and Tassone, Christopher J. and Calta, Nicholas P.}
-}
-
 @article{Paul:2021,
   doi = {10.1371/journal.pcbi.1008820},
   author = {Paul, Avishek and McLendon, Helen and Rally, Veronica and Sakata, Jon T. and Woolley, Sarah C.},
