### 2026-07-07 42ac90b
update referencing on docs index w/ new JOSS publication
