### 2026-07-07 4073287
add JOSS badge to index.rst for docs
