### 2026-07-09 fc31258
Fix pyOpenSci Peer-Reviewed badge link
Updated the link for pyOpenSci Peer-Reviewed badge in README.
