### 2026-07-09 8483ae3
add pyopensci badge to docs
