### 2026-07-16 b36f318
optimize spread function by reducing calls to percentile
