### 2026-07-16 0f1ddf1
optimized proportion_values by replacing with vectorized operations
