### 2026-07-16 f005b65
optimize min_max with vectorized operations
