### 2026-07-16 5b050fd
optimized cv with a single exponentiation
