### 2026-07-16 a90c502
optimized burstiness via caching results
