### 2026-07-17 3fdc7f8
make optimizations to the make_buffer function
