### 2026-07-17 bd96b42
make optimizations to the make_mat_buffer function
