### 2026-07-17 df77f90
vectorized pnn for speed
