### 2026-07-17 18ac1e9
vectorized polvar for speed
