### 2026-07-17 8b3f462
optimizations to correlation, distribution and entropy
