+++ pyhctsa/operations/correlation.py
@@ -4,6 +4,7 @@ from typing import Union
 
 import numpy as np
 from numpy.typing import ArrayLike
+from numpy.lib.stride_tricks import sliding_window_view
 from scipy.linalg import LinAlgError
 from scipy.optimize import curve_fit
 from scipy.stats import expon, gaussian_kde, kurtosis, skew
@@ -159,11 +160,13 @@ def add_noise(y: ArrayLike, tau: Union[int, str] = 1, ami_method: str = 'even',
 
     return out
 
-def first_under_fn(x: ArrayLike, m: ArrayLike, p: ArrayLike) -> float:
+def first_under_fn(x: float, m: np.ndarray, p: np.ndarray) -> float:
     """
     Find the value of m for the first time p goes under the threshold, x. 
     p and m are vectors of the same length
     """
+    m = np.asarray(m)
+    p = np.asarray(p)
     first_i = next((m_val for m_val, p_val in zip(m, p) if p_val < x), m[-1])
 
     return first_i
@@ -232,7 +235,7 @@ def crinkle_statistic(y: ArrayLike) -> float:
 
     return float(out)
 
-def time_rev_kaplan(y: ArrayLike, time_lag: int = 1) -> float:
+def time_rev_kaplan(y: np.ndarray, time_lag: int = 1) -> float:
     """
     Time reversal asymmetry statistic.
 
@@ -300,52 +303,42 @@ def embed2_angle_tau(y: ArrayLike, max_tau: int) -> dict:
         Dictionary containing statistics of the autocorrelation of angles for each tau,
         including mean, max, min, and autocorrelation at different lags.
     """
+    y = np.asarray(y).ravel()
+    dy = np.diff(y)
+
     tau_range = np.arange(1, max_tau + 1)
     num_tau = len(tau_range)
-
-    # Ensure y is a column vector
-    y = np.atleast_2d(y)
-    if y.shape[0] < y.shape[1]:
-        y = y.T
-
     stats_store = np.zeros((3, num_tau))
 
     for i, tau in enumerate(tau_range):
-        m = np.column_stack((y[:-tau], y[tau:]))
-        diff_x = np.diff(m[:, 0])
-        diff_y = np.diff(m[:, 1])
-        # Handle division by zero
-        with np.errstate(divide='ignore', invalid='ignore'):
-            theta = diff_y / diff_x
-        theta = np.arctan(theta)
+        # theta_t = arctan((y[t+tau+1]-y[t+tau]) / (y[t+1]-y[t])) = arctan(dy[t+tau]/dy[t])
+        with np.errstate(divide="ignore", invalid="ignore"):
+            theta = np.arctan(dy[tau:] / dy[:-tau])
 
-        if len(theta) == 0:
-            logger.warning(f'Time series (N={len(y)}) too short for embedding')
+        if theta.size == 0:
+            logger.warning(f"Time series (N={len(y)}) too short for embedding")
             return np.nan
 
-        stats_store[0, i] = autocorr(theta, 1, 'Fourier')[0]
-        stats_store[1, i] = autocorr(theta, 2, 'Fourier')[0]
-        stats_store[2, i] = autocorr(theta, 3, 'Fourier')[0]
-    # Compute output statistics
-    out = {
-        'ac1_thetaac1': autocorr(stats_store[0, :], 1, 'Fourier')[0],
-        'ac1_thetaac2': autocorr(stats_store[1, :], 1, 'Fourier')[0],
[…]
