### 2026-07-18 62cec99
further optimizations
