+++ pyhctsa/operations/physics.py
@@ -46,8 +46,8 @@ def walker(y: ArrayLike, walker_rule: str = 'prop',
     dict
         Summaries of the walker's trajectory and its relationship to the series.
     """
-    N = len(y)
     y = np.asarray(y, dtype=float)
+    N = len(y)
 
     # Default values and type requirements for each rule
     WALKER_CONFIGS = {
@@ -91,54 +91,64 @@ def walker(y: ArrayLike, walker_rule: str = 'prop',
     # Do the walk
     # ------------------------------------------------------------------
     w = np.zeros(N)
+    yl = y.tolist()
 
     if walker_rule == 'prop':
         # walker narrows the gap between its position and the series value
         # by the proportion walker_params at each step
         p = walker_params
-        w[0] = 0
+        w_prev = 0.0  # w[0] = 0
         for i in range(1, N):
-            w[i] = w[i-1] + p * (y[i-1] - w[i-1])
+            w_prev = w_prev + p * (yl[i-1] - w_prev)
+            w[i] = w_prev
 
     elif walker_rule == 'biasprop':
         # biased motion: [p_up, p_down]
         pup, pdown = walker_params
-        w[0] = 0
+        w_prev = 0.0  # w[0] = 0
+        yprev = yl[0]
         for i in range(1, N):
-            if y[i] > y[i-1]:  # time series increases
-                w[i] = w[i-1] + pup * (y[i-1] - w[i-1])
-            else:
-                w[i] = w[i-1] + pdown * (y[i-1] - w[i-1])
+            ycur = yl[i]
+            p = pup if ycur > yprev else pdown  # time series increases -> pup
+            w_prev = w_prev + p * (yprev - w_prev)
+            w[i] = w_prev
+            yprev = ycur
 
     elif walker_rule == 'momentum':
         # walker moves with inertia; the series acts as a force
         m = walker_params  # 'inertial mass'
-        w[0] = y[0]
-        w[1] = y[1]
+        w0 = yl[0]
+        w1 = yl[1]
+        w[0] = w0
+        w[1] = w1
+        w_pp = w0    # w[i-2]
+        w_prev = w1  # w[i-1]
         for i in range(2, N):
-            w_inert = w[i-1] + (w[i-1] - w[i-2])
-            w[i] = w_inert + (y[i] - w_inert) / m  # dissipative term
+            w_inert = w_prev + (w_prev - w_pp)
+            cur = w_inert + (yl[i] - w_inert) / m  # dissipative term
+            w[i] = cur
+            w_pp = w_prev
+            w_prev = cur
 
     elif walker_rule == 'runningvar':
-        # inertial motion rescaled by local standard deviation
         m, wl = walker_params
         wl = int(wl)
         w[0] = y[0]
         w[1] = y[1]
+        w_pp = yl[0]    # w[i-2]
+        w_prev = yl[1]  # w[i-1]
         for i in range(2, N):
-            w_inert = w[i-1] + (w[i-1] - w[i-2])
-            w_mom = w_inert + (y[i] - w_inert) / m  # dissipative term
-            # MATLAB: if im > wl, with im the 1-based index (= i + 1 here).
+            w_inert = w_prev + (w_prev - w_pp)
+            w_mom = w_inert + (yl[i] - w_inert) / m  # dissipative term
             if i + 1 > wl:
-                # MATLAB windows are y(im-wl:im) / w(im-wl:im): inclusive of the
[…]
