+++ pyhctsa/operations/wavelet.py
@@ -46,8 +46,6 @@ def _custom_cwt(data, scales, wavelet, sampling_period=1., method='conv', axis=-
         The method used to compute the CWT. Can be any of:
             - ``conv`` uses ``numpy.convolve``.
             - ``fft`` uses frequency domain convolution.
-            - ``auto`` uses automatic selection based on an estimate of the
-              computational complexity at each scale.
 
         The ``conv`` method complexity is ``O(len(scale) * len(data))``.
         The ``fft`` method is ``O(N * log2(N))`` with
