### 2026-08-05 0fb81da
add stepdetect feature
