+++ pyhctsa/toolboxes/Max_Little/ML_l1pwc.py
@@ -1,0 +1,314 @@
+"""Discrete total variation denoising (TVD) via a primal-dual interior-point solver.
+
+Python port of ``ML_l1pwc.m`` from Max Little's steps_bumps_toolkit.
+
+Minimizes
+
+    E = (1/2)||y - x||_2^2 + lambda*||Dx||_1
+
+over x, given input signal y, for each value of the regularization parameter
+lambda > 0.  D is the first difference matrix.  Hot-restarts from each value of
+lambda speed up convergence for subsequent values; best use of this feature is
+made by ensuring the chosen lambda values are close to each other.
+
+(c) Max Little, 2010.  Based around code originally written by S.J. Kim, K. Koh,
+S. Boyd and D. Gorinevsky.  If you use this code for your research, please cite:
+
+    M.A. Little, Nick S. Jones (2010) "Sparse Bayesian Step-Filtering for
+    High-Throughput Analysis of Molecular Machine Dynamics", in 2010 IEEE
+    International Conference on Acoustics, Speech and Signal Processing,
+    ICASSP 2010 Proceedings.
+
+Released under the terms of the GNU General Public License as published by the
+Free Software Foundation; version 2 or later.
+"""
+
+import numpy as np
+from scipy.linalg import solveh_banded, solve_banded
+from scipy.linalg.blas import ddot, dnrm2
+
+__all__ = ["l1pwc"]
+
+
+def _fmt(spec, *vals):
+    """
+    printf with MATLAB's spelling of the non-finite values.
+    """
+    return (spec % vals).replace("inf", "Inf").replace("nan", "NaN")
+
+
+def _d_mul(v):
+    """D*v for v of length n, returns length m = n-1.
+
+    Column order for row i is (col i, col i+1): 0 + v[i] - v[i+1].
+    """
+    return v[:-1] - v[1:]
+
+
+def _dt_mul(z):
+    """D'*z for z of length m, returns n.
+
+    Column j of D has entries at rows j-1 (-1) and j (+1), ascending, so
+    out[j] = -z[j-1] + z[j] with the appropriate end truncations.
+    """
+    m = z.shape[0]
+    out = np.empty(m + 1, dtype=np.float64)
+    out[0] = z[0]
+    out[1:m] = -z[:-1] + z[1:]
+    out[m] = -z[m - 1]
+    return out
+
+
+def _ddt_mul(z):
+    """DDT*z for z of length m.
+
+    Row i draws from columns i-1 (-1), i (+2), i+1 (-1) in that order:
+    out[i] = (-z[i-1] + 2*z[i]) - z[i+1].
+    """
+    m = z.shape[0]
+    out = np.empty(m, dtype=np.float64)
+    out[0] = 0.0
+    out[1:] = -z[:-1]
+    out += 2.0 * z
+    out[:-1] -= z[1:]
+    return out
+
+
+def _solve_tridiag(diag, b):
+    """Solve T*x = b where T is symmetric tridiagonal with the given diagonal
+    and -1 on both off-diagonals.
+    """
+    m = diag.shape[0]
+    if m == 1:
+        return b / diag[0]
+    # Lower-form banded storage for solveh_banded: row 0 = diagonal,
+    # row 1 = first subdiagonal (last entry unused).
+    ab = np.empty((2, m), dtype=np.float64)
+    ab[0] = diag
+    ab[1, : m - 1] = -1.0
+    ab[1, m - 1] = 0.0
+    try:
+        return solveh_banded(ab, b, lower=True, check_finite=False)
+    except np.linalg.LinAlgError:
+        ab_gen = np.zeros((3, m), dtype=np.float64)
+        ab_gen[0, 1:] = -1.0
+        ab_gen[1] = diag
+        ab_gen[2, : m - 1] = -1.0
[…]
