+++ pyhctsa/operations/changepoint.py
@@ -1,0 +1,215 @@
+import numpy as np
+
+from ..toolboxes.Max_Little.ML_l1pwc import l1pwc, l1pwc_lmax
+
+def stepdetect(y: np.ndarray, method: str = 'l1pwc', params: float | int = 10) -> dict:
+    """
+    Analysis of discrete steps in a time series.
+
+    Gives information about discrete steps in the signal, using the function
+    `l1pwc` from Max A. Little's step detection toolkit.
+
+    References
+    ----------
+    .. [1] S.J. Kim, K. Koh, S. Boyd and D. Gorinevsky (2007) "An Interior-Point Method for Large-Scale 
+        ℓ1 -Regularized Least Squares," in IEEE Journal of Selected Topics in Signal Processing, 
+        vol. 1, no. 4, pp. 606-617, Dec. 2007, doi: 10.1109/JSTSP.2007.910971. 
+    .. [2] M.A. Little, Nick S. Jones (2010) "Sparse Bayesian Step-Filtering for
+        High-Throughput Analysis of Molecular Machine Dynamics", in 2010 IEEE
+        International Conference on Acoustics, Speech and Signal Processing,
+        ICASSP 2010 Proceedings.
+
+    Parameters
+    ----------
+    y : array-like
+        The input time series.
+    method : str
+        The step-detection method. Currently only `l1pwc` is supported.
+    params : float or str
+        The parameters for the given method used. For `l1pwc` this corresponds to lambda. Default is 10.
+
+    Returns
+    -------
+    dict
+        Statistics on the output of the step-detection method, including the intervals
+        between change points, the proportion of constant segments, the reduction in
+        variance from removing the piece-wise constants, and stationarity in the
+        occurrence of change points.
+    """
+
+    y = np.asarray(y).ravel()
+    N = len(y)
+
+    out = {}
+    if method == 'l1pwc':
+        lamb = params
+        if lamb < 1:  # specify as a proportion of lambdamax
+            lamb = l1pwc_lmax(y) * lamb
+        steppedy, E, s, lambda_max = l1pwc(y, lamb, False)  # defaults for stoptol/maxiter
+
+        # l1pwc returns x as N x L; collapse to 1-D so diff/subtraction behave
+        steppedy = np.asarray(steppedy).ravel()
+        E = np.atleast_1d(E)
+        s = np.atleast_1d(s)
+
+        # Round to remove numerical fluctuations of order less than 1e-4
+        steppedy = np.round(steppedy * 1e4) / 1e4
+        out = {
+            "E": E[0],
+            "s": s[0],          # collapses to 1 for some parameter values
+            "lambdamax": lambda_max,
+        }
+        # Get step indices from steppedy; these give the index of the start of
+        # each run. Kept 1-based to match MATLAB, since every statistic below
+        # was written against 1-based indices.
+        whch = np.flatnonzero(np.diff(steppedy) != 0)
+        if whch.size:
+            chpts = np.concatenate(([1], whch + 2))
+        else:
+            chpts = np.array([1])   # no changes
+    elif method == 'kv':
+        raise NotImplementedError('step detection method not yet implemented.')
+    else:
+        raise ValueError(f'unknown step detection method: {method}')
+
+    num_change_points = len(chpts)
+
+    # Intervals -- of change
+    chints = np.diff(np.concatenate([chpts, [N]]))
+
+    # Number of constant segments per sample
+    out["nsegments"] = num_change_points / N  # will be 1 if there are no changes
+
+    # How much reduces variance
[…]
