+++ pyhctsa/configurations/hctsa.yaml
@@ -1616,4 +1616,21 @@ nonlinearity:
     legacy_name: NL_embed_PCA
     ordered_args: ['tau', 'm']
 
-    
+changepoint:
+  stepdetect:
+    base_name: step_detect
+    depedencies:
+    configs:
+      - {method: 'l1pwc', params: 10, zscore: True}
+      - {method: 'l1pwc', params: 0.05, zscore: True}
+      - {method: 'l1pwc', params: 0.2, zscore: True}
+    legacy_name: CP_ML_StepDetect
+    ordered_args: ['method', 'params']
+  
+  l1pwc_sweep_lambda:
+    base_name: l1pwc_sweep_lambda
+    depedencies:
+    configs:
+      - {lambdar: [!range [0, 0.95, 0.05]], zscore: True}
+    legacy_name: CP_l1pwc_sweep_lambda
+    ordered_args: ['lambdar']
