### 2026-08-06 77352b2
drop requirements + add new range constructor for configs
