+++ pyhctsa/operations/scaling.py
@@ -4,6 +4,7 @@ from numpy.typing import ArrayLike
 import numpy as np
 from scipy.stats import iqr
 from scipy.interpolate import interp1d
+from scipy.linalg import qr, solve_triangular
 import statsmodels.api as sm
 import logging
 logger = logging.getLogger('pyhctsa')
@@ -268,4 +269,261 @@ def _robust_linear_fit(log_tt, log_ff, the_range, field_name):
     out[f'{field_name}se2'] = results.bse[1]  # standard error in slope
     out[f'{field_name}ssr'] = np.mean(results.resid ** 2)  # mean squares residual
     out[f'{field_name}resac1'] = autocorr(results.resid, 1, 'Fourier')[0]  # autocorr at lag 1
+    return out
+
+
+def _colon(base, step, limit):
+    base = float(base)
+    step = float(step)
+    limit = float(limit)
+    if step == 0 or (step > 0 and base > limit) or (step < 0 and base < limit):
+        return np.zeros(0)
+
+    n = int(np.floor((limit - base) / step + 0.5))
+    # Guard against the rounding above overshooting the limit:
+    while n > 0 and abs(base + n * step) > abs(limit) + abs(step) * 0.5:
+        n -= 1
+
+    i = np.arange(n + 1, dtype=float)
+    out = np.empty(n + 1)
+    lower = i <= n / 2.0
+    out[lower] = base + i[lower] * step
+    out[~lower] = limit - (n - i[~lower]) * step
+    out[0] = base
+    if n > 0:
+        out[n] = limit
+    return out
+
+def _linspace(d1, d2, n):
+    d1 = float(d1)
+    d2 = float(d2)
+    n1 = n - 1
+    if np.isinf((d2 - d1) * (n1 - 1)):
+        i = np.arange(n1 + 1, dtype=float)
+        y = d1 + (d2 / n1) * i - (d1 / n1) * i
+    else:
+        y = d1 + np.arange(n1 + 1, dtype=float) * (d2 - d1) / n1
+    if y.size:
+        if d1 == d2:
+            y[:] = d1
+        else:
+            y[n - 1] = d2
+    return y
+
+def _round(x):
+    x = np.asarray(x, dtype=float)
+    return np.sign(x) * np.floor(np.abs(x) + 0.5)
+
+def _polyfit(x, y, deg):
+    x = np.asarray(x, dtype=float).ravel()
+    y = np.asarray(y, dtype=float).ravel()
+
+    n = deg + 1
+    V = np.ones((x.size, n))
+    for j in range(deg - 1, -1, -1):
+        V[:, j] = x * V[:, j + 1]
+
+    Q, R, perm = qr(V, mode="economic", pivoting=True)
+
+    # Rank from the pivoted diagonal, using Matlab's mldivide tolerance:
+    diagR = np.abs(np.diag(R))
+    if diagR.size == 0 or diagR[0] == 0:
+        return np.zeros(n)
+    tol = max(V.shape) * np.spacing(diagR[0])
+    rank = int(np.sum(diagR > tol))
+
+    p = np.zeros(n)
+    p[perm[:rank]] = solve_triangular(
+        R[:rank, :rank], (Q.T @ y)[:rank], lower=False
+    )
+    return p
+
+def _polyval(p, x):
+    x = np.asarray(x, dtype=float)
+    y = np.zeros_like(x)
+    for coeff in p:
+        y = x * y + coeff
+    return y
+
+def _std(x, axis=None):
+    x = np.asarray(x, dtype=float)
+    n = x.size if axis is None else x.shape[axis]
+    if n < 2:
+        return np.zeros(()) if axis is None else np.zeros(
[…]
