+++ .github/workflows/run_cross_platform.yaml
@@ -26,7 +26,6 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install -e .
-          # If needed: pip install -r requirements.txt
 
       - name: Run extractor
         run: |
