+++ .github/workflows/run_unit_tests.yaml
@@ -32,7 +32,6 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install setuptools
-          pip install -r requirements.txt
           pip install -e .[test]
           pip install pytest-cov
 

[…]
