### 2026-08-07 4d6db89
update workflows to run unit tests on all branches
