+++ .github/workflows/run_cross_platform.yaml
@@ -2,8 +2,6 @@ name: Cross-platform consistency
 
 on:
   push:
-    branches: [ main ]
-  pull_request:
 
 jobs:
   run-per-os:
