### 2026-08-07 15c1583
add back updated hctsa config
