+++ pyhctsa/configurations/hctsa.yaml
@@ -1,1190 +1,1190 @@
-# medical:
-#   hrv_classic:
-#     base_name: hrv_classic
-#     dependencies:
-#     configs:
-#       - {zscore: True}
-#     legacy_name: MD_hrv_classic
+medical:
+  hrv_classic:
+    base_name: hrv_classic
+    dependencies:
+    configs:
+      - {zscore: True}
+    legacy_name: MD_hrv_classic
 
-#   pnn:
-#     base_name: pnn
-#     dependencies:
-#     configs:
-#       - {zscore: False}
-#     legacy_name: MD_pNN 
+  pnn:
+    base_name: pnn
+    dependencies:
+    configs:
+      - {zscore: False}
+    legacy_name: MD_pNN 
 
-#   pol_var:
-#     base_name: pol_var
-#     dependencies:
-#     configs:
-#       - {d: 1.0, D: 3, zscore: True}
-#       - {d: 1.0, D: 4, zscore: True}
-#       - {d: 1.0, D: 5, zscore: True}
-#       - {d: 1.0, D: 6, zscore: True}
+  pol_var:
+    base_name: pol_var
+    dependencies:
+    configs:
+      - {d: 1.0, D: 3, zscore: True}
+      - {d: 1.0, D: 4, zscore: True}
+      - {d: 1.0, D: 5, zscore: True}
+      - {d: 1.0, D: 6, zscore: True}
 
-#       - {d: 0.5, D: 3, zscore: True}
-#       - {d: 0.5, D: 4, zscore: True}
-#       - {d: 0.5, D: 5, zscore: True}
-#       - {d: 0.5, D: 6, zscore: True}
+      - {d: 0.5, D: 3, zscore: True}
+      - {d: 0.5, D: 4, zscore: True}
+      - {d: 0.5, D: 5, zscore: True}
+      - {d: 0.5, D: 6, zscore: True}
 
-#       - {d: 0.1, D: 3, zscore: True}
-#       - {d: 0.1, D: 4, zscore: True}
-#       - {d: 0.1, D: 5, zscore: True}
-#       - {d: 0.1, D: 6, zscore: True}
-#     legacy_name: MD_polvar
-#     ordered_args: ['d', 'D']
+      - {d: 0.1, D: 3, zscore: True}
+      - {d: 0.1, D: 4, zscore: True}
+      - {d: 0.1, D: 5, zscore: True}
+      - {d: 0.1, D: 6, zscore: True}
+    legacy_name: MD_polvar
+    ordered_args: ['d', 'D']
 
-#   raw_hrv_meas:
-#     base_name: hrv_meas
-#     dependencies:
-#     configs:
-#       - {zscore: False}
-#     legacy_name: MD_rawHRVmeas
+  raw_hrv_meas:
+    base_name: hrv_meas
+    dependencies:
+    configs:
+      - {zscore: False}
+    legacy_name: MD_rawHRVmeas
     
-# extreme_events:
-#   moving_threshold:
-#     base_name: moving_threshold
-#     dependencies:
-#     configs:
-#       - {a: 0.1, b: 0.1, zscore: True}
-#       - {a: 0.1, b: 0.02, zscore: True}
-#       - {a: 1.0, b: 0.1, zscore: True}
-#       - {a: 1.0, b: 0.02, zscore: True}
-#     legacy_name: EX_MovingThreshold
-#     ordered_args: ['a', 'b']
+extreme_events:
+  moving_threshold:
+    base_name: moving_threshold
+    dependencies:
+    configs:
+      - {a: 0.1, b: 0.1, zscore: True}
+      - {a: 0.1, b: 0.02, zscore: True}
+      - {a: 1.0, b: 0.1, zscore: True}
+      - {a: 1.0, b: 0.02, zscore: True}
[…]
