### 2026-08-07 8f7df15
add nonlinearity functions
