+++ pyhctsa/operations/nonlinearity.py
@@ -6,9 +6,12 @@ import logging
 logger = logging.getLogger('pyhctsa')
 
 from sklearn.decomposition import PCA
+from scipy.linalg import lapack, qr, solve_triangular
+from sklearn.neighbors import NearestNeighbors
 
 from ..operations.model_fit import residual_analysis
-from ..operations.correlation import first_crossing, first_min
+from ..operations.correlation import first_crossing, first_min, autocorr
+from ..toolboxes.Tisean_3_0_1 import tisean as _tisean
 
 def _first_fn(p, threshold, over_or_under='under'):
     if over_or_under == 'under':
@@ -356,3 +359,552 @@ def embed_pca(y: ArrayLike, tau: Union[str, int] = 'ac', m: int = 3) -> dict:
     out['fb001'] = _first_fn(perc, 0.01, 'under')
 
     return out
+
+def _time_delay_embed(y: ArrayLike, tau: Union[str, int], m: int) -> np.ndarray:
+    """
+    Time-delay embedding of a univariate time series into an m-dimensional space.
+    """
+    y = np.asarray(y, dtype=float)
+    n = len(y)
+
+    if isinstance(tau, str):
+        if tau == 'ac':
+            tau = first_crossing(y, 'ac', 0, 'discrete')
+            if np.isnan(tau):
+                raise ValueError('Could not get time delay by ACF (time series too short?)')
+            tau = int(tau)
+        else:
+            raise ValueError(f"Invalid time-delay method: '{tau}'. Only 'ac' (or an integer) is supported.")
+    tau = int(tau)
+    m = int(m)
+
+    n_embed = n - (m-1)*tau
+    if n_embed <= 0:
+        raise ValueError(f'Time series (N = {n}) too short to embed with these embedding parameters.')
+
+    y_embed = np.zeros((n_embed, m))
+    for i in range(m):
+        y_embed[:, i] = y[i*tau : n_embed + i*tau]
+
+    return y_embed
+
+def local_density(y: ArrayLike, nnr: int = 3, past: int = 40,
+                  tau: Union[str, int] = 'ac', m: int = 2) -> dict:
+    """
+    Local density estimates in the time-delay embedding space.
+    
+    Computes a standard k-nearest-neighbor local density estimate at each
+    point of the time-delay embedding: density(i) is proportional to
+    1/r_NNR(i)^m, where r_NNR(i) is the distance from point i to its NNR-th
+    nearest neighbor (excluding temporally-close points within a Theiler
+    window of "past" samples) and m is the embedding dimension. This
+    operation previously used TSTOOL's 'localdensity', which the original
+    author noted was "very poorly documented in the TSTOOL package" -- its
+    exact algorithm was never confirmed, only assumed to be some form of
+    local density estimate in the embedding space, which is what's computed
+    here natively (no toolbox dependency at all). The missing normalizing
+    constant (relating 1/r^m to a true probability density) is the same for
+    every point in a given call, so it cancels out of all of the relative
+    statistics below (min/max/std/mean/median/autocorrelation).
+
+    Parameters
+    ----------
+    y : array-like
+        Input time series.
+    nnr : int, optional
+        Number of nearest neighbours to compute. Default is 3.
+    past : int, optional
+        Number of time-correlated points to discard (samples), i.e., the
+        Theiler window. Default is 40.
+    tau : str or int, optional
+        The time-delay of the embedding, either an integer or ``'ac'`` for the
[…]
