### 2026-08-07 8789848
add nonlinearity functions
