+++ pyhctsa/toolboxes/Tisean_3_0_1/TS_d2.c
@@ -1,0 +1,561 @@
+/*
+ *   Python C-extension wrapper around the correlation-sum kernel of TISEAN's
+ *   `d2` program (TISEAN 3.0.1, source_c/d2.c, author Rainer Hegger).
+ *
+ *   The original is a command-line program: it parses argv into file-scope
+ *   globals, reads the series from a text file, and writes the .c2/.d2/.h2/.stat
+ *   files itself (calling exit() when the smallest length scale is exhausted).
+ *   None of that survives repeated in-process calls, so the algorithm has been
+ *   made re-entrant here: every global lives in a `d2_ctx`, the series arrives
+ *   as a numpy array, and the raw pair counts are handed back to Python, which
+ *   assembles the c2/d2/h2 tables using the same expressions d2.c prints.
+ *
+ *   The numerics are a line-by-line transcription of d2.c -- including the
+ *   integer truncation in the length-scale index and the hand-rolled PRNG in
+ *   scramble() -- so the counts are identical to those of a stock TISEAN build.
+ *   The one deliberate change is that scramble() always uses the 64-bit
+ *   constants (d2.c picks them via sizeof(long)), so results no longer depend
+ *   on the platform's `long` width.
+ *
+ *   TISEAN is Copyright (c) 1998-2007 Rainer Hegger, Holger Kantz,
+ *   Thomas Schreiber, and is distributed under the GNU General Public License
+ *   version 2 or later.
+ */
+
+#include <Python.h>
+#include <numpy/arrayobject.h>
+
+#include <math.h>
+#include <stdint.h>
+#include <stdlib.h>
+#include <string.h>
+
+/* Size of the field for box assisted neighbour searching (power of 2) */
+#define NMAX 256
+/* Size of the box for the scramble routine */
+#define SCBOX 4096
+
+typedef struct {
+    const double *series;      /* single-column time series (DIM == 1)       */
+    int64_t length;
+
+    unsigned int DIM;          /* kept for fidelity with d2.c; always 1 here */
+    unsigned int EMBED;
+    unsigned int HOWOFTEN;
+    unsigned int DELAY;
+    int64_t MINDIST;           /* Theiler window                             */
+    double MAXFOUND;           /* compared against the (double) pair counts  */
+
+    double EPSMAX, EPSMIN, EPSMAX1;
+    double epsinv, epsfactor, lneps, lnfac;
+
+    int imax, howoften1, imin;
+
+    int64_t *scr, *oscr, *list, *listc1;
+    int64_t *box;              /* NMAX * NMAX, row-major                     */
+    int64_t *boxc1;
+    double *hs;                /* scratch delay vector, EMBED * DIM          */
+    double **found;            /* EMBED*DIM rows of HOWOFTEN counters        */
+    double *found_flat;
+    double *norm;
+    double *epsm;
+} d2_ctx;
+
+static void ctx_free(d2_ctx *c)
+{
+    free(c->scr);
+    free(c->oscr);
+    free(c->list);
+    free(c->listc1);
+    free(c->box);
+    free(c->boxc1);
+    free(c->hs);
+    free(c->found);
+    free(c->found_flat);
+    free(c->norm);
+    free(c->epsm);
+    memset(c, 0, sizeof(*c));
+}
+
+/* ---------------------------------------------------------------------------
+ * scramble(): d2.c's deterministic shuffle of the centre-point order.
+ * ------------------------------------------------------------------------ */
[…]
