+++ pyhctsa/toolboxes/Tisean_3_0_1/tisean.py
@@ -1,0 +1,313 @@
+"""
+Python front-end to the TISEAN 3.0.1 routines used by pyhctsa.
+
+hctsa drives TISEAN through the shell: it writes the time series to a temporary
+file, runs ``d2``, and then pipes the resulting ``.c2`` file through ``c2g`` and
+``c2t``.  Here the same three steps happen in-process:
+
+* :func:`d2` calls the C kernel in ``TS_d2.c`` (a re-entrant transcription of
+  TISEAN's ``d2.c``) and assembles the ``.c2``/``.d2``/``.h2`` tables from the
+  raw pair counts, using the expressions ``d2.c`` prints.
+* :func:`c2g` and :func:`c2t` are ports of ``source_f/c2g.f`` and
+  ``source_f/c2t.f``.  Those are Fortran, so they are reimplemented rather than
+  wrapped -- both are short, and this keeps the package free of a Fortran
+  toolchain.
+
+TISEAN is Copyright (c) 1998-2007 Rainer Hegger, Holger Kantz, Thomas
+Schreiber, and is distributed under the GNU General Public License v2 or later.
+"""
+
+from __future__ import annotations
+
+import math
+from typing import List, Optional
+
+import numpy as np
+from numpy.typing import ArrayLike
+
+from . import d2 as _d2_c
+
+__all__ = ["d2", "c2g", "c2t"]
+
+
+# 15-point Gauss-Kronrod rule, as tabulated in SLATEC's dqk15.f (which is what
+# c2g.f integrates with).  Only the positive abscissae are listed there.
+_XGK = np.array([
+    0.9914553711208126, 0.9491079123427585, 0.8648644233597691,
+    0.7415311855993945, 0.5860872354676911, 0.4058451513773972,
+    0.2077849550078985, 0.0,
+])
+_WGK = np.array([
+    0.02293532201052922, 0.06309209262997855, 0.10479001032225018,
+    0.14065325971552592, 0.16900472663926790, 0.19035057806478540,
+    0.20443294007529889, 0.20948214108472782,
+])
+# Mirrored to the full 15 nodes/weights on (-1, 1).
+_GK_X = np.concatenate([-_XGK[:7], [0.0], _XGK[6::-1]])
+_GK_W = np.concatenate([_WGK[:7], [_WGK[7]], _WGK[6::-1]])
+
+# c2g.f / c2t.f hold at most this many length scales per embedding dimension.
+_MEPS = 1000
+
+
+def _e(x: float) -> float:
+    """Round through C's ``%e`` -- what TISEAN writes and hctsa reads back."""
+    return float("%e" % x)
+
+
+def _round_significant(y: np.ndarray, digits: int) -> np.ndarray:
+    """Round to `digits` significant figures, as ``dlmwrite('precision', n)``."""
+    fmt = "%%.%dg" % digits
+    return np.array([float(fmt % v) for v in y], dtype=float)
+
+
+def d2(
+    y: ArrayLike,
+    delay: int = 1,
+    embed: int = 10,
+    theiler: int = 0,
+    howoften: int = 100,
+    maxfound: int = 1000,
+    epsmax: Optional[float] = None,
+    epsmin: Optional[float] = None,
+    write_precision: Optional[int] = 7,
+) -> dict:
+    """
+    Estimate correlation sums, dimensions and entropies (TISEAN's ``d2``).
+
+    Parameters
+    ----------
+    y : array-like
+        Input time series.
+    delay : int, optional
+        Time delay of the embedding (``d2 -d``). Default is 1.
+    embed : int, optional
+        Maximum embedding dimension (``d2 -M1,<embed>``). Default is 10.
+    theiler : int, optional
[…]
