### 2026-08-07 d29fadc
update docs with new methods
