### 2026-08-07 14d3277
update docs and function name mappings
