+++ pyhctsa/operations/nonlinearity.py
@@ -12,6 +12,23 @@ from ..operations.model_fit import residual_analysis
 from ..operations.correlation import first_crossing, first_min, autocorr
 from ..toolboxes.Tisean_3_0_1 import tisean as _tisean
 
+def _resolve_time_delay(y: ArrayLike, tau: Union[int, str]) -> Union[int, float]:
+    """Resolve a string time-delay spec to a lag.
+
+    ``'ac'`` uses the first zero-crossing of the autocorrelation function and
+    ``'mi'`` the first minimum of the automutual information. An integer is
+    returned unchanged. The resolved value may be NaN (time series too short);
+    callers are responsible for handling that.
+    """
+    if not isinstance(tau, str):
+        return tau
+    if tau == 'ac':
+        return first_crossing(y, 'ac', 0, 'discrete')
+    if tau == 'mi':
+        return first_min(y, 'mi')
+    raise ValueError(f'Invalid time-delay method: {tau}. Choose either mi or ac.')
+
+
 def _first_fn(p, threshold, over_or_under='under'):
     if over_or_under == 'under':
         indices = np.where(p < threshold)[0]
@@ -75,7 +92,6 @@ def _ms_embed(z, v, w):
         lags = np.array([0, 1, 2])
         auto_neg = False
     elif w is not None:
-        # MS_embed(z, dim, lag)  →  lags = 0 : w : w*(v-1)
         v = int(v)
         lags = np.arange(0, w * v, w)          # length v
         auto_neg = False
@@ -242,13 +258,8 @@ def nlpe(y: ArrayLike, de: int = 3, tau: Union[int, str] = 1, max_n: int = 5000)
     n = len(y)
 
     if isinstance(tau, str):
-        if tau == 'ac':
-            tau = first_crossing(y, 'ac', 0, 'discrete')
-        elif tau == 'mi':
-            tau = first_min(y, 'mi')
-        else:
-            raise ValueError("tau can be either 'mi' or 'ac'")
-        # check the tau 
+        tau = _resolve_time_delay(y, tau)
+        # check the tau
         if np.isnan(tau):
             logger.warning('Time series cannot be embedded (too short?)')
             return np.nan
@@ -307,28 +318,16 @@ def embed_pca(y: ArrayLike, tau: Union[str, int] = 'ac', m: int = 3) -> dict:
         Various statistics summarizing the obtained eigenvalue distribution.
 
     """
-    n = len(y)
     if isinstance(tau, str):
-        if tau == 'ac':
-            tau = first_crossing(y, 'ac', 0, 'discrete')
-            if np.isnan(tau):
-                logger.warning('Could not get time delay by ACF (time series too short?)')
-                return np.nan
-        elif tau == 'mi':
-            tau = first_min(y, 'mi')
-            if np.isnan(tau):
-                logger.warning('Could not get time delay by mutual information (time series too short?)')
-                return np.nan
-        else:
-            raise ValueError(f'Invalid time-delay method: {tau}. Choose either mi or ac.')
-    n_embed = n - (m-1)*tau
-    if n_embed <= 0:
-        logger.warning(f'Time series (N = {n}) too short to embed with these embedding parameters.')
+        tau = _resolve_time_delay(y, tau)
+        if np.isnan(tau):
+            logger.warning('Could not get time delay (time series too short?)')
+            return np.nan
+    try:
[…]
