+++ docs/source/methods/index.rst
@@ -117,7 +117,7 @@ Select a card below to explore a category of time-series analysis methods implem
 
    
    .. grid-item-card:: Changepoint
-      :link: _changepointmeths
+      :link: changepointmeths
       :link-type: ref
       :img-bottom: ../_static/module_icons/nonlinear.png
 
