### 2026-08-07 193636b
update icons for methods list
