### 2026-08-08 b75deac
update hctsa yaml
