+++ pyhctsa/operations/nonlinearity.py
@@ -30,56 +30,38 @@ def _resolve_time_delay(y: ArrayLike, tau: Union[int, str]) -> Union[int, float]
 
 
 def _first_fn(p, threshold, over_or_under='under'):
+    """Position (counting from one) of the first element of ``p`` on the given
+    side of ``threshold``, or ``len(p) + 1`` if there is none."""
     if over_or_under == 'under':
         indices = np.where(p < threshold)[0]
     elif over_or_under == 'over':
         indices = np.where(p > threshold)[0]
     else:
         raise ValueError(f'Unknown setting: {over_or_under}')
-    
-    return indices[0] if len(indices) > 0 else len(p) + 1
 
-def _normed_single_curve_length(x: ArrayLike, lag: int, fs: Union[float, int],
-                                nrmdegree: int) -> tuple:
+    return indices[0] + 1 if len(indices) > 0 else len(p) + 1
+
+def _normed_single_curve_length(x: np.ndarray, lag: int, nrmdegree: int) -> np.ndarray:
     # helper function for _normed_single_curve_length_windowed and nsamdf
-    # compute crossed curve length for a single time series x
-    x = np.asarray(x, dtype=float)
+    # crossed curve length of x at each delay 0, 1, ..., lag
     rx1 = np.zeros(lag+1)
-    for delay in range(0, lag+1):
-        if delay == 0:
-            rx1[delay] = np.linalg.norm(x - x, ord=nrmdegree)
-        else:
-            rx1[delay] = np.linalg.norm(x[:-delay] - x[delay:], ord=nrmdegree)
-    # mirror to form full symmetric sequence [rx1_reversed, rx1[1:]]
-    rx = np.concatenate([rx1[::-1], rx1[1:]])
-
-    pxy = np.abs(np.fft.fft(rx))
-    frq = np.linspace(0, 0.5 * fs, lag + 1)
-    return rx1, pxy, frq
+    for delay in range(1, lag+1):  # rx1[0] is the norm of a zero vector, i.e., zero
+        rx1[delay] = np.linalg.norm(x[:-delay] - x[delay:], ord=nrmdegree)
+    return rx1
 
 def _normed_single_curve_length_windowed(x: ArrayLike, win_len: int,
                                          shift_len: int, lag: int,
-                                         fs: Union[float, int], nrmdegree: int) -> tuple:
+                                         nrmdegree: int) -> np.ndarray:
     # helper function for nsamdf
-    # shiftlen is winlen - overlaplen
+    # mean curve length over sliding windows; shiftlen is winlen - overlaplen
+    x = np.asarray(x, dtype=float)
     m = int(np.floor((len(x) - win_len)/shift_len) + 1)
-    
-    p_sum = np.zeros(2*lag+1)
+
     r_sum = np.zeros(lag+1)
+    for start in range(0, m*shift_len, shift_len):
+        r_sum += _normed_single_curve_length(x[start:start+win_len], lag, nrmdegree)
 
-    for ii in range(1, m+1):
-        x_seg = x[(ii-1)*shift_len:(ii-1)*shift_len+win_len]
-        rx, px, _ = _normed_single_curve_length(x_seg, lag=lag, fs=fs, nrmdegree=nrmdegree)
-        p_sum += px
-        r_sum += rx
-    
-    p_mean = p_sum / m
-    r_mean = r_sum / m
-
-    p_mean2 = np.abs(np.fft.fft(np.concatenate([r_mean[::-1], r_mean[1:]])))
-    frq = np.linspace(0, 0.5 * fs, lag+1)
-
-    return r_mean, p_mean, frq, p_mean2
+    return r_sum / m
 
 def _ms_embed(z, v, w):
     # helper function for nlpe
[…]
