### 2026-08-08 ea35b6c
Consolidate three time-delay embedding implementations
`entropy._embed`, `nonlinearity._time_delay_embed` and `correlation._lag_embed`
were three copies of the same construction. The first two produce exactly equal
arrays; `_lag_embed` is the same embedding with its columns reversed.

Replace all three with `utils.time_delay_embed(y, m, tau, reverse=False)`.

The copies disagreed on how they signalled an over-short series: two raised
ValueError, `_lag_embed` logged a warning and returned a bare nan, which forced
`time_rev_kaplan` to test its result with `np.isscalar`. The shared helper always

[…]
