### 2026-08-08 b8b2a2b
Remove _prctile in favour of utils.matlab_quantile
`distribution._prctile` was a third implementation of the quantile already
provided by `utils.matlab_quantile` (and by numpy's 'hazen' method). Verified
zero maximum absolute difference against both across a range of percentages.

Its one caller, `outlier_test`, now takes both bounds from a single
`matlab_quantile` call; a stale comment noting the numpy equivalent goes with it.

Verified bit-identical over the 192-result baseline.

Co-Authored-By: Claude Opus 5 <noreply@anthropic.com>
