### 2026-08-08 502714f
Replace correlation._stat_av with stationarity.sliding_window
`_stat_av` reimplemented the subset of `sliding_window` covering window_stat in
{'mean','std'} with across_win_stat='std'; its nested `get_window` was a
byte-for-byte copy of `stationarity._get_window`. Verified bit-identical output
across both statistics and several segmentations.

Its six call sites in `translate_shape` collapse to a loop over num_seg. The
import is function-local because stationarity already imports from this module,
so a top-level import would close the cycle -- the same approach information.py
uses for its corr
[…]
