### 2026-08-08 6da3eb6
Remove unused _first_min_mi_gaussian
Zero references anywhere in the package, tests, configs or docs. Its behaviour
remains reachable via first_min(y, 'mi-gaussian').

This deletion was already present as an uncommitted working-tree edit; committing
it here as the last item of the Tier A dedupe.

Co-Authored-By: Claude Opus 5 <noreply@anthropic.com>
