### 2026-08-08 43b737c
Fix translate_shape crashing on numpy < 2.0
`np.unique_counts` was added in numpy 2.0, but pyproject.toml declares
numpy>=1.24, so translate_shape raised AttributeError on every call for any
environment inside the supported range (including the numpy 1.26 used here).

Use np.unique(..., return_counts=True), which is equivalent and available across
the whole declared range.

Verified against the previous behaviour by shimming np.unique_counts onto the
pre-fix code: 40 full translate_shape outputs (5 series x 2 shapes x 4 offsets)
are bit-identical, and the fixed version now runs unshimmed on nu
[…]
