### 2026-08-08 c4ec58d
add transition_p_alphabet port
