+++ pyhctsa/calculator.py
@@ -16,7 +16,7 @@ import yaml
 from numpy.typing import ArrayLike
 from rich.progress import Progress, TextColumn, BarColumn, TaskProgressColumn, TimeElapsedColumn
 
-from .utils import _check_optional_deps, _preprocess_decorator, _validate_data, _build_repr_html
+from .utils import _check_optional_deps, _preprocess_decorator, _validate_data
 from .distribute import _compute_features_for_chunk, _extract_features_single_series
 
 class RangeList(list):
@@ -262,9 +262,6 @@ class FeatureCalculator:
         self.feature_funcs = self._build_feature_funcs()
         print(f"Loaded {len(self.feature_funcs)} master operations.")
     
-    def _repr_html_(self):
-        return _build_repr_html(self.feature_funcs, self._skipped_functions, self.config, self.config_path)
-
     def _check_deps(self, module_key, feature_name, config):
         raw_deps = config.get("dependencies", None)
         if not raw_deps:
