+++ pyhctsa/toolboxes/matlab/matlab_fit.py
@@ -1,0 +1,486 @@
+"""
+A port of the nonlinear least-squares machinery behind MATLAB's ``fit(x, y, fittype(...))``.
+
+This module ports that path for the unconstrained case, which is all the Curve Fitting
+Toolbox uses when no ``Lower``/``Upper`` bounds are given. Without finite bounds the
+reflective machinery collapses considerably: the scaling matrix ``D`` is the identity,
+``DG`` vanishes, no step is ever truncated at a bound, and the reflected trial direction is
+never evaluated. What remains is the two-dimensional subspace trust-region iteration.
+"""
+
+import numpy as np
+from scipy.linalg import solve_triangular
+
+__all__ = ['lsqcurvefit_trr', 'fit_exp1', 'fit_poly1', 'goodness_of_fit']
+
+_EPS = np.finfo(float).eps
+_SQRT_EPS = np.sqrt(_EPS)
+
+
+# ------------------------------------------------------------------------------
+# Trust-region subproblem: min { g'*s + 0.5*s'*H*s : ||s|| <= delta }
+# ------------------------------------------------------------------------------
+
+def _seceqn(lam, eigval, alpha, delta):
+    """The secular equation 1/delta - 1/||s(lambda)||, evaluated at a scalar lambda."""
+    w = eigval + lam
+    with np.errstate(divide='ignore', invalid='ignore'):
+        m = np.where(w != 0, alpha / np.where(w != 0, w, 1), np.inf)
+    m = m * m
+    with np.errstate(divide='ignore', invalid='ignore'):
+        value = np.sqrt(1.0 / np.sum(m))
+    if np.isnan(value):
+        value = 0.0
+    return 1.0 / delta - value
+
+
+def _secular_eqn_root(x, itbnd, eigval, alpha, delta, tol):
+    """
+    Zero of the secular equation to the RIGHT of the starting point x.
+    A port of the modified `fzero` used by MATLAB's `trust`.
+    """
+    itfun = 0
+    dx = abs(x) / 2 if x != 0 else 0.5
+
+    a = x
+    c = a
+    fa = _seceqn(a, eigval, alpha, delta)
+    itfun += 1
+
+    b = x + 1
+    fb = _seceqn(b, eigval, alpha, delta)
+    itfun += 1
+
+    # Find change of sign
+    while (fa > 0) == (fb > 0):
+        dx = 2 * dx
+        if (fa > 0) != (fb > 0):
+            break
+        b = x + dx
+        fb = _seceqn(b, eigval, alpha, delta)
+        itfun += 1
+        if itfun > itbnd:
+            break
+
+    fc = fb
+    d = e = b - a
+    # Main loop, exit from middle of the loop
+    while fb != 0:
+        # Insure that b is the best result so far, a is the previous value of b,
+        # and c is on the opposite side of the zero from b
+        if (fb > 0) == (fc > 0):
+            c = a
+            fc = fa
+            d = b - a
+            e = d
+        if abs(fc) < abs(fb):
+            a, b, c = b, c, b
+            fa, fb, fc = fb, fc, fb
+
+        if itfun > itbnd:
+            break
+        m = 0.5 * (c - b)
+        toler = 2.0 * tol * max(abs(b), 1.0)
+        if abs(m) <= toler or fb == 0.0:
+            break
+
+        # Choose bisection or interpolation
+        if abs(e) < toler or abs(fa) <= abs(fb):
+            d = m
+            e = m
+        else:
+            s = fb / fa
+            if a == c:
+                p = 2.0 * m * s  # linear interpolation
+                q = 1.0 - s
[…]
