+++ pyhctsa/operations/symbolic.py
@@ -1,5 +1,5 @@
 import numpy as np
-from typing import Union
+from typing import Optional, Union
 from numpy.typing import ArrayLike
 import logging
 logger = logging.getLogger('pyhctsa')
@@ -8,7 +8,8 @@ from scipy.stats import mstats
 from scipy.signal import resample_poly
 
 from ..operations.correlation import first_crossing
-from ..utils import binarize, sign_change
+from ..toolboxes.matlab.matlab_fit import fit_exp1, fit_poly1, goodness_of_fit
+from ..utils import binarize, matlab_quantile, sign_change
 
 def surprise(y: ArrayLike, what_prior: str = 'dist', memory: float = 0.2, num_groups: int = 3,
              coarse_grain_method: str = 'quantile', num_iters: int = 500,
@@ -672,6 +673,280 @@ def transition_matrix(y: ArrayLike, how_to_cg: str = 'quantile',
     out['mineigcov'] = np.min(np.real(eig_cov_t))
 
     return out
+
+
+def _seq_sum(x: ArrayLike) -> complex:
+    """
+    Sequential summation, as MATLAB's `sum` performs it over a short vector.
+    NumPy sums pairwise instead, which can differ in the last bit -- enough to flip a
+    downstream `>` comparison against a threshold derived from these same sums.
+    """
+    total = 0.0
+    for v in np.asarray(x).ravel():
+        total = total + v
+    return total
+
+
+def _seq_mean(x: ArrayLike) -> complex:
+    """MATLAB's `mean`, i.e. its sequential `sum` divided by the count."""
+    x = np.asarray(x)
+    return _seq_sum(x) / x.size if x.size else np.nan
+
+
+def _seq_std(x: ArrayLike) -> float:
+    """
+    MATLAB's sample standard deviation: sqrt(sum(abs(x - mean(x)).^2)/(n-1)), where the
+    std of a scalar is 0 and the std of an empty vector is NaN. Accepts complex input,
+    for which it returns the real spread about the complex mean.
+    """
+    x = np.asarray(x)
+    if x.size == 0:
+        return np.nan
+    if x.size == 1:
+        return 0.0
+    xc = x - _seq_mean(x)
+    return float(np.sqrt(np.real(_seq_sum(xc * np.conj(xc))) / (x.size - 1)))
+
+
+def _exp_fit_gof(x: np.ndarray, y: np.ndarray, start_point) -> dict:
+    """
+    Fit ``a*exp(b*x)`` by nonlinear least squares and return the coefficients along
+    with the goodness-of-fit statistics reported by MATLAB's Curve Fitting Toolbox.
+    """
+    a, b = fit_exp1(x, y, start_point)
+    gof = goodness_of_fit(y, a * np.exp(b * x), num_coeffs=2)
+    return {'a': a, 'b': b, 'r2': gof['rsquare'],
+            'adjr2': gof['adjrsquare'], 'rmse': gof['rmse']}
+
+
+def _discretize_quantile(y: np.ndarray, num_groups: int) -> np.ndarray:
+    """
+    Discretize a time series into `num_groups` equiprobable groups by quantile separation.
+
+    This mirrors the (right-inclusive) binning of SB_TransitionpAlphabet's SUB_discretize,
+    which differs from the (left-inclusive) binning used by `coarse_grain`.
+    """
+    # thresholds for dividing the time-series values
+    th = matlab_quantile(y, np.linspace(0, 1, num_groups + 1))
+    th[0] = th[0] - 1  # ensures the first point is included
+
+    # turn the time series into a set of numbers from 1:num_groups
+    yth = np.zeros(len(y), dtype=int)
[…]
