### 2026-08-08 c290520
update add_noise
