### 2026-08-09 1fb0502
add gaussian process functions
