+++ pyhctsa/operations/symbolic.py
@@ -600,77 +600,71 @@ def transition_matrix(y: ArrayLike, how_to_cg: str = 'quantile',
         transition matrix.
     """
     # check inputs
-    y = np.asarray(y)
+    y = np.asarray(y, dtype=float)
     if num_groups < 2:
-        logger.warning("Too few groups for coarse-graining")
-        return np.nan
-    if tau == 'ac':
-        # determine the tau from first zero of the ACF
+        raise ValueError('Too few groups for coarse-graining')
+    if isinstance(tau, str):
+        if tau != 'ac':
+            raise ValueError(f"Unknown tau '{tau}'")
+        # determine tau from the first zero-crossing of the ACF
         tau = first_crossing(y, 'ac', 0, 'discrete')
-        if np.isnan(tau):
-            logger.warning("Time series too short to estimate tau")
-            return np.nan
-    if tau > 1:
-        y = resample_poly(y, 1, int(tau))
-    
-    N = len(y)
+    if np.isnan(tau):
+        raise ValueError('Time series too short to estimate tau')
+    tau = int(tau)
 
+    if tau > 1:  # calculate the transition matrix at a non-unit lag
+        y = resample_poly(y, 1, tau)  # downsample at rate 1:tau
+
+    N = len(y)  # time-series length
+
+    # (((1))) Discretize the time series to a symbolic string, containing
+    # integers from 1 to num_groups
     yth = coarse_grain(y, how_to_cg, num_groups)
-    # At this point we should have:
-    # (*) yth: a thresholded y containing integers from 1 to num_groups
-    yth = np.ravel(yth)
 
-    T = np.zeros((num_groups,num_groups))
-    for i in range(num_groups):
-        ri = (yth == i + 1)
-        if sum(ri) == 0:
-            T[i,:] = 0
-        else:
-            ri_next = np.r_[False, ri[:-1]]
-            for j in range(num_groups):
-                T[i, j] = np.sum(yth[ri_next] == j + 1)
+    # (((2))) Compute the tau-step transition matrix (Markov for tau = 1)
+    T = _transition_matrix(yth, num_groups)
 
+    # (((3))) Output measures from the transition matrix
     out = {}
-    # Normalize from counts to probabilities:
-    T = T/(N - 1) # N-1 is appropriate because it's a 1-time transition matrix
 
-    if num_groups == 2:
-        for i in range(4):
-            out[f'T{i+1}'] = T.transpose().flatten()[i]
-
-    elif num_groups == 3:
-        for i in range(9):
-            out[f'T{i+1}'] = T.transpose().flatten()[i]
-
-    elif num_groups > 3:
+    # (i) Raw values of the transition matrix; only for num_groups = 2, 3 are all
+    # elements returned (in MATLAB's column-major order), otherwise just the diagonal
+    if num_groups in (2, 3):
+        for i, v in enumerate(T.flatten(order='F')):
+            out[f'T{i+1}'] = v
+    else:
         for i in range(num_groups):
-            out[f'TD{i+1}'] = T.transpose()[i, i]
+            out[f'TD{i+1}'] = T[i, i]
 
     # (ii) Measures on the diagonal
-    out['ondiag'] = np.trace(T) # trace
-    out['stddiag'] = np.std(np.diag(T), ddof=1) # std of diagonal elements
+    diag_t = np.diag(T)
+    out['ondiag'] = _seq_sum(diag_t)  # trace
+    out['stddiag'] = _seq_std(diag_t)  # std of diagonal elements
 
     # (iii) Measures of symmetry:
[…]
