+++ pyhctsa/operations/model_fit.py
@@ -14,6 +14,8 @@ logger = logging.getLogger('pyhctsa')
 
 from ..operations.correlation import autocorr, first_crossing
 from ..operations.stationarity import sliding_window
+from ..toolboxes.matlab.gpml.gpml import CovSEisoNoise, gp_predict, gp_train
+from ..toolboxes.matlab.optimizers import minimize
 from ..utils import z_score, ljung_box_pvalue
 
 def hmm_fit(y: ArrayLike, train_p: float = 0.8, num_states: int = 3, random_seed: int = 0) -> dict:
@@ -833,3 +835,338 @@ def is_seasonal(y: ArrayLike) -> int:
         out = 1 # test thinks the time series has strong periodicities
     
     return out
+
+def _gp_learn_hyperp(tt: np.ndarray, yt: np.ndarray, cov, nfevals: int = -50) -> np.ndarray:
+    """
+    learn GP hyperparameters for the time series ``(tt, yt)``.
+
+    The GP is a mean-zero process with a Gaussian likelihood and Laplace
+    inference; ``nfevals`` is negative, so it caps the number of function
+    evaluations rather than the number of line searches.
+
+    Returns the flattened hyperparameter vector ``[cov..., lik]`` -- gpml
+    unwraps the hyperparameter struct with its fields alphabetised (cov, lik,
+    mean), and the mean is empty for a mean-zero process.
+
+    Raises ``numpy.linalg.LinAlgError`` if the covariance loses positive
+    definiteness, the counterpart of gpml's ``MATLAB:posdef`` error.
+    """
+    nhps = cov.n_hyp
+    # Initial values: the length parameter is in the ballpark of the difference
+    # between time elements; the remaining log-hyperparameters start at zero and
+    # the likelihood noise at log(0.1).
+    hyp0 = np.concatenate([[np.log(np.mean(np.diff(tt)))],
+                           np.zeros(nhps - 1), [np.log(0.1)]])
+
+    def _nlz(theta):
+        hyp = {'cov': theta[:nhps], 'lik': theta[nhps], 'mean': np.zeros(0)}
+        nlZ, dnlZ = gp_train(hyp, cov, tt, yt)
+        return nlZ, np.concatenate([dnlZ['cov'], dnlZ['lik'], dnlZ['mean']])
+
+    theta, _, _ = minimize(hyp0, _nlz, nfevals)
+    return theta
+
+
+def gp_fit_across(y: ArrayLike, cov_func: str = 'covSEiso_covNoise',
+                  npoints: int = 20) -> dict:
+    """
+    Gaussian Process time-series modeling for local prediction.
+
+    Trains a Gaussian Process model on equally-spaced points throughout the time
+    series and uses the model to predict its intermediate values.
+
+    Parameters
+    ----------
+    y : array-like
+        The input time series.
+    cov_func : str
+        The covariance function. Only ``'covSEiso_covNoise'``, the gpml
+        ``covSum`` of a squared exponential and a noise term, is supported -- it
+        is the only configuration hctsa instantiates. Default is
+        ``'covSEiso_covNoise'``.
+    npoints : int
+        The number of points through the time series to fit the GP model to.
+        Default is 20.
+
+    Returns
+    -------
+    dict
+        Dictionary summarising the error and the fitted hyperparameters.
+    """
+    if cov_func != 'covSEiso_covNoise':
+        raise ValueError(
+            "Only cov_func='covSEiso_covNoise' is supported "
+            f"(got {cov_func!r}); it is the only variant used by hctsa.")
+
+    y = np.asarray(y, dtype=float).ravel()
[…]
