### 2026-08-10 740d5b6
Simplify dependency installation in CI workflow
Removed redundant pip install commands for requirements and pytest-cov.
