+++ .github/workflows/run_unit_tests.yaml
@@ -32,9 +32,7 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install setuptools
-          pip install -r requirements.txt
-          pip install -e .[test]
-          pip install pytest-cov
+          pip install -e ".[test]"
 
       # Run normal tests everywhere EXCEPT the single coverage runner
       - name: Run pyhctsa unit tests
