+++ requirements.txt
@@ -1,14 +1,0 @@
-scipy
-numpy
-pyyaml
-statsmodels
-numba
-scikit-learn
-antropy
-PyWavelets
-ts2vg
-arch
-hmmlearn
-rich
-pathos
-lmfit
