### 2026-08-10 dad614a
add gaussian process functions
