### 2026-08-10 7080330
refactor model_fit
