+++ pyhctsa/operations/model_fit.py
@@ -146,7 +146,6 @@ def fit_subsegments(y: ArrayLike, model: str = 'ar', order: int = 2, subset_how:
     y = np.asarray(y)
     N = len(y)
     num_pred = sample_p[0]
-    r = np.zeros((num_pred, 2))
     if subset_how == 'uniform':
         if len(sample_p) == 1:  # size will depend on number of unique subsegments
             # num_pred+1 boundaries = num_pred portions
@@ -223,11 +222,7 @@ def loop_local_simple(y: ArrayLike, forecast_meth: str = 'mean') -> dict:
         raise ValueError(f"Unknown prediction method: {forecast_meth}")
     stats_st = np.zeros((len(train_length_range), 5))
     for i in range(len(train_length_range)):
-        outtmp = None
-        if forecast_meth == 'mean':
-            outtmp = local_simple(y, 'mean', train_length_range[i])
-        elif forecast_meth == 'median':
-            outtmp = local_simple(y, 'median', train_length_range[i])
+        outtmp = local_simple(y, forecast_meth, train_length_range[i])
         stats_st[i, 0] = outtmp['stderr']
         stats_st[i, 1] = outtmp['sws']
         stats_st[i, 2] = outtmp['swm']
@@ -261,25 +256,13 @@ def loop_local_simple(y: ArrayLike, forecast_meth: str = 'mean') -> dict:
         out['stderr_peakpos'] = np.nan
         out['stderr_peaksize'] = np.nan
 
-    #% (2) Sliding Window Stationarity
-    out['sws_chn'] = np.mean(np.diff(stats_st[:,1]))/(np.ptp(stats_st[:,1]))
-    out['sws_meansgndiff'] = np.mean(np.sign(np.diff(stats_st[:,1])))
-    out['sws_stdn'] = np.std(stats_st[:,1], ddof=1)/np.ptp(stats_st[:,1])
-
-    #% (3) sliding window mean
-    out['swm_chn'] = np.mean(np.diff(stats_st[:, 2]))/(np.ptp(stats_st[:, 2]))
-    out['swm_meansgndiff'] = np.mean(np.sign(np.diff(stats_st[:,2])))
-    out['swm_stdn'] = np.std(stats_st[:,2], ddof=1)/np.ptp(stats_st[:,2])
-    
-    # (4) AC1
-    out['ac1_chn'] = np.mean(np.diff(stats_st[:, 3]))/(np.ptp(stats_st[:, 3]))
-    out['ac1_meansgndiff'] = np.mean(np.sign(np.diff(stats_st[:,3])))
-    out['ac1_stdn'] = np.std(stats_st[:,3], ddof=1)/np.ptp(stats_st[:,3])
-
-    # (5) AC2
-    out['ac2_chn'] = np.mean(np.diff(stats_st[:, 4]))/(np.ptp(stats_st[:, 4]))
-    out['ac2_meansgndiff'] = np.mean(np.sign(np.diff(stats_st[:,4])))
-    out['ac2_stdn'] = np.std(stats_st[:,4], ddof=1)/np.ptp(stats_st[:,4])
+    #% (2)-(5) Curve statistics for the remaining metrics:
+    #%   sws (sliding window stationarity), swm (sliding window mean), ac1, ac2
+    for name, col in (('sws', 1), ('swm', 2), ('ac1', 3), ('ac2', 4)):
+        curve = stats_st[:, col]
+        out[f'{name}_chn'] = np.mean(np.diff(curve)) / np.ptp(curve)
+        out[f'{name}_meansgndiff'] = np.mean(np.sign(np.diff(curve)))
+        out[f'{name}_stdn'] = np.std(curve, ddof=1) / np.ptp(curve)
 
     return out
 
[…]
