### 2026-08-10 fd46e42
refactor physics
