### 2026-08-10 f322bc7
refactor entropy
