+++ pyhctsa/operations/entropy.py
@@ -16,6 +16,18 @@ from ..toolboxes.Max_Little import close_returns as _close_returns_c
 from ..toolboxes.physionet import sampen as _sampen_c
 from ..utils import bin_picker, histc, make_buffer, time_delay_embed, z_score
 
+
+def _entropy_summary(ents: np.ndarray) -> dict:
+    """Summary statistics on a set of entropy values."""
+    return {
+        'maxent': np.max(ents),
+        'minent': np.min(ents),
+        'medent': np.median(ents),
+        'meanent': np.mean(ents),
+        'stdent': np.std(ents, ddof=1),
+    }
+
+
 def shannon_entropy(
     y: ArrayLike,
     num_bins: Union[int, list[int]] = 2,
@@ -61,34 +73,22 @@ def shannon_entropy(
 
     if bin_range_size == 1:
         if depth_range_size == 1:
-            # %% Evaluate the shannon entropy of discretization - scales with depth, so it's nice to normalize by this factor
+            # Entropy scales with depth, so normalize by this factor
             out = shannon.entropy(y, int(num_bins), int(depth)) / int(depth)
         elif depth_range_size > 1:
-            # % Range over depths specified in the vector and return statistics on results
-            num_depths = depth_range_size
-            ents = np.zeros(num_depths)
-            for i in range(num_depths):
-                ents[i] = shannon.entropy(y, int(num_bins), int(depth[i])) / int(depth[i])
-            out = {}
-            #% Output statistics on variation across the range tested:
-            out['maxent'] = np.max(ents)
-            out['minent'] = np.min(ents)
-            out['medent'] = np.median(ents)
-            out['meanent'] = np.mean(ents)
-            out['stdent'] = np.std(ents, ddof=1)
+            # Range over depths and return statistics on the results
+            ents = np.array([
+                shannon.entropy(y, int(num_bins), int(d)) / int(d) for d in depth
+            ])
+            out = _entropy_summary(ents)
 
     elif bin_range_size > 1:
         if depth_range_size == 1:
-            #%% (*) Statistics over different bin numbers (constant depth)
-            ents = np.zeros(bin_range_size)
-            for i in range(bin_range_size):
-                ents[i] = shannon.entropy(y, int(num_bins[i]), int(depth))
-            out = {}
-            out['maxent'] = np.max(ents)
-            out['minent'] = np.min(ents)
-            out['medent'] = np.median(ents)
-            out['meanent'] = np.mean(ents)
-            out['stdent'] = np.std(ents, ddof=1)
+            # Statistics over different bin numbers (constant depth)
+            ents = np.array([
+                shannon.entropy(y, int(n), int(depth)) for n in num_bins
+            ])
+            out = _entropy_summary(ents)
         elif depth_range_size > 1:
             raise NotImplementedError("Comparing both bins and depth not implemented.")
 
@@ -142,12 +142,10 @@ def distribution_entropy(
         ]
         if y_hat.size == 0:
             return np.nan
-        else:
-            out = (
-                distribution_entropy(y, hist_or_ks, num_bins)
-                - distribution_entropy(y_hat, hist_or_ks, num_bins)
-            )
-            return out
+        return (
+            distribution_entropy(y, hist_or_ks, num_bins)
[…]
