+++ pyhctsa/operations/information.py
@@ -1,6 +1,5 @@
 import logging
 logger = logging.getLogger('pyhctsa')
-import os
 from typing import Any, Dict, List, Optional, Union, Callable
 
 import numpy as np
@@ -542,316 +541,117 @@ def rm_automutual_information(y: ArrayLike, tau: int = 1) -> float:
 
     return out
 
-def _rm_info(*args):
-    """
-    Estimates the mutual information of two stationary signals with independent 
-    pairs of samples using various approaches.
+def _rm_info(x: ArrayLike, y: ArrayLike):
+    """Unbiased mutual-information estimate of two equal-length signals via a 2D histogram.
+
+    Moddemeijer's histogram estimator (natural-log base, unbiased correction).
+    Original MATLAB by R. Moddemeijer; Python translation by Tucker Cullen.
 
     Parameters
     ----------
-    x : array-like
-        First input time series (row vector).
-    y : array-like
-        Second input time series (row vector).
-    descriptor : array-like, optional
-        Histogram descriptor array of shape (2, 3):
-            [[LOWERBOUNDX, UPPERBOUNDX, NCELLX],
-             [LOWERBOUNDY, UPPERBOUNDY, NCELLY]]
-        If not provided, will be computed automatically.
-    approach : {'unbiased', 'mmse', 'biased'}, optional
-        Method for estimating mutual information:
-            - 'unbiased': The unbiased estimate (default)
-            - 'mmse': The minimum mean square error estimate
-            - 'biased': The biased estimate
-    base : float, optional
-        The base of the logarithm (default: e).
+    x, y : array-like
+        Equal-length 1-D input vectors.
 
     Returns
     -------
     estimate : float
-        The mutual information estimate.
+        The (unbiased) mutual information estimate in nats.
     nbias : float
-        The N-bias of the estimate.
+        The N-bias of the estimate (0 after the unbiased correction).
     sigma : float
         The standard error of the estimate.
     descriptor : np.ndarray
-        The descriptor of the histogram used, see also _rm_histogram_2.
-
-    Notes
-    -----
-    - See also: http://www.cs.rug.nl/~rudy/matlab/
-    - Original MATLAB function by R. Moddemeijer, minor modifications by Ben Fulcher.
-    - Python translation by Tucker Cullen.
+        The histogram descriptor used (see :func:`_rm_histogram_2`).
     """
-    n_args = len(args)
-    if n_args < 1:
-        print("Takes in 2-5 parameters: ")
-        print("rm_information(x, y)")
-        print("rm_information(x, y, descriptor)")
-        print("rm_information(x, y, descriptor, approach)")
-        print("rm_information(x, y, descriptor, approach, base)")
-        print()
+    x = np.asarray(x)
+    y = np.asarray(y)
+    if x.ndim != 1 or y.ndim != 1:
+        raise ValueError("x and y must be 1-D vectors")
+    if x.size != y.size:
+        raise ValueError("Unequal length of x and y")
 
-        print("Returns a tuple containing: ")
-        print("estimate, nbias, sigma, descriptor")
-        return
+    h, descriptor = _rm_histogram_2(x, y)
+    n_cell_x = int(descriptor[0, 2])
+    n_cell_y = int(descriptor[1, 2])
 
-    # some initial tests on the input arguments
-    x = np.array(args[0])  # make sure the inputs are in numpy array form
-    y = np.array(args[1])
[…]
