### 2026-08-10 2b3acd4
optimizations to medical
