### 2026-08-10 de80400
optimize leastsq fit
