+++ pyhctsa/toolboxes/matlab/matlab_fit.py
@@ -1,95 +1,231 @@
 """
 A port of the nonlinear least-squares machinery behind MATLAB's
 ``fit(x, y, fittype(...))``.
-
-This module ports that path for the unconstrained case, which is all the Curve
-Fitting Toolbox uses when no ``Lower``/``Upper`` bounds are given. Without
-finite bounds the reflective machinery collapses considerably: the scaling
-matrix ``D`` is the identity, ``DG`` vanishes, no step is ever truncated at a
-bound, and the reflected trial direction is never evaluated. What remains is
-the two-dimensional subspace trust-region iteration.
 """
 
+import math
+
 import numpy as np
-from scipy.linalg import solve_triangular
+from numpy.linalg import LinAlgError
+from scipy.linalg import get_lapack_funcs
+
+_dtrtrs, _dgeqrf = get_lapack_funcs(
+    ("trtrs", "geqrf"),
+    (np.empty(0, dtype=np.float64),),
+)
+
+_TRIL_IDX = {}
+
+
+def _qr_r(A):
+    """
+    ``np.linalg.qr(A, mode="r")``.
+
+    ``qr`` copies ``A``, runs ``dgeqrf`` and returns ``triu`` of the leading
+    ``min(m, n)`` rows. ``dgeqrf`` here is the same LAPACK routine on the same
+    matrix, so the factor is identical bit for bit, and the explicit copy +
+    zero-fill below reproduces ``np.triu``'s C-contiguous result exactly.
+    """
+    qr, _tau, _work, _info = _dgeqrf(A)
+
+    m, n = A.shape
+
+    R = qr[: (n if m >= n else m), :].copy()
+
+    k = R.shape[0]
+
+    idx = _TRIL_IDX.get(k)
+
+    if idx is None:
+        idx = _TRIL_IDX[k] = np.tril_indices(k, -1)
+
+    R[idx] = 0.0
+
+    return R
+
+
+def _solve_tri(a1, b1, trans, overwrite_b=False):
+    """
+    ``scipy.linalg.solve_triangular(a1, b1, trans=trans, lower=False,
+    check_finite=False)``.
+
+    This reproduces scipy's dispatch verbatim, including the trick of solving
+    the transposed system when ``a1`` is not Fortran-ordered (``dtrtrs``
+    expects column-major input), so the exact same LAPACK call is made.
+
+    The shape validation is load-bearing, not defensive: when there are fewer
+    residuals than parameters the R factor is rectangular and scipy raises
+    ``ValueError``, which ``_newton_direction`` catches to fall back on the
+    steepest-descent direction.
+    """
+    if a1.shape[0] != a1.shape[1]:
+        raise ValueError("expected square matrix")
+
+    if a1.shape[0] != b1.shape[0]:
+        raise ValueError(
+            f"shapes of a {a1.shape} and b {b1.shape} are incompatible"
+        )
+
+    if a1.flags.f_contiguous:
+        x, info = _dtrtrs(
+            a1,
+            b1,
+            overwrite_b=overwrite_b,
+            lower=False,
+            trans=trans,
+            unitdiag=False,
+        )
+
+    else:
+        x, info = _dtrtrs(
+            a1.T,
+            b1,
+            overwrite_b=overwrite_b,
+            lower=True,
+            trans=not trans,
+            unitdiag=False,
+        )
+
+    if info == 0:
+        return x
+
+    if info > 0:
+        raise LinAlgError(
+            f"singular matrix: resolution failed at diagonal {info - 1}"
+        )
+
+    raise ValueError(
+        f"illegal value in {-info}-th argument of internal trtrs"
+    )
[…]
