### 2026-08-10 2ecb9a8
add optimized periodicitywang c code
