+++ pyhctsa/toolboxes/c22/PD_PeriodicityWang.c
@@ -15,18 +15,46 @@
 #include "splinefit.h"
 #include "stats.h"
 
-// Structure to hold multiple threshold results
-typedef struct {
-    int th1;  // threshold 0
-    int th2;  // threshold 0.01
-    int th3;  // threshold 0.1
-    int th4;  // threshold 0.2
-    int th5;  // threshold 1/sqrt(N)
-    int th6;  // threshold 5/sqrt(N)
-    int th7;  // threshold 10/sqrt(N)
-} PD_PeriodicityWang_Results;
+#ifndef ACF_B
+#define ACF_B 4
+#endif
 
-// Your existing C function (keep exactly as is)
+static void PD_acf_block(const double * restrict y, const int n, const int tLo,
+                         const int tHi, double * restrict acf)
+{
+    int t = tLo;
+
+    for (; t + (ACF_B-1) <= tHi; t += ACF_B) {
+        const int tmax = t + (ACF_B-1);
+
+        int iEnd = n - tmax;
+        if (iEnd < 0) iEnd = 0;
+
+        double a[ACF_B];
+        for (int k = 0; k < ACF_B; k++) a[k] = 0.0;
+
+        for (int i = 0; i < iEnd; i++) {
+            const double v = y[i];
+            const double * restrict w = y + i + t;
+            for (int k = 0; k < ACF_B; k++) a[k] += v * w[k];
+        }
+
+        for (int k = 0; k < ACF_B; k++) {
+            const int tau = t + k;
+            const int m   = n - tau;
+            for (int i = iEnd; i < m; i++) a[k] += y[i] * y[i+tau];
+            acf[tau-1] = a[k] / m;
+        }
+    }
+
+    for (; t <= tHi; t++) {
+        const int m = n - t;
+        double acc = 0.0;
+        for (int i = 0; i < m; i++) acc += y[i] * y[i+t];
+        acf[t-1] = acc / m;
+    }
+}
+
 PD_PeriodicityWang_Results PD_PeriodicityWang(const double * y, const int size){
     
     // Initialize results structure with default value 1 (matching MATLAB)
@@ -53,87 +81,87 @@ PD_PeriodicityWang_Results PD_PeriodicityWang(const double * y, const int size){
     double * ySpline = malloc(size * sizeof(double));
     // fit a spline with 3 nodes to the data
     splinefit(y, size, ySpline);
-    
+
     // subtract spline from data to remove trend
     double * ySub = malloc(size * sizeof(double));
     for(int i = 0; i < size; i++){
         ySub[i] = y[i] - ySpline[i];
     }
-    
-    // compute autocorrelations up to 1/3 of the length of the time series
-    int acmax = (int)ceil((double)size/3);
+    free(ySpline);
+
+    const int acmax = (int)ceil((double)size/3);
     double * acf = malloc(acmax*sizeof(double));
-    for(int tau = 1; tau <= acmax; tau++){
-        // correlation/ covariance the same, don't care for scaling (cov would be more efficient)
-        acf[tau-1] = autocov_lag(ySub, size, tau);
-    }
-    
-    // find troughs and peaks
     int * troughs = malloc(acmax * sizeof(int));
-    int * peaks = malloc(acmax * sizeof(int));
-    int nTroughs = 0;
-    int nPeaks = 0;
-    
-    double slopeIn = 0;
-    double slopeOut = 0;
-    for(int i = 1; i < acmax-1; i ++){
-        slopeIn = acf[i] - acf[i-1];
-        slopeOut = acf[i+1] - acf[i];
-        if(slopeIn < 0 && slopeOut > 0) {
-            troughs[nTroughs] = i + 1; // +1 to match MATLAB 1-based indexing
-            nTroughs += 1;
[…]
