+++ pyhctsa/toolboxes/c22/PD_PeriodicityWang.h
@@ -9,8 +9,19 @@
 #ifndef PD_PeriodicityWang_h
 #define PD_PeriodicityWang_h
 
-#include <stdio.h>
+typedef struct {
+    int th1;  // threshold 0
+    int th2;  // threshold 0.01
+    int th3;  // threshold 0.1
+    int th4;  // threshold 0.2
+    int th5;  // threshold 1/sqrt(N)
+    int th6;  // threshold 5/sqrt(N)
+    int th7;  // threshold 10/sqrt(N)
+} PD_PeriodicityWang_Results;
 
-extern int PD_PeriodicityWang_th0_01(const double * y, const int size);
+PD_PeriodicityWang_Results PD_PeriodicityWang(
+    const double *y,
+    int size
+);
 
 #endif /* PD_PeriodicityWang_h */
