### 2026-08-10 7546d3c
add periodogram to spectral summaries
