### 2026-08-10 f1eea4a
update polysat feature, compute using opt least sq
