+++ pyhctsa/operations/spectral.py
@@ -1,8 +1,9 @@
 import numpy as np
 from numpy.typing import ArrayLike
 import scipy.fft
-from scipy.optimize import curve_fit
 import statsmodels.api as sm
+
+from ..toolboxes.matlab.matlab_fit import lsqcurvefit_trr, goodness_of_fit
 
 from ..operations.correlation import autocorr, first_crossing
 from ..operations.distribution import moments
@@ -223,15 +224,13 @@ def spectral_summaries(y: ArrayLike, psd_meth: str = 'fft', window_type: str = '
     out['fpoly2_r2'] = 1 - (sum_sq_err / (np.sum((cs_s - np.mean(cs_s))**2)))
 
     # Fit polysat a*x^2/(b+x^2) (has zero derivative at zero, though)
-    polysat = lambda x, a, b : (a*(x**2))/(b + x**2)
-    popt, _ = curve_fit(polysat, w, cs_s, p0=[cs_s[-1], 100])
-    a, b = popt
+    polysat = lambda p, x: (p[0] * (x**2)) / (p[1] + x**2)
+    a, b = lsqcurvefit_trr(polysat, [cs_s[-1], 100], w, cs_s)
     out['fpolysat_a'] = a
     out['fpolysat_b'] = b
-    residuals = polysat(w, a, b) - cs_s
-    sum_sq_err = np.sum(residuals**2)
-    out['fpolysat_r2'] = 1 - (sum_sq_err/(np.sum((cs_s - np.mean(cs_s))**2)))
-    out['fpolysat_rmse'] = np.sqrt(np.mean(residuals**2))
+    gof = goodness_of_fit(cs_s, polysat([a, b], w), 2)
+    out['fpolysat_r2'] = gof['rsquare']
+    out['fpolysat_rmse'] = gof['rmse']
 
     # Shannon spectral entropy
     h_shann = -s * np.log(s)
