### 2026-08-10 63a1582
add missing features in spectral summaries
