+++ pyhctsa/toolboxes/matlab/matlab_fit.py
@@ -1137,3 +1137,164 @@ def goodness_of_fit(y, y_fit, num_coeffs):
         "adjrsquare": 1 - (1 - rsquare) * (n - 1) / dfe,
         "rmse": np.sqrt(sse / dfe),
     }
+
+# ------------------------------------------------------------------------------
+# Robust linear fitting (MATLAB's robustfit)
+# ------------------------------------------------------------------------------
+
+
+def _bisquare(u):
+    """Tukey's bisquare weight function, as used by MATLAB's ``robustfit``."""
+    u = np.asarray(u, dtype=float)
+    return (np.abs(u) < 1) * (1 - u**2) ** 2
+
+
+def _madsigma(r, p):
+    """MAD-based scale estimate (``madsigma`` in MATLAB's ``statrobustfit``)."""
+    rs = np.sort(np.abs(r))
+    return np.median(rs[max(1, p) - 1:]) / 0.6745
+
+
+def _robustsigma(r, p, s, tune, h):
+    """
+    Robust residual scale estimate of DuMouchel & O'Brien (1989), following
+    MATLAB's ``statrobustsigma``.
+    """
+    st = s * tune
+    n = len(r)
+    u = r / st
+
+    phi = u * _bisquare(u)
+    delta = 0.0001
+    u1 = u - delta
+    phi0 = u1 * _bisquare(u1)
+    u1 = u + delta
+    phi1 = u1 * _bisquare(u1)
+
+    dphi = (phi1 - phi0) / (2 * delta)
+    m1 = np.mean(dphi)
+    m2 = np.sum((1 - h) * phi**2) / (n - p)
+    K = 1 + (p / n) * (1 - m1) / m1
+
+    return K * np.sqrt(m2) * st / m1
+
+
+def robustfit(x, y, tune=4.685):
+    """
+    Iteratively reweighted least-squares linear fit, a port of MATLAB's
+    ``robustfit(x, y)`` with the default bisquare weight function and a
+    constant term.
+
+    Parameters
+    ----------
+    x : array-like
+        Predictor (a single column; the constant term is added internally).
+
+    y : array-like
+        Response.
+
+    tune : float, optional
+        Tuning constant of the bisquare weight function (MATLAB default
+        4.685).
+
+    Returns
+    -------
+    b : ndarray
+        Coefficient estimates, ``[intercept, gradient]``.
+
+    stats : dict
+        Keys are ``ols_s`` (sigma estimate from ordinary least squares),
+        ``robust_s`` (robust sigma estimate), ``mad_s`` (MAD-based sigma
+        estimate), ``s`` (the larger of ``robust_s`` and a weighted average
+        of ``ols_s`` and ``robust_s``) and ``se`` (standard errors of the
+        coefficient estimates).
+    """
+    x = np.asarray(x, dtype=float).ravel()
+    y = np.asarray(y, dtype=float).ravel()
+
+    n = len(y)
+    X = np.column_stack((np.ones(n), x))
+    p = X.shape[1]
+    if n <= p:
+        raise ValueError("Not enough data points for a robust fit.")
+
+    from scipy.linalg import qr as _qr
+
+    Q, R, perm = _qr(X, mode="economic", pivoting=True)
+
+    eps = np.finfo(float).eps
+    tol = abs(R[0, 0]) * max(n, p) * eps
+    xrank = int(np.sum(np.abs(np.diag(R)) > tol))
+
+    R1 = R[:xrank, :xrank]
+    b = np.zeros(p)
+    if xrank == p:
[…]
