### 2026-08-15 4221d8e
add poincare_section
