+++ pyhctsa/toolboxes/Tisean_3_0_1/TS_poincare.c
@@ -1,0 +1,280 @@
+/*
+ *   Python C-extension wrapper around TISEAN's `poincare` program
+ *   (TISEAN 3.0.1, source_c/poincare.c, author Rainer Hegger).
+ *
+ *   The original is a command-line program: it parses argv into file-scope
+ *   globals, reads the series from a text file, and prints one line per cut to
+ *   stdout or to a `.poin` file (calling exit() when the crossing level lies
+ *   outside the data). None of that survives repeated in-process calls, so the
+ *   algorithm has been made re-entrant here: every global lives in a
+ *   `poin_ctx`, the series arrives as a numpy array, and the cuts are handed
+ *   back as a single (n_cuts, dim) array -- the rows the program would have
+ *   written, at full double precision rather than through `%e`.
+ *
+ *   The numerics are a line-by-line transcription of poincare.c, including the
+ *   `lasttime > 0.0` gate that drops the first crossing and the naive summation
+ *   in routines/variance.c, which fixes the default crossing level (and hence,
+ *   at the last ulp, which samples count as crossings).
+ *
+ *   TISEAN is Copyright (c) 1998-2007 Rainer Hegger, Holger Kantz,
+ *   Thomas Schreiber, and is distributed under the GNU General Public License
+ *   version 2 or later.
+ */
+
+#include <Python.h>
+#include <numpy/arrayobject.h>
+
+#include <math.h>
+#include <stdint.h>
+#include <stdlib.h>
+#include <string.h>
+
+typedef struct {
+    const double *series;
+    int64_t length;
+
+    int dim;                   /* embedding dimension (poincare -m)          */
+    int delay;                 /* embedding delay (poincare -d)              */
+    int comp;                  /* component to cut (poincare -q)             */
+    int dir;                   /* 0: crossing from below, 1: from above      */
+    double where;              /* crossing level (poincare -a)               */
+
+    double *rows;              /* nrows * dim, row-major                     */
+    int64_t nrows, cap;
+} poin_ctx;
+
+static void ctx_free(poin_ctx *c)
+{
+    free(c->rows);
+    memset(c, 0, sizeof(*c));
+}
+
+/* Hand out the next row of the output table, growing it as needed. */
+static double *next_row(poin_ctx *c)
+{
+    if (c->nrows == c->cap) {
+        int64_t cap = (c->cap == 0) ? 256 : c->cap * 2;
+        double *rows = (double *)realloc(c->rows,
+                                         sizeof(double) * (size_t)(cap * c->dim));
+        if (rows == NULL)
+            return NULL;
+        c->rows = rows;
+        c->cap = cap;
+    }
+    return c->rows + c->nrows++ * c->dim;
+}
+
+/* ---------------------------------------------------------------------------
+ * poincare(): the body of poincare.c's poincare(), writing into `rows` instead
+ * of to a file. Returns 0 on success, -1 on allocation failure.
+ * ------------------------------------------------------------------------ */
+static int poincare_run(poin_ctx *c)
+{
+    const double *series = c->series;
+    const double where = c->where;
+    const int dim = c->dim, delay = c->delay, comp = c->comp;
+    int64_t i, jd, start, stop;
+    long j;
+    int k, cut;
+    double delta, xcut, *row;
+    double time = 0.0, lasttime = 0.0;
+
+    start = (int64_t)(comp - 1) * delay;
[…]
