### 2026-08-15 b580d0f
add poincare_section
