### 2026-08-15 014784a
Update numpy dependency version in pyproject.toml
