+++ pyproject.toml
@@ -18,7 +18,7 @@ maintainers = [
 ]
 
 dependencies = [
-  "numpy",
+  "numpy>=2",
   "scipy",
   "pyyaml",
   "statsmodels",
