### 2026-08-15 203ce07
Update numpy requirement to version 2
