+++ pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["setuptools>=68", "wheel", "numpy>=1.24"]
+requires = ["setuptools>=68", "numpy>=2"]
 build-backend = "setuptools.build_meta"
 
 [project]
