### 2026-08-15 0b85e35
fix module configs
