### 2026-08-15 152f828
update stationarity
