+++ pyhctsa/operations/pre_process.py
@@ -82,7 +82,6 @@ def preproc_compare(y: ArrayLike, detrend_meth: str = 'medianf') -> dict:
     """
     y = np.asarray(y)
     N = len(y)
-    r = np.arange(N)
 
     # ------------------------------------------------------------------
     # Apply preprocessing: y (raw) -> y_d (detrended/transformed)
