+++ pyhctsa/operations/scaling.py
@@ -1,5 +1,3 @@
-from typing import Union
-
 from numpy.typing import ArrayLike
 import numpy as np
 from scipy.interpolate import interp1d
