### 2026-08-15 777402b
drop explicit module-level configs
