+++ pyhctsa/operations/hypothesis_tests.py
@@ -1,13 +1,23 @@
+import logging
 from typing import Union
 
 import numpy as np
 from numpy.typing import ArrayLike
 from arch.unitroot import VarianceRatio
-from scipy.stats import jarque_bera, norm, wilcoxon
+from scipy.interpolate import PchipInterpolator
+from scipy.optimize import brentq, minimize
+from scipy.stats import beta as beta_dist
+from scipy.stats import chi2, expon
+from scipy.stats import gamma as gamma_dist
+from scipy.stats import gumbel_l, jarque_bera, kstwo, lognorm, norm, rayleigh, uniform, weibull_min, wilcoxon
 from statsmodels.sandbox.stats.runs import runstest_1samp
 from statsmodels.stats.descriptivestats import sign_test
 
 from ..utils import ljung_box_pvalue
+from ..toolboxes.distribution_fits._lilliefors_tables import LILLIE_ALPHAS, LILLIE_TABLES
+from ..toolboxes.distribution_fits.distfits import betafit, evfit
+
+logger = logging.getLogger('pyhctsa')
 
 def variance_ratio_test(y: ArrayLike, periods: Union[int, list[int], float] = 2,
                         iids: Union[int, list[int]] = 0) -> dict:
@@ -128,3 +138,247 @@ def hypothesis_test(x: ArrayLike, the_test: str = 'signtest') -> float:
     else:
         raise ValueError(f"Unknown test: {the_test}.")
     return p
+
+def distribution_test(x: ArrayLike, the_test: str = 'chi2gof', the_distn: str = 'norm',
+                      num_bins: int = 10) -> float:
+    """
+    Hypothesis test for distributional fits to a data vector.
+
+    Fits a distribution to the data and then performs an appropriate hypothesis
+    test to quantify the difference between the two distributions.
+
+    Parameters
+    ----------
+    x : array-like
+        The input data vector.
+    the_test : str, optional
+        The hypothesis test to perform:
+
+        - 'chi2gof': chi^2 goodness of fit test
+        - 'ks': Kolmogorov-Smirnov test
+        - 'lillie': Lilliefors test (only defined for 'norm', 'ev', and 'exp')
+
+        Default is ``'chi2gof'``.
+    the_distn : str, optional
+        The distribution to fit:
+
+        - 'norm' (Normal)
+        - 'ev' (Extreme value)
+        - 'uni' (Uniform)
+        - 'beta' (Beta)
+        - 'rayleigh' (Rayleigh)
+        - 'exp' (Exponential)
+        - 'gamma' (Gamma)
+        - 'logn' (Log-normal)
+        - 'wbl' (Weibull)
+
+        Default is ``'norm'``.
+    num_bins : int, optional
+        The number of bins to use for the chi^2 goodness of fit test.
+        Default is 10.
+
+    Returns
+    -------
+    float
+        P-value from the hypothesis test. NaN when the fit is not valid for
+        the data (e.g., a positive-only distribution fitted to data with
+        negative values).
+    """
+    x = np.asarray(x, dtype=float)
+    num_bins = int(num_bins)
+
+    if the_distn == 'beta':
+        # clumsily scale to the range (0,1), as in MATLAB
+        sd = np.std(x, ddof=1)
+        x = (x - np.min(x) + 0.01 * sd) / (np.max(x) - np.min(x) + 0.01 * sd)
+    elif the_distn in ('rayleigh', 'exp', 'gamma'):
+        if np.any(x < 0):
[…]
