+++ pyhctsa/toolboxes/distribution_fits/_lilliefors_tables.py
@@ -1,0 +1,90 @@
+"""Tabulated critical values for the Lilliefors goodness-of-fit test.
+
+These reproduce the tables used by MATLAB's ``lillietest``: Monte-Carlo
+quantiles for sample sizes 4..20 (rows) at each significance level in
+``LILLIE_ALPHAS`` (columns), plus coefficient triples ``(c1, c2, c3)`` for the
+asymptotic approximation ``CV = c1/sqrt(n) - c2/n - c3/n**1.5`` used for n > 20.
+
+Kept in a separate module so the bulk of static data does not bloat
+``hypothesis_tests``. Consumed by ``hypothesis_tests._lilliefors_pvalue``.
+"""
+
+import numpy as np
+
+LILLIE_ALPHAS = np.array(
+    [0.001, 0.0015, 0.002, 0.005, 0.01, 0.015, 0.02, 0.05, 0.10, 0.15, 0.20, 0.50])
+
+LILLIE_TABLES = {
+    'norm': {
+        'quantiles': np.array([
+            [0.4328, 0.4308, 0.4292, 0.4217, 0.4131, 0.4065, 0.4008, 0.3754, 0.3453, 0.3213, 0.3028, 0.2581],
+            [0.4386, 0.4335, 0.4291, 0.4128, 0.3966, 0.3851, 0.3756, 0.3431, 0.3189, 0.3026, 0.2893, 0.2332],
+            [0.4225, 0.4152, 0.4091, 0.3878, 0.3703, 0.3599, 0.3521, 0.3236, 0.2973, 0.2810, 0.2688, 0.2187],
+            [0.4006, 0.3931, 0.3873, 0.3677, 0.3506, 0.3397, 0.3317, 0.3041, 0.2802, 0.2643, 0.2523, 0.2062],
+            [0.3828, 0.3749, 0.3692, 0.3494, 0.3326, 0.3223, 0.3146, 0.2880, 0.2651, 0.2502, 0.2387, 0.1947],
+            [0.3656, 0.3582, 0.3524, 0.3332, 0.3171, 0.3071, 0.2997, 0.2740, 0.2520, 0.2379, 0.2271, 0.1851],
+            [0.3509, 0.3428, 0.3377, 0.3185, 0.3034, 0.2938, 0.2866, 0.2620, 0.2410, 0.2274, 0.2171, 0.1767],
+            [0.3374, 0.3305, 0.3252, 0.3067, 0.2915, 0.2823, 0.2754, 0.2515, 0.2312, 0.2181, 0.2082, 0.1695],
+            [0.3253, 0.3184, 0.3131, 0.2955, 0.2808, 0.2714, 0.2647, 0.2418, 0.2224, 0.2098, 0.2002, 0.1631],
+            [0.3147, 0.3074, 0.3021, 0.2849, 0.2706, 0.2618, 0.2553, 0.2333, 0.2145, 0.2025, 0.1932, 0.1574],
+            [0.3033, 0.2970, 0.2923, 0.2756, 0.2619, 0.2536, 0.2473, 0.2257, 0.2075, 0.1958, 0.1868, 0.1522],
+            [0.2960, 0.2893, 0.2844, 0.2673, 0.2539, 0.2457, 0.2398, 0.2189, 0.2012, 0.1898, 0.1811, 0.1475],
+            [0.2884, 0.2821, 0.2772, 0.2604, 0.2472, 0.2390, 0.2331, 0.2126, 0.1953, 0.1843, 0.1759, 0.1433],
[…]
