### 2026-08-15 7f06e31
add missing features in pre_process
