+++ docs/source/api.rst
@@ -205,6 +205,7 @@ Nonlinearity
    pyhctsa.operations.nonlinearity.local_density
    pyhctsa.operations.nonlinearity.tisean_d2
    pyhctsa.operations.nonlinearity.poincare_section
+   pyhctsa.operations.nonlinearity.delay_time
 
 .. _physicsmeths:
 
