+++ pyhctsa/toolboxes/Tisean_3_0_1/tisean.py
@@ -2,16 +2,20 @@
 Python front-end to the TISEAN 3.0.1 routines used by pyhctsa.
 
 hctsa drives TISEAN through the shell: it writes the time series to a temporary
-file, runs ``d2``, and then pipes the resulting ``.c2`` file through ``c2g`` and
-``c2t``.  Here the same three steps happen in-process:
+file, runs ``d2`` or ``c1``, and then pipes the resulting ``.c2``/``.c1`` file
+through ``c2g``, ``c2t`` or ``c2d``.  Here the same steps happen in-process:
 
 * :func:`d2` calls the C kernel in ``TS_d2.c`` (a re-entrant transcription of
   TISEAN's ``d2.c``) and assembles the ``.c2``/``.d2``/``.h2`` tables from the
   raw pair counts, using the expressions ``d2.c`` prints.
-* :func:`c2g` and :func:`c2t` are ports of ``source_f/c2g.f`` and
-  ``source_f/c2t.f``.  Those are Fortran, so they are reimplemented rather than
-  wrapped -- both are short, and this keeps the package free of a Fortran
-  toolchain.
+* :func:`c1` calls the C kernel in ``TS_c1.c``, a re-entrant transcription of
+  ``source_f/c1.f`` and its estimator ``source_f/d1.f``.  Those are Fortran, and
+  single-precision Fortran at that, so the kernel is written in ``float``; see
+  the header of ``TS_c1.c`` for what else that transcription has to pin down
+  (in particular the random draw of the reference points).
+* :func:`c2g`, :func:`c2t` and :func:`c2d` are ports of ``source_f/c2g.f``,
+  ``source_f/c2t.f`` and ``source_f/c2d.f``.  Those are Fortran too, but short
+  enough to reimplement, which keeps the package free of a Fortran toolchain.
 
 TISEAN is Copyright (c) 1998-2007 Rainer Hegger, Holger Kantz, Thomas
 Schreiber, and is distributed under the GNU General Public License v2 or later.
@@ -25,10 +29,11 @@ from typing import List, Optional
 import numpy as np
 from numpy.typing import ArrayLike
 
+from . import c1 as _c1_c
 from . import d2 as _d2_c
 from . import poincare as _poincare_c
 
-__all__ = ["d2", "c2g", "c2t", "poincare"]
+__all__ = ["c1", "c2d", "d2", "c2g", "c2t", "poincare"]
 
 
 # 15-point Gauss-Kronrod rule, as tabulated in SLATEC's dqk15.f (which is what
@@ -160,6 +165,142 @@ def d2(
             h2_blocks.append(np.array(rows, dtype=float).reshape(-1, 2))
 
     return {"c2": c2_blocks, "d2": d2_blocks, "h2": h2_blocks}
+
+
+def c1(
+    y: ArrayLike,
+    delay: int = 1,
+    mmin: int = 2,
+    mmax: int = 10,
+    tsep: int = 1,
+    nref: int = 100,
+    resolution: float = 2.0,
+    kmax: int = 100,
+    seed: int = 1,
+    write_precision: Optional[int] = 7,
+) -> List[np.ndarray]:
+    """
+    Fixed-mass curves for the information dimension (TISEAN's ``c1``).
+
+    For a geometrically spaced sequence of neighbour masses, and for each
+    embedding dimension, this reports the mean radius that encloses that mass
+    around a set of randomly chosen reference points. The local slope of the
+    resulting curve -- see :func:`c2d` -- estimates the information dimension
+    D1.
+
+    Parameters
+    ----------
+    y : array-like
+        Input time series.
+    delay : int, optional
+        Time delay of the embedding (``c1 -d``). Default is 1.
+    mmin, mmax : int, optional
[…]
