+++ pyhctsa/operations/nonlinearity.py
@@ -836,6 +836,156 @@ def tisean_d2(y: ArrayLike, tau: Union[int, str] = 1, maxm: int = 10,
 
     return out
 
+def _findscalingr_c1(x: np.ndarray) -> np.ndarray:
+    # As _findscalingr_ind, but for the fixed-mass curves: the candidate ranges
+    # start in the first quarter rather than the first half, the reward for a
+    # longer range is larger, and the spread is measured on the raw values
+    # rather than on a rescaled copy. Returns
+    # [start, end, goodness, dimension, dimension std].
+    v = np.asarray(x, dtype=float).ravel()
+    l = v.size
+    gamma = 0.005  # regularization parameter, chosen empirically
+    stptr = np.arange(1, int(np.floor(l / 4)))         # must be in the first quarter
+    endptr = np.arange(int(np.ceil(l / 4)) + 1, l + 1)  # and in the last three
+    if stptr.size == 0 or endptr.size == 0:
+        raise ValueError('Error finding scaling range')
+
+    mybad = np.empty((stptr.size, endptr.size))
+    for i, s in enumerate(stptr):
+        for j, e in enumerate(endptr):
+            mybad[i, j] = np.std(v[s - 1:e], ddof=1)
+    mybad -= gamma * (endptr[np.newaxis, :] - stptr[:, np.newaxis] + 1)
+
+    a, b, best = _argmin_first_colmajor(mybad)
+    if a is None:
+        raise ValueError('Error finding scaling range')
+    ri1, ri2 = int(stptr[a]), int(endptr[b])
+    sub = v[ri1 - 1:ri2]
+    return np.array([ri1, ri2, best, np.mean(sub), np.std(sub, ddof=1)])
+
+
+def tisean_c1(y: ArrayLike, tau: Union[int, str] = 1, mmm: tuple = (2, 10),
+              tsep: Union[int, float] = 0.02,
+              nref: Union[int, float] = 0.5) -> Union[dict, float]:
+    """
+    Information dimension from the TISEAN package's ``c1`` routine.
+
+    Runs the fixed-mass computation of the information dimension [1]_: for a
+    geometrically spaced sequence of neighbour masses, ``c1`` reports the mean
+    radius that encloses that mass, and ``c2d`` turns each curve into local
+    slopes. Every embedding dimension then gets its own scaling range, found by
+    a penalized search that trades the flatness of the slopes against the span
+    of scales they hold over, and the returned statistics summarise the
+    dimension estimates and the scaling ranges across dimensions.
+
+    References
+    ----------
+    .. [1] R. Hegger, H. Kantz and T. Schreiber, "Practical implementation of
+        nonlinear time series methods: The TISEAN package", Chaos 9(2) 413 (1999)
+
+    Parameters
+    ----------
+    y : array-like
+        Input time series.
+    tau : int or str, optional
+        The time-delay. Can be an integer, or ``'ac'`` for the first
+        zero-crossing of the autocorrelation function, or ``'mi'`` for the first
+        minimum of the automutual information. Default is 1.
+    mmm : tuple of int, optional
+        Minimum and maximum embedding dimension. Default is ``(2, 10)``.
+    tsep : int or float, optional
+        Time separation between a reference point and the neighbours counted
+        around it. A value in ``(0, 1)`` is taken as a proportion of the
+        time-series length. Default is 0.02, i.e. 2% of the data length.
+    nref : int or float, optional
+        Number of reference points. A value in ``(0, 1]`` is taken as a
[…]
