+++ pyhctsa/configurations/hctsa.yaml
@@ -1733,7 +1733,15 @@ nonlinearity:
       - {tau: 'ac', maxm: 10, theiler_win: 0.01, zscore: True}
     legacy_name: NL_TISEAN_d2
     ordered_args: ['tau', 'maxm', 'theiler_win']
-    
+
+  tisean_c1:
+    base_name: tisean_c1
+    dependencies:
+    configs:
+      - {tau: 1, mmm: [[2, 6]], tsep: 25, nref: 0.1, zscore: True}
+    legacy_name: NL_TISEAN_c1
+    ordered_args: ['tau', 'mmm', 'tsep', 'nref']
+
   poincare_section:
     base_name: poincare_section
     dependencies:

[…]
