### 2026-08-19 1c174b0
add wentropy
