+++ pyhctsa/toolboxes/matlab/_pptest_tables.py
@@ -1,0 +1,128 @@
+"""Tabulated critical values for the Phillips-Perron unit root test.
+
+Simulated quantiles of the test statistic for each combination of
+regression model (``'ar'``, ``'ard'``, ``'ts'``) and test statistic
+(``'t1'``, ``'t2'``). Rows are effective sample sizes ``PP_SAMP_SIZES``,
+columns are cumulative probabilities ``PP_SIG_LEVELS``.
+
+These values derive from the Dickey-Fuller simulations described in
+Hamilton (1994), Table B.6.
+"""
+
+import numpy as np
+
+# Effective sample sizes indexing the table rows. The final entry is a stand-in
+# for "asymptotic"; when T exceeds 10000 the top row is stretched rather than extrapolated.
+PP_SAMP_SIZES = np.array(
+    [10, 15, 20, 25, 30, 40, 50, 75, 100, 150, 200, 300, 500, 1000, 10000], dtype=float)
+
+# Cumulative probabilities indexing the table columns:
+# 0.001, 0.005:0.005:0.10, 0.125:0.025:0.20, 0.80:0.025:0.875, 0.90:0.005:0.995, 0.999
+PP_SIG_LEVELS = np.array(
+    [0.001, *np.arange(0.005, 0.1001, 0.005), *np.arange(0.125, 0.2001, 0.025),
+     *np.arange(0.80, 0.8751, 0.025), *np.arange(0.90, 0.9951, 0.005), 0.999])
+
+PP_CV_TABLES = {
+    ('ar', 't1'): np.array([
+        [-3.9976, -3.1423, -2.7891, -2.5834, -2.4351, -2.3201, -2.226, -2.146, -2.0766, -2.0148, -1.9594, -1.9095, -1.8629, -1.8201, -1.7799, -1.7427, -1.7075, -1.6744, -1.6427, -1.6126, -1.5839, -1.4569, -1.3501, -1.2567, -1.173, 0.4848, 0.5897, 0.7056, 0.8373, 0.9913, 1.0258, 1.0621, 1.0996, 1.1392, 1.181, 1.2254, 1.2722, 1.323, 1.3775, 1.4361, 1.5004, 1.572, 1.652, 1.7437, 1.8514, 1.9818, 2.1485, 2.3836, 2.784, 3.7494],  # T = 10
+        [-3.7444, -3.0262, -2.7138, -2.5289, -2.3954, -2.2909, -2.2044, -2.1305, -2.0659, -2.0083, -1.956, -1.9084, -1.8648, -1.8239, -1.7858, -1.7501, -1.7161, -1.6841, -1.6537, -1.6248, -1.597, -1.4738, -1.369, -1.2773, -1.195, 0.4548, 0.5589, 0.6733, 0.8024, 0.9521, 0.9856, 1.0203, 1.0568, 1.0946, 1.1352, 1.1775, 1.2225, 1.2703, 1.3218, 1.3771, 1.4378, 1.5044, 1.579, 1.6643, 1.7627, 1.8807, 2.031, 2.2387, 2.5818, 3.3667],  # T = 15
[…]
