### 2026-08-19 53fb368
bug fix for moments - make scale invariant as per HTCSA
