### 2026-08-19 44e2b23
update hctsa.yaml
