### 2026-08-19 c0a3e95
bug fix surprise memory_data[inmem2+2] — inmem2 indexes into inmem1, not memory_data
