### 2026-08-19 3bce534
bug fix moment_corr - density = ptp(M1)*ptp(M2)/N is area-per-point, not a density
