### 2026-08-19 7a1d84e
bug fix viz graph
