+++ pyhctsa/operations/graph.py
@@ -152,7 +152,7 @@ def visibility_graph(y: ArrayLike, meth: str = 'horiz', max_l: int = 5000) -> di
     out = {}
     # Degree distribution: basic statistics
     out['mode'] = scipy.stats.mode(k).mode
-    out['propmode'] = np.sum(k == out['mode'])/np.sum(k)
+    out['propmode'] = np.sum(k == out['mode'])/len(k)
     out['meank'] = meank # mean number of links per node
     out['mediank'] = mediank
     out['stdk'] = stdk
