+++ pyhctsa/operations/spectral.py
@@ -50,20 +50,39 @@ def spectral_summaries(y: ArrayLike, psd_meth: str = 'fft', window_type: str = '
     y = np.asarray(y)
     ny = len(y)
 
-    window = None
+    # Welch's method needs a window *shorter* than the full series so that the
+    # signal is actually split into multiple overlapping segments to average
+    # over -- that segment-averaging is the entire point of the method. A window
+    # spanning the whole series (correct for a periodogram, a genuine
+    # single-segment whole-series estimate) collapses Welch to one segment with
+    # no averaging, making it near-indistinguishable from an unwindowed
+    # periodogram.
+    #
+    # Capping the length at a fixed 256 rather than a fraction of ny keeps the
+    # segment COUNT growing with N at FIXED frequency resolution, which is
+    # Welch's actual convergence guarantee. Tying it to ny/4 instead held the
+    # count fixed and let the resolution grow without bound, which is what made
+    # maxWidth (and other bandwidth-like fields) drift with series length.
+    # 256 matches both scipy's and MATLAB's own default segment length.
+    if psd_meth == 'welch':
+        win_length = max(min(256, int(np.floor(ny / 4 + 0.5))), 16)
+    else:
+        win_length = ny
+
     # Set window (for periodogram and welch):
     if window_type == 'none':
-        window = []
+        # MATLAB's pwelch falls back to a Hamming window when passed []
+        window = np.hamming(win_length) if psd_meth == 'welch' else []
     elif window_type == 'hamming':
-        window = np.hamming(ny)
+        window = np.hamming(win_length)
     elif window_type == 'hann':
-        window = np.hanning(ny)
+        window = np.hanning(win_length)
     elif window_type == 'bartlett':
-        window = np.bartlett(ny)
+        window = np.bartlett(win_length)
     elif window_type == 'boxcar':
-        window = scipy.signal.windows.boxcar(ny)
+        window = scipy.signal.windows.boxcar(win_length)
     elif window_type == 'rect':
-        window = np.ones(ny)
+        window = np.ones(win_length)
     else:
         raise ValueError(f"Unknown window: {window_type}")
 
@@ -80,8 +99,17 @@ def spectral_summaries(y: ArrayLike, psd_meth: str = 'fft', window_type: str = '
     elif psd_meth == 'welch':
         # welch power spectral density estimate
         fs = 1
-        n = 2 ** (int(np.ceil(np.log2(ny))))
-        f, s = scipy.signal.welch(y, window=window, noverlap=0, nfft=n, fs=fs)
+        # nfft was previously tied to ny, independently of the window length, so
+        # even with the window capped the frequency GRID kept getting finer with
+        # N while the smoothing bandwidth stayed fixed. That mismatch shows up as
+        # its own N-dependence in bin-lag statistics (ac1/ac2, which correlate
+        # adjacent *bins*, not a physical frequency lag) and inflates bin-count
+        # sums that are meant to describe the spectrum's shape, not its number of
+        # samples. MATLAB's pwelch default is max(256, 2^nextpow2(winLength)),
+        # i.e. resolution set by the window, not the series.
+        n = max(256, 2 ** int(np.ceil(np.log2(len(window)))))
+        # noverlap=None gives the standard 50% overlap
[…]
