### 2026-08-19 27f2fba
bug fixes spectral summaries
