+++ pyhctsa/operations/spectral.py
@@ -13,7 +13,7 @@ def spectral_summaries(y: ArrayLike, psd_meth: str = 'fft', window_type: str = '
     Statistics of the power spectrum of a time series.
 
     Computes a range of statistics summarizing the power spectrum of a time series.
-    The spectrum can be estimated using a periodogram, fast Fourier transform (FFT), 
+    The spectrum can be estimated using a periodogram, fast Fourier transform (FFT),
     or Welch's method.
 
     Parameters
@@ -31,7 +31,7 @@ def spectral_summaries(y: ArrayLike, psd_meth: str = 'fft', window_type: str = '
 
     window_type : {'boxcar', 'rect', 'bartlett', 'hann', 'hamming', 'none'}, optional
         The window to use for spectral estimation:
-            
+
         - 'boxcar'
         - 'rect'
         - 'bartlett'
@@ -50,26 +50,11 @@ def spectral_summaries(y: ArrayLike, psd_meth: str = 'fft', window_type: str = '
     y = np.asarray(y)
     ny = len(y)
 
-    # Welch's method needs a window *shorter* than the full series so that the
-    # signal is actually split into multiple overlapping segments to average
-    # over -- that segment-averaging is the entire point of the method. A window
-    # spanning the whole series (correct for a periodogram, a genuine
-    # single-segment whole-series estimate) collapses Welch to one segment with
-    # no averaging, making it near-indistinguishable from an unwindowed
-    # periodogram.
-    #
-    # Capping the length at a fixed 256 rather than a fraction of ny keeps the
-    # segment COUNT growing with N at FIXED frequency resolution, which is
-    # Welch's actual convergence guarantee. Tying it to ny/4 instead held the
-    # count fixed and let the resolution grow without bound, which is what made
-    # maxWidth (and other bandwidth-like fields) drift with series length.
-    # 256 matches both scipy's and MATLAB's own default segment length.
     if psd_meth == 'welch':
         win_length = max(min(256, int(np.floor(ny / 4 + 0.5))), 16)
     else:
         win_length = ny
 
-    # Set window (for periodogram and welch):
     if window_type == 'none':
         # MATLAB's pwelch falls back to a Hamming window when passed []
         window = np.hamming(win_length) if psd_meth == 'welch' else []
@@ -86,32 +71,21 @@ def spectral_summaries(y: ArrayLike, psd_meth: str = 'fft', window_type: str = '
     else:
         raise ValueError(f"Unknown window: {window_type}")
 
-    # Compute the Fourier Transform
     if psd_meth == 'fft':
-        fs = 1  # sampling freq
-        nfft = 2 ** (int(np.ceil(np.log2(ny))))  # next power of 2
-        f = (fs / 2) * np.linspace(0, 1, int(nfft / 2) + 1)  # freq
-        w = 2 * np.pi * f  # angular freq
-        s = scipy.fft.fft(y, nfft)  # do the fourier transform
-        s = 2 * np.abs(s[:int(nfft / 2) + 1]) ** 2 / ny  # single-sided power spectral density
-        s = s / (2 * np.pi)  # convert to angular freq space
+        fs = 1
+        nfft = 2 ** (int(np.ceil(np.log2(ny))))
+        f = (fs / 2) * np.linspace(0, 1, int(nfft / 2) + 1)
+        w = 2 * np.pi * f
[…]
