+++ pyhctsa/operations/stationarity.py
@@ -93,93 +93,6 @@ def local_distributions(y: ArrayLike, num_segs: int = 5, each_or_par: str = 'par
 
     return out
 
-def dyn_win(y: ArrayLike, max_num_segments: int = 10) -> dict:
-    """
-    How stationarity estimates depend on the number of time-series subsegments.
-
-    Specifically, variation in a range of local measures are implemented: mean,
-    standard deviation, skewness, kurtosis, ApEn(1,0.2), SampEn(1,0.2), AC(1),
-    AC(2), and the first zero-crossing of the autocorrelation function.
-
-    The standard deviation of local estimates of these quantities across the time
-    series are calculated as an estimate of the stationarity in this quantity as a
-    function of the number of splits, n_{seg}, of the time series.
-
-    Parameters
-    -----------
-    y: array-like
-        The time series to analyze.
-    max_num_segments: int, optional
-        The maximum number of segments to consider. Sweeps from 2 to
-        max_num_segments. Defaults to 10.
-
-    Returns
-    -------
-    out
-        The standard deviation of this set of 'stationarity' estimates across these window sizes
-    """
-    y = np.asarray(y)
-    nsegr = np.arange(2, max_num_segments + 1, 1)  # range of nseg to sweep across
-    nmov = 1  # controls window overlap
-    num_features = 11  # num of features
-    fs = np.zeros((len(nsegr), num_features))  # standard deviation of feature values over windows
-    tau_g = first_crossing(y, 'ac', 0, 'discrete')  # global tau
-
-    for i, nseg in enumerate(nsegr):
-        wlen = int(np.floor(len(y) / nseg))  # window length
-        inc = int(np.floor(wlen / nmov))  # increment to move at each step
-        # if increment is rounded to zero, prop it up
-        if inc == 0:
-            inc = 1
-
-        num_steps = int(np.floor((len(y) - wlen) / inc) + 1)
-        qs = np.zeros((num_steps, num_features))
-
-        for j in range(num_steps):
-            y_sub = y[j * inc:j * inc + wlen]
-            tau_l = first_crossing(y_sub, 'ac', 0, 'discrete')
-
-            qs[j, 0] = np.mean(y_sub)
-            qs[j, 1] = np.std(y_sub, ddof=1)
-            qs[j, 2] = skew(y_sub)
-            qs[j, 3] = kurtosis(y_sub)
-            sampen_out = sample_entropy(y_sub, 2, 0.15)
-            qs[j, 4] = sampen_out['quadSampEn1']  # SampEn_1_015
-            #qs[j, 5] = sampen_out['quadSampEn2'] # SampEn_2_015
-            # One FFT autocorrelation per window instead of four; index the lags.
-            # acf_w[t] is bit-identical to autocorr(y_sub, t, 'Fourier')[0]; the guards
-            # reproduce autocorr's out-of-range -> NaN behaviour exactly.
-            acf_w = autocorr(y_sub, [], 'Fourier')
-            Lw = len(acf_w)
-            _pick = lambda t: acf_w[t] if 0 <= t <= Lw - 1 else np.nan
-            qs[j, 6] = acf_w[1] if Lw > 1 else np.nan  # AC1
-            qs[j, 7] = acf_w[2] if Lw > 2 else np.nan  # AC2
-            # (Sometimes tau_g or taul can be longer than ySub; then these will output NaNs:)
-            qs[j, 8] = _pick(tau_g)  # AC_glob_tau
-            qs[j, 9] = _pick(tau_l)  # AC_loc_tau
-            qs[j, 10] = tau_l
-
[…]
