### 2026-08-22 1f8d386
add autocorr_shapex2 and falling sticks
