+++ pyhctsa/operations/correlation.py
@@ -760,11 +760,11 @@ def stick_angles(y: ArrayLike) -> dict:
 
     # Initialise output dictionary
     out = {}
-    out['std_p'] = np.nanstd(angles[0], ddof=1) 
+    # std_p/std_n dropped: r=0.98 with pos_neg_asymmetry's volPos/volNeg on
+    # real EEG data, which measures regime-conditional volatility more directly.
     out['mean_p'] = np.nanmean(angles[0]) 
     out['median_p'] = np.nanmedian(angles[0])
 
-    out['std_n'] = np.nanstd(angles[1], ddof=1)
     out['mean_n'] = np.nanmean(angles[1])
     out['median_n'] = np.nanmedian(angles[1])
 
@@ -869,37 +869,36 @@ def stick_angles(y: ArrayLike) -> dict:
     
     # All angles
     
+    # statav2_all_s/statav3_all_s/statav4_all_s dropped: mutually r=0.97-0.98
+    # with statav5_all_s (and each other) on real EEG data, so only one
+    # representative of the StatAv-spread-of-all-angles family is kept.
     # StatAv2
-    out['statav2_all_m'], out['statav2_all_s'] = _sub_statav(zallAngles, 2)
+    out['statav2_all_m'], _ = _sub_statav(zallAngles, 2)
     # StatAv3
-    out['statav3_all_m'], out['statav3_all_s'] = _sub_statav(zallAngles, 3)
+    out['statav3_all_m'], _ = _sub_statav(zallAngles, 3)
     # StatAv4
-    out['statav4_all_m'], out['statav4_all_s'] = _sub_statav(zallAngles, 4)
+    out['statav4_all_m'], _ = _sub_statav(zallAngles, 4)
     # StatAv5
     out['statav5_all_m'], out['statav5_all_s'] = _sub_statav(zallAngles, 5)
     
     # correlations? 
+    # Note: ac2_p/ac2_n/ac2_all dropped (r=0.96-0.99 with the corresponding
+    # ac1_*), and tau_all/ac1_all dropped (each r=0.95-0.97 with its own p/n
+    # split) -- keeping the p/n split (rather than the pooled 'all') preserves
+    # the asymmetry signal that's the actual point of this feature.
     if len(zangles[0]) > 0:
         out['tau_p'] = first_crossing(zangles[0], 'ac', 0, 'continuous')
         out['ac1_p'] = autocorr(zangles[0], 1, 'Fourier')[0]
-        out['ac2_p'] = autocorr(zangles[0], 2, 'Fourier')[0]
     else:
         out['tau_p'] = np.nan
         out['ac1_p'] = np.nan
-        out['ac2_p'] = np.nan
     
     if len(zangles[1]) > 0:
         out['tau_n'] = first_crossing(zangles[1], 'ac', 0, 'continuous')
         out['ac1_n'] = autocorr(zangles[1], 1, 'Fourier')[0]
-        out['ac2_n'] = autocorr(zangles[1], 2, 'Fourier')[0]
     else:
         out['tau_n'] = np.nan
         out['ac1_n'] = np.nan
-        out['ac2_n'] = np.nan
-    
-    out['tau_all'] = first_crossing(zallAngles, 'ac', 0, 'continuous')
-    out['ac1_all'] = autocorr(zallAngles, 1, 'Fourier')[0]
-    out['ac2_all'] = autocorr(zallAngles, 2, 'Fourier')[0]
 
     # What does the distribution look like?
     # Some quantiles and moments
@@ -950,6 +949,225 @@ def _sub_statav(x: ArrayLike, n: int) -> tuple:
         statavstd = np.std(np.std(x_buff, axis=0), ddof=1, axis=0)/np.std(x, ddof=1, axis=0)
 
     return statavmean, statavstd
+
[…]
