### 2026-08-22 4a35c6b
add oversampling
