+++ pyhctsa/operations/correlation.py
@@ -5,17 +5,36 @@ from typing import Union
 import numpy as np
 from numpy.typing import ArrayLike
 from scipy.linalg import LinAlgError
+from scipy.linalg import solve_triangular
 from scipy.optimize import curve_fit
-from scipy.stats import expon, gaussian_kde, kurtosis, skew
+from scipy.stats import chi2, expon, gaussian_kde, kurtosis, skew
 from scipy.stats import mode as smode
 from statsmodels.tsa.stattools import pacf
 
+from ..operations.hypothesis_tests import _kstest_statistic
 from ..operations.information import first_min, automutual_info
 from ..toolboxes.c22 import periodicity_wang_wrapper
 from ..toolboxes.matlab.matlab_fit import fit_exp1, goodness_of_fit
 from ..utils import (_first_index_past_threshold, bin_picker, histc,
-                     make_mat_buffer, point_of_crossing, sign_change,
-                     time_delay_embed, z_score)
+                     make_mat_buffer, matlab_quantile, point_of_crossing,
+                     sign_change, time_delay_embed, z_score)
+
+def _resolve_time_delay(y: ArrayLike, tau: Union[int, str]) -> Union[int, float]:
+    """Resolve a string time-delay spec to a lag.
+
+    ``'ac'`` uses the first zero-crossing of the autocorrelation function and
+    ``'mi'`` the first minimum of the automutual information. An integer is
+    returned unchanged. The resolved value may be NaN (time series too short);
+    callers are responsible for handling that.
+    """
+    if not isinstance(tau, str):
+        return tau
+    if tau == 'ac':
+        return first_crossing(y, 'ac', 0, 'discrete')
+    if tau == 'mi':
+        return first_min(y, 'mi')
+    raise ValueError(f'Invalid time-delay method: {tau}. Choose either mi or ac.')
+
 
 def add_noise(y: ArrayLike, tau: Union[int, str] = 1, ami_method: str = 'even',
               extra_param: Union[int, None] = None, random_seed = None,
@@ -950,6 +969,209 @@ def _sub_statav(x: ArrayLike, n: int) -> tuple:
 
     return statavmean, statavstd
 
+def joint_non_gaussianity(y: ArrayLike, tau: Union[int, str] = 'ac', m: int = 2,
+                          theiler_win: int = 1,
+                          max_n: Union[int, str] = 10000) -> dict:
+    """
+    Tests for non-Gaussianity of the joint, time-lagged embedding distribution.
+
+    Embeds the time series in m dimensions at time delay tau (e.g., the pair
+    (x_t,x_{t+tau}) for m=2, or the triple (x_t,x_{t+tau},x_{t+2tau}) for
+    m=3) and tests whether the resulting point cloud is consistent with a
+    multivariate Gaussian.
+
+    cf. existing Gaussianity tests acting on marginal distribution
+    (distribution_test, compare_ks_fit), which all
+    A linear (e.g., AR(1)) Gaussian process has a Gaussian marginal
+    *and* a Gaussian joint embedding distribution;
+    a nonlinear or non-reversible process can look
+    Gaussian marginally while its lagged joint distribution is visibly
+    non-elliptical (curved, multimodal, or heavy/light-tailed along
+    directions the marginal alone cannot see).
+    cf. also time-irreversibility metrics like trev/tc3
+    (which use a single third-moment statistic of lagged
+    pairs/triples as a nonlinearity probe) but this is perhaps more general as
+    : it tests the whole joint shape rather than one moment combination.
+
+    Two complementary statistics are computed, both based on Mardia's
+    (1970) classical multivariate normality measures, chosen because they
+    generalize to any embedding dimension m via the same formula (so m=2
[…]
