### 2026-08-22 0180ecc
add posnegasymmetry
