+++ pyhctsa/operations/correlation.py
@@ -1491,6 +1491,141 @@ def oversampling(y: ArrayLike) -> dict:
 
     return out
 
+def _pn_corr(x: np.ndarray, y: np.ndarray) -> float:
+    """The off-diagonal entry of MATLAB's ``corrcoef(x, y)``.
+
+    Mirrors the arithmetic of MATLAB's ``correl`` subfunction: build the
+    covariance matrix with the (n-1) normalisation, take the square root of
+    the diagonal first (to avoid under/overflow) and divide twice, then limit
+    the result to [-1, 1].
+
+    A constant input has no defined correlation, and is reported as NaN. The
+    regime subsets here are routinely constant on quantised series (a whole
+    regime sitting on one level), and centring such a subset leaves only
+    rounding dust, so dividing it through -- as MATLAB does -- returns a sign
+    determined entirely by summation order rather than by the data.
+    """
+    if x[0] == x.min() and x[0] == x.max():
+        return np.nan
+    if y[0] == y.min() and y[0] == y.max():
+        return np.nan
+    n = x.size
+    xc = x - np.mean(x)
+    yc = y - np.mean(y)
+    cxy = np.dot(xc, yc) / (n - 1)
+    cxx = np.dot(xc, xc) / (n - 1)
+    cyy = np.dot(yc, yc) / (n - 1)
+    r = cxy / np.sqrt(cyy) / np.sqrt(cxx)
+    if abs(r) > 1:
+        r = np.sign(r)
+    return float(r)
+
+
+def _pn_std(x: np.ndarray) -> float:
+    """MATLAB's ``std``: zero for a single observation, NaN for none."""
+    if x.size == 0:
+        return np.nan
+    if x.size == 1:
+        return 0.0
+    return float(np.std(x, ddof=1))
+
+
+def pos_neg_asymmetry(y: ArrayLike) -> dict:
+    """
+    Asymmetry of local dynamics between positive and negative regimes.
+
+    Splits the time series by the sign of each value (assumes y is z-scored, so
+    the split is around the mean) and asks whether the one-step-ahead dynamics
+    differ between the two regimes: is the series more volatile, or more
+    persistent (higher one-step autocorrelation), when the current value is
+    above vs. below its mean? This targets a form of distribution-dynamics
+    interaction not captured by ``stick_angles``, which instead compares the
+    distribution of local slopes *within* each same-sign subsequence.
+
+    Parameters
+    ----------
+    y : array-like
+        The input time series (assumed z-scored: the regime split threshold
+        is 0).
+
+    Returns
+    -------
+    dict
+        The conditional volatility and one-step autocorrelation of the
+        positive/negative regimes, and normalized contrasts between them (the
+        volatility contrast is a leverage-effect-style statistic; the
+        autocorrelation contrast is a threshold-AR(1)-style statistic). Also
+        isolates the two zero-crossing transition types (positive-to-negative,
+        negative-to-positive) -- posMask/negMask are conditioned on the current
+        value only, so they mix crossing and non-crossing steps together; the
+        crossing-specific fields ask instead whether the jump *at* a regime
+        switch is itself asymmetric (e.g. sharper downward crossings than
+        upward ones).
+    """
+    y = np.asarray(y, dtype=float).ravel()
+
[…]
