### 2026-08-22 4b34060
add time_rev_kld
