+++ pyhctsa/operations/correlation.py
@@ -7,6 +7,7 @@ from numpy.typing import ArrayLike
 from scipy.linalg import LinAlgError
 from scipy.linalg import solve_triangular
 from scipy.optimize import curve_fit
+from scipy.spatial import cKDTree
 from scipy.stats import chi2, expon, gaussian_kde, kurtosis, skew
 from scipy.stats import mode as smode
 from statsmodels.tsa.stattools import pacf
@@ -1169,6 +1170,204 @@ def joint_non_gaussianity(y: ArrayLike, tau: Union[int, str] = 'ac', m: int = 2,
                        for k in range(-theiler_win, theiler_win + 1)
                        if abs(k) < n_skew)
         out['mardiaSkew'] = float((G3.sum() - band_sum) / n_off_band)
+
+    return out
+
+def _theiler_kth(idx: np.ndarray, dist: np.ndarray, k: int, theiler_win: int,
+                 ref_set: np.ndarray, query_set: np.ndarray) -> np.ndarray:
+    # For each query point i, returns the k-th nearest-neighbor distance within
+    # ref_set, excluding any candidate with |i-j| <= theiler_win. Falls back to an
+    # exact brute-force search for any point whose over-fetched candidate list
+    # (idx, dist) didn't leave at least k valid (non-excluded) neighbors.
+    n = idx.shape[0]
+    n_ref = ref_set.shape[0]
+    i = np.arange(n)[:, None]
+    valid = np.abs(idx - i) > theiler_win
+    # position of the k-th valid candidate in each (distance-sorted) row:
+    kth_hit = valid & (np.cumsum(valid, axis=1) == k)
+    has_kth = kth_hit.any(axis=1)
+    pos = np.argmax(kth_hit, axis=1)
+
+    kth = np.full(n, np.nan)
+    rows = np.flatnonzero(has_kth)
+    kth[rows] = dist[rows, pos[rows]]
+
+    time_idx = np.arange(n_ref)
+    for i in np.flatnonzero(~has_kth):
+        all_dists = np.sqrt(np.sum((ref_set - query_set[i]) ** 2, axis=1))
+        all_dists[np.abs(time_idx - i) <= theiler_win] = np.inf
+        kth_dist = np.partition(all_dists, k - 1)[k - 1]
+        if np.isfinite(kth_dist):
+            kth[i] = kth_dist
+
+    return kth
+
+
+def _knn_kld(A: np.ndarray, B: np.ndarray, k: int, theiler_win: int) -> float:
+    # Wang-Kulkarni-Verdu (2009) k-NN estimator of KL(P_A||P_B), where A and B
+    # are equal-sized (n x d) point sets whose rows are in temporal
+    # correspondence (row i of A and row i of B share the same underlying time
+    # index), so a Theiler window can be applied consistently in both searches.
+    n, d = A.shape
+    k_fetch = min(n - 1, k + 2 * theiler_win + 5)
+
+    # Self-search within A (for r_k(x_i), the k-th NN of x_i within A\{x_i}):
+    dist_a, idx_a = cKDTree(A).query(A, k=k_fetch + 1, workers=-1)
+    rk = _theiler_kth(idx_a, dist_a, k, theiler_win, A, A)
+
+    # Cross-search from A into B (for s_k(x_i), the k-th NN of x_i within B):
+    dist_b, idx_b = cKDTree(B).query(A, k=k_fetch, workers=-1)
+    sk = _theiler_kth(idx_b, dist_b, k, theiler_win, B, A)
+
+    good = np.isfinite(rk) & np.isfinite(sk) & (rk > 0) & (sk > 0)
+    n_good = int(np.count_nonzero(good))
+    if n_good < 0.5 * n:
+        return np.nan
+
[…]
