### 2026-08-22 c69f588
update wavelet_entropy
