+++ pyhctsa/operations/entropy.py
@@ -4,6 +4,7 @@ import logging
 logger = logging.getLogger('pyhctsa')
 
 import numpy as np
+import pywt
 from numpy.typing import ArrayLike
 from numba import njit
 from antropy.entropy import _xlogx
@@ -603,61 +604,91 @@ def complexity_invariant_distance(y: ArrayLike) -> dict:
         'CE2_norm': CE2 / min_CE2,
     }
 
-def wavelet_entropy(y: ArrayLike, whaten: str = 'shannon',
-                    p: Optional[float] = None) -> float:
+def _modwt(x: np.ndarray, wavelet_name: str, level: int) -> np.ndarray:
+    """Maximal-overlap discrete wavelet transform, with periodic boundary handling.
+
+    Returns a ``(level + 1, len(x))`` array whose rows are the wavelet
+    coefficients at scales ``1, ..., level`` followed by the scaling
+    (approximation) coefficients at scale ``level``.
     """
-    Wavelet-toolbox entropy of a time series.
+    w = pywt.Wavelet(wavelet_name)
+    # MODWT filters are the DWT filters scaled by 1/sqrt(2), applied without
+    # downsampling (the filters are reversed for a convolution, which leaves
+    # each scale's energy unchanged):
+    h = np.asarray(w.dec_hi[::-1], dtype=float) / np.sqrt(2)
+    g = np.asarray(w.dec_lo[::-1], dtype=float) / np.sqrt(2)
 
-    For a series :math:`y` of length :math:`N` the four entropy types are
+    v = np.asarray(x, dtype=float)
+    coeffs = np.empty((level + 1, v.size))
+    for j in range(level):
+        up = 1 << j # the filters are upsampled by 2^(j-1) at scale j
+        wj = np.zeros(v.size)
+        vj = np.zeros(v.size)
+        for ell, (hl, gl) in enumerate(zip(h, g)):
+            shifted = np.roll(v, up * ell)
+            wj += hl * shifted
+            vj += gl * shifted
+        coeffs[j] = wj
+        v = vj
+    coeffs[level] = v
 
-    - ``'shannon'``: :math:`-\\frac{1}{N}\\sum_i y_i^2 \\log(y_i^2)`,
-      using the convention :math:`0\\log 0 = 0`.
-    - ``'logenergy'``: :math:`\\frac{1}{N}\\sum_i \\log(y_i^2)`,
-      using the convention :math:`\\log 0 = 0`.
-    - ``'threshold'``: :math:`\\frac{1}{N}\\#\\{i : |y_i| > p\\}`.
-    - ``'sure'``: Stein's Unbiased Risk Estimate,
-      :math:`\\frac{1}{N}\\left(N - 2\\#\\{i : |y_i| \\leq p\\} +
-      \\sum_i \\min(y_i^2, p^2)\\right)`.
+    return coeffs
+
+
+def wavelet_entropy(y: ArrayLike, wavelet_name: str = 'sym4',
+                    level: int = 5) -> float:
+    """
+    Wavelet entropy of a time series.
+
+    Decomposes y via the maximal-overlap discrete wavelet transform (MODWT)
+    into a set of scales, computes each scale's share of the signal's total
+    energy, p_j = E_j / sum(E), and returns the Shannon entropy of this
+    relative-energy distribution across scales, normalized to [0,1] by its
+    maximum possible value, log2(numLevels).
+
+    References
+    ----------
+    .. [1] O.A. Rosso, S. Blanco, J. Yordanova, V. Kolev, A. Figliola, M.
+           Schuermann, E. Basar, "Wavelet entropy: a new tool for analysis of
+           short duration brain electrical signals", J. Neurosci. Methods
+           105(1) 65 (2001).
 
     Parameters
     ----------
     y : array-like
-        One-dimensional time series input.
[…]
