+++ pyhctsa/operations/nonlinearity.py
@@ -8,7 +8,9 @@ logger = logging.getLogger('pyhctsa')
 from sklearn.decomposition import PCA
 from sklearn.neighbors import NearestNeighbors
 
-from ..operations.model_fit import residual_analysis, _ml_rng
+from scipy.spatial.distance import cdist
+
+from ..operations.model_fit import residual_analysis, _ml_rng, _ml_randperm_k
 from ..operations.correlation import first_crossing, autocorr, _resolve_time_delay
 from ..operations.information import first_min
 from ..operations.scaling import _round
@@ -1222,5 +1224,216 @@ def delay_time(y: ArrayLike, max_delay: Union[int, float] = 0.2, past: int = 1,
     out['stdtau'] = np.std(tau, ddof=1)
     out['mintau'] = np.min(tau)
     out['maxtau'] = np.max(tau)
+
+    return out
+
+
+_EVT_POLE_BLOCK = 64  # poles per distance-matrix block
+
+def _evt_quantile(x: np.ndarray, p: float) -> float:
+    """The `p`-quantile of `x`, from the two order statistics it interpolates."""
+    n = x.size
+    r = (100.0*p/100.0)*n
+    k = np.floor(r + 0.5)
+    kp1 = k + 1
+    r = r - k
+
+    k = 1 if (k < 1 or np.isnan(k)) else int(k)
+    kp1 = min(int(kp1), n)
+    xs = np.partition(x, (k - 1, kp1 - 1))
+    xk, xkp1 = xs[k - 1], xs[kp1 - 1]
+
+    if xk == xkp1 or r == -0.5:
+        return xk
+
+    return (0.5 + r)*xkp1 + (0.5 - r)*xk
+
+def _omitnan_mean(x: np.ndarray) -> float:
+    x = x[~np.isnan(x)]
+
+    return float(np.mean(x)) if x.size > 0 else np.nan
+
+def _omitnan_std(x: np.ndarray) -> float:
+    x = x[~np.isnan(x)]
+    if x.size == 0:
+        return np.nan
+
+    return float(np.std(x, ddof=1)) if x.size > 1 else 0.0
+
+def evt_local_dim(y: ArrayLike, tau: Union[int, str] = 'ac', m: int = 3,
+                  q: float = 0.98, theiler_win: Union[int, float] = 0.01,
+                  n_poles: int = 200, m_order: int = 5, max_n: int = 10000,
+                  random_seed: Union[int, None] = 0) -> dict:
+    """
+    Extreme value theory local dimension and persistence of the attractor.
+
+    Time-delay embeds the series and, for a sample of reference points
+    ("poles") on the reconstructed orbit, treats close returns of the orbit
+    to each pole as extreme events: g_i = -log(||Y_i - pole||) is large
+    exactly when the orbit passes close to the pole. Extreme value theory
+    applied to this observable gives two local quantities per pole
+    (Freitas, Freitas & Todd 2010; Lucarini et al., "Extremes and
+    Recurrence in Dynamical Systems", 2016; Faranda, Messori & Yiou 2017):
+
+      - a local dimension d(pole): under the Freitas-Freitas-Todd theorem,
+        the Gumbel-law scale parameter of the extreme value law for g_i
+        equals the local dimension of the attractor at that pole exactly
+        (Faranda, Freitas, Guiraud & Vaienti 2014, Props 1-2). In practice
+        this is estimated by a peaks-over-threshold fit: take the
+        exceedances of g_i above a high quantile q, and set
+        d(pole) = 1/mean(exceedances) (the reciprocal of the exponential
+        MLE scale, i.e. the GPD fit with shape fixed at its ansatz-implied
+        value of 0 -- appropriate here because g_i = -log(distance) is
[…]
