### 2026-08-23 5ec5741
improve distribute code and add SSA method
