+++ pyhctsa/distribute.py
@@ -1,100 +1,271 @@
+"""Parallel execution backends for :class:`~pyhctsa.calculator.FeatureCalculator`."""
+import logging
+import os
 from abc import ABC, abstractmethod
-from typing import Union
-import multiprocessing as mp
-import numpy as np
-from functools import partial
-from rich.progress import Progress, TextColumn, BarColumn, TaskProgressColumn, TimeRemainingColumn, MofNCompleteColumn
-from pathos.helpers import mp as pathos_mp
+from contextlib import contextmanager
+from typing import Any, Callable, Union
 
-def _extract_features_single_series(ts, feature_funcs):
+import dill
+import numpy as np
+from pathos.helpers import mp as _mp
+from rich.progress import (BarColumn, MofNCompleteColumn, Progress,
+                           TaskProgressColumn, TextColumn, TimeRemainingColumn)
+
+logger = logging.getLogger('pyhctsa')
+
+# Environment variables read by the threading layers of the BLAS/OpenMP
+# backends that NumPy, SciPy and friends link against. They are consulted when
+# the backend is *loaded*, so they must be set before the worker interpreter
+# imports NumPy -- see ``_thread_env``.
+_BLAS_THREAD_VARS = (
+    "OMP_NUM_THREADS",
+    "OPENBLAS_NUM_THREADS",
+    "MKL_NUM_THREADS",
+    "NUMEXPR_NUM_THREADS",
+    "VECLIB_MAXIMUM_THREADS",
+)
+
+# 1-D array outputs no longer than this are spread across one column per
+# element; anything larger is left as a single array-valued cell.
+_MAX_INLINE_ARRAY = 16
+
+
+def _flatten_into(values: dict, name: str, val: Any) -> None:
+    """Write a single feature output into ``values`` under one or more keys."""
+    if isinstance(val, dict):
+        for k, v in val.items():
+            values[f"{name}.{k}"] = v
+    elif isinstance(val, np.ndarray) and val.ndim == 1 and val.size <= _MAX_INLINE_ARRAY:
+        for i, v in enumerate(val):
+            values[f"{name}_{i}"] = v
+    else:
+        values[name] = val
+
+
+def _extract_features_single_series(ts, feature_funcs) -> tuple[dict, dict]:
     """
     Worker task.
+
+    Returns a ``(values, errors)`` pair. A feature that raises contributes
+    nothing to ``values`` and one entry to ``errors``, keyed by feature label.
+    Keeping failures out of ``values`` is what makes the resulting column set
+    depend only on the configuration, not on which series happened to fail.
     """
-    row = {}
+    values: dict = {}
+    errors: dict = {}
     for name, func in feature_funcs.items():
         try:
             val = func(ts)
-            if isinstance(val, dict):
-                for k, v in val.items():
-                    row[f"{name}.{k}"] = v
-            elif isinstance(val, np.ndarray) and val.ndim == 1 and val.size <= 16:
-                for i, v in enumerate(val):
-                    row[f"{name}_{i}"] = v
-            else:
-                row[name] = val
         except Exception as e:
-            row[name] = f"Error: {e}"
-    return row
+            errors[name] = f"{type(e).__name__}: {e}"
+            # the traceback only exists here, in the process that raised
+            logger.debug("Feature '%s' failed", name, exc_info=True)
+            continue
+        _flatten_into(values, name, val)
+    return values, errors
 
-def _compute_features_for_chunk(chunk, feature_funcs):
-    return [_extract_features_single_series(ts, feature_funcs) for ts in chunk]
+
+@contextmanager
+def _thread_env(n_threads: Union[int, None]):
+    """Temporarily pin the BLAS/OpenMP thread-count environment variables.
+
[…]
