+++ pyhctsa/operations/nonlinearity.py
@@ -9,6 +9,7 @@ from sklearn.decomposition import PCA
 from sklearn.neighbors import NearestNeighbors
 
 from scipy.spatial.distance import cdist
+from scipy.stats import spearmanr
 
 from ..operations.model_fit import residual_analysis, _ml_rng, _ml_randperm_k
 from ..operations.correlation import first_crossing, autocorr, _resolve_time_delay
@@ -321,6 +322,146 @@ def embed_pca(y: ArrayLike, tau: Union[str, int] = 'ac', m: int = 3) -> dict:
     out['fb02'] = _first_fn(perc, 0.2, 'under')
     out['fb01'] = _first_fn(perc, 0.1, 'under')
     out['fb001'] = _first_fn(perc, 0.01, 'under')
+
+    return out
+
+def _ssa_diagonal_average(U: np.ndarray, sigma: np.ndarray, V: np.ndarray,
+                          components, counts: np.ndarray) -> np.ndarray:
+    """
+    Reconstructs a length-N component series from an elementary matrix
+    by averaging over its anti-diagonals ("Hankelization").
+    """
+    # The anti-diagonal sums of the rank-one matrix sigma_i*u_i*v_i' are the
+    # (linear) convolution of u_i with v_i, so the elementary matrices never
+    # have to be formed:
+    diag_sums = np.zeros(len(counts))
+    for i in components:
+        diag_sums += sigma[i]*np.convolve(U[:, i], V[:, i])
+
+    return diag_sums/counts
+
+def ssa(y: ArrayLike, L: Union[int, None] = None) -> dict:
+    """
+    Singular Spectrum Analysis of a time series.
+
+    Constructs the trajectory (Hankel) matrix of the time series using a
+    window length L (i.e., a time-delay embedding with delay tau = 1), and
+    performs an uncentered singular value decomposition of the result.
+
+    Unlike ``embed_pca`` (which centers the embedded data before decomposing it,
+    and allows a general embedding delay), this implements classic "Basic SSA":
+    a fixed delay of 1, no centering (so that a genuine trend is not removed
+    before decomposition), and diagonal averaging ("Hankelization") of the
+    leading elementary matrices back into component time series. Statistics
+    are computed on the singular-value pairing structure, and on the
+    reconstructed leading trend/oscillatory components themselves, rather than
+    on the raw eigenvalue spectrum (which ``embed_pca`` already covers).
+
+    References
+    ----------
+    .. [1] "Extracting qualitative dynamics from experimental data"
+        D. S. Broomhead and G. P. King, Physica D 20(2-3) 217 (1986)
+    .. [2] "Analysis of Time Series Structure: SSA and Related Techniques"
+        N. Golyandina, V. Nekrutkin, A. Zhigljavsky, Chapman & Hall/CRC (2001)
+
+    Parameters
+    ----------
+    y : array-like
+        The input time series.
+    L : int, optional
+        The window length (default: floor(N/4)). Must satisfy
+        ``4 <= L <= floor(N/2)``.
+
+    Returns
+    -------
+    dict
+        Statistics on the singular-value pairing/decay structure (gap1-gap5,
+        sepIdx), and on the reconstructed leading block of components as a
+        whole (trend strength trend_r2/trend_rho, dominant period of its most
+        tightly-paired internal mode pairperiod, and w-correlation-based
+        separability from the residual wcorr_leadresid). Individual
+        eigentriples within a near-degenerate pair are not uniquely determined
[…]
