### 2026-08-24 a6efac0
add spectral features and higuchi_fd
