+++ pyhctsa/operations/nonlinearity.py
@@ -10,6 +10,7 @@ from sklearn.neighbors import NearestNeighbors
 
 from scipy.spatial.distance import cdist
 from scipy.stats import spearmanr
+from scipy.signal import correlate
 
 from ..operations.model_fit import residual_analysis, _ml_rng, _ml_randperm_k
 from ..operations.correlation import first_crossing, autocorr, _resolve_time_delay
@@ -17,6 +18,198 @@ from ..operations.information import first_min
 from ..operations.scaling import _round
 from ..toolboxes.Tisean_3_0_1 import tisean as _tisean
 from ..utils import _first_index_past_threshold, matlab_quantile, time_delay_embed
+
+def zero_one_test(y, num_c=20, max_n=10000):
+    """Modified 0-1 test for chaos.
+
+    Parameters
+    ----------
+    y : array_like
+        Input scalar time series.
+    num_c : int, default=20
+        Number of frequencies c over which the statistics are computed.
+    max_n : int or "full", default=10000
+        Maximum number of samples to analyze. If "full", use the complete
+        time series.
+
+    Returns
+    -------
+    out : dict
+        K : float
+            Median correlation statistic across frequencies.
+        Kstd : float
+            Standard deviation of K across frequencies.
+        D : float
+            Median diffusion rate across frequencies.
+        Dstd : float
+            Standard deviation of D across frequencies.
+
+    Notes
+    -----
+    K ~ 0 indicates bounded/regular dynamics, whereas K ~ 1 indicates
+    unbounded dynamics. Unbounded dynamics may be deterministic chaos
+    or stochastic noise; the 0-1 test alone does not distinguish them.
+    """
+
+    y = np.asarray(y, dtype=float).reshape(-1)
+    n = y.size
+
+    # ------------------------------------------------------------------
+    # Input handling
+    # ------------------------------------------------------------------
+    if isinstance(max_n, str):
+        if max_n != "full":
+            raise ValueError("maxN must be an integer or 'full'.")
+
+        if n > 50_000:
+            warnings.warn(
+                f"Time series ({n} samples) exceeds 50000 with "
+                "maxN='full'; computation may be slow.",
+                RuntimeWarning,
+            )
+
+    else:
+        max_n = int(max_n)
+
+        if n > max_n:
+            warnings.warn(
+                f"Time series ({n} samples) exceeds maxN={max_n}; "
+                f"analyzing the first {max_n} samples.",
+                RuntimeWarning,
+            )
+            y = y[:max_n]
+            n = max_n
+
+    if n < 200:
+        warnings.warn(
+            f"Time series (N={n}) too short for a meaningful "
+            "0-1 test (need >= 200).",
+            RuntimeWarning,
+        )
+
+        return {
+            "K": np.nan,
+            "Kstd": np.nan,
+            "D": np.nan,
+            "Dstd": np.nan,
+        }
+
+    # ------------------------------------------------------------------
+    # Set up test
+    # ------------------------------------------------------------------
+    tcut = n // 10
+
+    # Number of displacement pairs used for every lag.
+    n_pairs = n - tcut
+
+    # j is a mathematical index in cos(j*c), not a Python array index.
+    j = np.arange(1, n + 1, dtype=float)
+
+    lags_idx = np.arange(1, tcut + 1)
+    lags = lags_idx.astype(float)
+
+    mean_y = np.mean(y)
+
+    cs = np.linspace(
+        np.pi / 5,
+        4 * np.pi / 5,
+        num_c,
+    )
+
[…]
