### 2026-08-24 02de473
add ramping window
