+++ pyhctsa/operations/stationarity.py
@@ -1,12 +1,13 @@
 import logging
 logger = logging.getLogger('pyhctsa')
 import warnings
+from itertools import permutations
 from typing import Union
 
 import numpy as np
 from numpy.typing import ArrayLike
 from scipy.signal import detrend
-from scipy.stats import gaussian_kde, kurtosis, skew
+from scipy.stats import gaussian_kde, kurtosis, norm, pearsonr, rankdata, skew
 from statsmodels.tools.sm_exceptions import InterpolationWarning
 from statsmodels.tsa.stattools import kpss
 
@@ -1203,3 +1204,208 @@ def _get_window(step_ind, inc, win_length):
     end_idx = (step_ind) * inc + win_length
     
     return np.arange(start_idx, end_idx).astype(int)
+
+def _kendall_tie_adj(r: np.ndarray) -> np.ndarray:
+    # Tie adjustments accompanying the midranks, as returned by MATLAB's tiedrank:
+    # [sum(t*(t-1))/2, sum(t*(t-1)*(t-2)), sum(t*(t-1)*(2*t+5))] over tied-group sizes t
+    _, counts = np.unique(r, return_counts=True)
+    t = counts[counts > 1].astype(float)
+    return np.array([np.sum(t * (t - 1)) / 2,
+                     np.sum(t * (t - 1) * (t - 2)),
+                     np.sum(t * (t - 1) * (2 * t + 5))])
+
+def _kendall(x: np.ndarray, y: np.ndarray) -> tuple:
+    # Kendall's tau-b and its two-tailed p-value, following MATLAB's corr(...,'type','Kendall'):
+    # the p-value is exact (permutation distribution of K) for small samples and a
+    # continuity-corrected normal approximation otherwise
+    n = len(x)
+    xrank, yrank = rankdata(x), rankdata(y)
+    xadj, yadj = _kendall_tie_adj(xrank), _kendall_tie_adj(yrank)
+    n2const = n * (n - 1) // 2
+
+    K = int(round(np.sum(np.sign(xrank[:, None] - xrank[None, :])
+                         * np.sign(yrank[:, None] - yrank[None, :])) / 2))
+
+    denom = np.sqrt((n2const - xadj[0]) * (n2const - yadj[0]))
+    tau = K / denom if denom > 0 else np.nan
+
+    ties = (xadj[0] > 0) or (yadj[0] > 0)
+    if (xadj[0] == n2const) or (yadj[0] == n2const):
+        return tau, np.nan
+
+    exact = (n < 10) if ties else (n < 50)
+    if exact:
+        nfact = 1.0
+        for i in range(2, n + 1):
+            nfact *= i
+        if ties:
+            # With ties, take permutations of the midranks
+            yperms = np.array(list(permutations(yrank)))
+            kperm = np.zeros(yperms.shape[0])
+            for w in range(n - 1):
+                U = np.sign(xrank[w] - xrank[w+1:])
+                V = np.sign(yperms[:, [w]] - yperms[:, w+1:])
+                kperm += V @ U
+            freq = np.bincount(np.rint(kperm).astype(int) + n2const,
+                               minlength=2*n2const+1).astype(float)[:-1]
+        else:
+            # No ties, use recursion to get the cumulative distribution of the number, C,
+            # of positive (xi-xj)*(yi-yj), i<j. K = #pos-#neg = C-Q, and C+Q = n(n-1)/2
+            freq = np.array([1.0, 1.0])
+            for i in range(3, n + 1):
+                freq = np.convolve(freq, np.ones(i))
+            interleaved = np.zeros(2 * freq.size)
+            interleaved[::2] = freq  # bins at integers, starting at -n2const
+            freq = interleaved[:-1]
[…]
