### 2026-08-26 8b366fe
add modwtvar
