+++ pyhctsa/operations/wavelet.py
@@ -515,6 +515,199 @@ def wl_coeffs(y: ArrayLike, w_name: str = 'db3', level: Union[int, str] = 3) -> 
 
     return out
 
+def _modwt(x: ArrayLike, w_name: str, level: int) -> np.ndarray:
+    """
+    Maximal overlap discrete wavelet transform.
+
+    Parameters
+    ----------
+    x : array-like
+        The input time series.
+    w_name : str
+        The name of the orthogonal mother wavelet (e.g., ``'db3'``, ``'sym2'``).
+    level : int
+        The level of the transform.
+
+    Returns
+    -------
+    numpy.ndarray
+        A ``(level+1, len(x))`` array whose first ``level`` rows are the wavelet
+        (detail) coefficients by level and whose final row holds the level-``level``
+        scaling coefficients.
+    """
+    x = np.asarray(x, dtype=float).ravel()
+    siglen = x.size
+    wavelet = pywt.Wavelet(w_name)
+
+    # Scale the scaling and wavelet filters for the MODWT
+    Lo = np.asarray(wavelet.rec_lo, dtype=float)/np.sqrt(2)
+    Hi = np.asarray(wavelet.rec_hi, dtype=float)/np.sqrt(2)
+
+    # If the signal length is less than the filter length, need to
+    # periodize the signal in order to use the DFT algorithm
+    if siglen < Lo.size:
+        x = np.tile(x, 1 + (Lo.size - siglen))
+    Nrep = x.size
+
+    # Obtain the DFT of the filters
+    G = np.fft.fft(Lo, Nrep)
+    H = np.fft.fft(Hi, Nrep)
+
+    # Obtain the DFT of the data
+    Vhat = np.fft.fft(x)
+    w = np.zeros((level+1, Nrep))
+    idx0 = np.arange(Nrep)
+
+    # Main MODWT algorithm
+    for jj in range(1, level+1):
+        upfactor = 2**(jj-1)
+        # Dilated filters modulo N
+        idx = (upfactor*idx0) % Nrep
+        What = H[idx]*Vhat
+        Vhat = G[idx]*Vhat
+        w[jj-1] = np.fft.ifft(What).real
+    w[level] = np.fft.ifft(Vhat).real
+
+    # Truncate data to length of boundary condition
+    return w[:, :siglen]
+
+
+def _modwt_var(w: np.ndarray, w_name: str) -> np.ndarray:
+    """
+    Unbiased multiscale variance estimates from a MODWT.
+
+    Parameters
+    ----------
+    w : numpy.ndarray
+        MODWT coefficients, as returned by :func:`_modwt`.
+    w_name : str
+        The name of the orthogonal mother wavelet used for the transform.
+
+    Returns
+    -------
+    numpy.ndarray
+        The unbiased wavelet variance by scale, computed from the boundary-unaffected
+        coefficients only. Levels with no such coefficients are omitted, and the
+        scaling variance is appended only when every requested level is retained.
+    """
+    L = pywt.Wavelet(w_name).dec_len
+
+    # Get the level -- the final row of w are the scaling coefficients
+    level = w.shape[0] - 1
+
+    # Extract scaling coefficients
+    VJ = w[-1]
+
+    # Extract wavelet coefficients
+    w = w[:-1]
+    N = w.shape[1]
+
+    # For an unbiased estimate
+    Jmax = int(np.floor(np.log2((N-1)/(L-1) + 1)))
+    if Jmax < 1:
+        return np.array([])
+    Jmax = min(Jmax, level)
+    w = w[:Jmax]
+
+    # Determine if we use the scaling coefficients
+    scalingvar = (Jmax == level)
+
[…]
