### 2026-08-26 e614fe8
add modwtvar
