+++ pyhctsa/operations/stationarity.py
@@ -1409,3 +1409,164 @@ def ramping_windows(y: ArrayLike, num_seg: int = 10) -> dict:
     out['asymac1_pearson_r'], out['asymac1_pearson_p'] = _pearson(segIdx, segAsymAC1)
 
     return out
+
+def slow_feature_analysis(y: ArrayLike, num_windows: int = 20) -> dict:
+    """
+    Slow feature analysis of windowed statistics.
+
+    Splits the time series into ``num_windows`` non-overlapping segments (same
+    segmentation and per-segment statistics as :func:`ramping_windows`: mean,
+    variance, skewness, and lag-1 autocorrelation, forming a ``num_windows`` x 4
+    matrix), then applies Slow Feature Analysis (SFA) to find the linear combination
+    of these four statistics that varies as *slowly* as possible across the sequence
+    of windows -- i.e., minimizes the variance of its own increments, subject to unit
+    variance. This is a fundamentally different criterion to :func:`ramping_windows`'
+    per-statistic monotonic-trend tests: SFA is multivariate (can find a slow
+    combination spread across mean/variance/skewness/AC1 that no single one of them
+    shows individually) and detects any slow, low-frequency evolution, not just a
+    monotonic ramp (e.g., a slow rise-then-fall hump across the series, invisible to
+    Kendall's tau/Pearson's r, still registers as "slow" here). Validated on
+    Empirical1000 to be essentially uncorrelated (max \\|r\\| ~ 0.14) with all
+    ``ramping_windows`` (numSeg = 10) trend fields.
+
+    The companion comparison to ordinary PCA addresses a different question: PCA
+    finds the combination of the four statistics with *maximal variance* across
+    windows, which need not be the slowest one -- a large-amplitude but choppy/noisy
+    statistic can dominate variance while a small-amplitude but smooth, genuinely
+    slow drift is buried underneath it and missed by PCA. ``pc1VarFrac`` and
+    ``slowPCA1corr`` quantify this: how concentrated the variance is in a single PCA
+    direction, and whether the slow direction found by SFA coincides with that
+    high-variance PCA direction (high overlap) or is a separate, lower-variance mode
+    that ordinary variance-based analysis would miss (low overlap).
+
+    Let ``z(t)`` be the ``num_windows`` x 4 matrix of [mean, variance, skewness, AC1]
+    computed per window, whitened (centered, then linearly transformed to unit
+    covariance). SFA finds the orthogonal directions ``u_i`` that minimize
+    ``var(diff(z*u_i))``, i.e. the "slowness" eigenvalues ``eta_i = var(diff(z*u_i))``
+    of the covariance of the whitened derivative signal (ascending: ``eta_1`` is the
+    slowest direction). For reference, i.i.d. (white-noise) windows give ``eta ~ 2``
+    on average; substantially smaller values indicate genuinely slow (smooth,
+    low-frequency) structure in some combination of the four per-window statistics.
+
+    Parameters
+    ----------
+    y : array-like
+        The input time series.
+    num_windows : int, optional
+        The number of non-overlapping segments to divide the time series into.
+        Non-overlapping segments are used for the same reason as
+        :func:`ramping_windows`: overlap would induce artificial serial correlation
+        between adjacent window statistics, which would make the derivative-based
+        slowness measure spuriously small regardless of any real slow structure in
+        the data. 20 was chosen (rather than :func:`ramping_windows`' default of 10)
+        because SFA needs enough windows to estimate the underlying 4x4 covariance
[…]
