+++ docs/source/api.rst
@@ -284,6 +284,7 @@ Stationarity
     pyhctsa.operations.stationarity.stat_av
     pyhctsa.operations.stationarity.sliding_window
     pyhctsa.operations.stationarity.ramping_windows
+    pyhctsa.operations.stationarity.slow_feature_analysis
 
 .. _surrogatesmeths:
 
