+++ pyhctsa/operations/stationarity.py
@@ -14,6 +14,7 @@ from statsmodels.tsa.stattools import kpss
 from ..operations.correlation import autocorr, first_crossing
 from ..operations.distribution import moments
 from ..operations.entropy import approximate_entropy, distribution_entropy, sample_entropy
+from ..toolboxes.matlab.matlab_fit import polyfit, robustfit
 from ..toolboxes.matlab._pptest_tables import PP_CV_TABLES, PP_SAMP_SIZES, PP_SIG_LEVELS
 from ..utils import make_mat_buffer, sign_change, z_score
 
@@ -1568,5 +1569,129 @@ def slow_feature_analysis(y: ArrayLike, num_windows: int = 20) -> dict:
 
     out['pc1VarFrac'] = pcaEigs[0] / np.sum(pcaEigs)
     out['slowPCA1corr'] = abs(_pearson(slowScores[:, 0], pcScores[:, 0])[0])
+
+    return out
+
+def _is_zscored(x: np.ndarray) -> bool:
+    numericThreshold = 100 * np.finfo(float).eps
+    return (abs(np.mean(x)) < numericThreshold) and (abs(np.std(x, ddof=1) - 1) < numericThreshold)
+
+def _cum_sum_bridge_stats(p: np.ndarray) -> Union[dict, float]:
+    # CUSUM/bridge stationarity statistics on a cumulative sum.
+    #
+    # Given a series p_t whose mean is being tested for stationarity, forms
+    # cumsum(p) and computes: linear-fit statistics on the cumsum (cf. trend),
+    # a CUSUM 'bridge' relative to the endpoint-to-endpoint line (cf. Inclan-Tiao's
+    # test for a change point in variance), and a comparison between an ordinary
+    # least-squares and a robust (bisquare) linear fit to the cumsum -- large
+    # disagreement between the two indicates the OLS trend is either outlier-driven
+    # or genuinely curved (accelerating/decelerating drift) rather than a clean
+    # linear trend.
+    p = np.asarray(p, dtype=float).ravel()
+    Np = len(p)
+    if Np < 20:
+        return np.nan
+
+    t = np.arange(1, Np + 1, dtype=float)
+    yC = np.cumsum(p)
+
+    out = {}
+    out['meanYC'] = np.mean(yC)
+    coeffsOLS = polyfit(t, yC, 1)
+    out['gradient'] = coeffsOLS[0] # ~ std(yC) too (r > 0.99 empirically); kept as the interpretable one
+    out['intercept'] = coeffsOLS[1]
+    residOLS = yC - np.polyval(coeffsOLS, t)
+
+    out['meanYC12'] = np.mean(yC[:Np // 2])
+    out['meanYC22'] = np.mean(yC[Np // 2:])
+
+    bridge = yC - (t / Np) * yC[-1]
+    scaleFactor = np.std(p, ddof=1) * np.sqrt(Np)
+    if scaleFactor > 0:
+        out['maxBridge'] = np.max(np.abs(bridge)) / scaleFactor
+    else:
+        out['maxBridge'] = np.nan
+    idxMax = int(np.argmax(np.abs(bridge))) + 1
+    out['posMaxBridge'] = idxMax / Np # where the largest deviation from stationarity occurs
+    out['stdBridge'] = np.std(bridge, ddof=1)
+
+    robCoeffs, robStats = robustfit(t, yC)
+    robustGradient = robCoeffs[1] # not output directly: r > 0.98 with out['gradient']
+    robResid = yC - (robCoeffs[0] + robCoeffs[1] * t)
+    if robStats['se'][1] > 0:
[…]
