### 2026-08-27 e9a68e9
update stationarity
