+++ pyhctsa/operations/stationarity.py
@@ -1354,60 +1354,60 @@ def ramping_windows(y: ArrayLike, num_seg: int = 10) -> dict:
     N = len(y)
     num_seg = int(num_seg)
 
-    minNumSeg = 5 # need enough segments for a meaningful trend statistic
-    minSegLength = 20 # heuristic minimum for meaningful skewness/kurtosis/AC1 estimates
+    min_num_seg = 5 # need enough segments for a meaningful trend statistic
+    min_seg_length = 20 # heuristic minimum for meaningful skewness/kurtosis/AC1 estimates
 
-    if num_seg < minNumSeg:
+    if num_seg < min_num_seg:
         raise ValueError(f"num_seg = {num_seg} is too few segments for a meaningful "
-                         f"trend statistic (need >= {minNumSeg})")
+                         f"trend statistic (need >= {min_num_seg})")
 
-    segLength = N // num_seg
-    if segLength < minSegLength:
+    seg_length = N // num_seg
+    if seg_length < min_seg_length:
         logger.warning(f"Time series (N = {N}) too short for {num_seg} segments of a meaningful length")
         return np.nan
 
     # ------------------------------------------------------------------------------
     # Segment the time series (non-overlapping, discarding any remainder)
     # ------------------------------------------------------------------------------
-    z = y[:segLength * num_seg].reshape(num_seg, segLength) # num_seg x segLength
+    z = y[:seg_length * num_seg].reshape(num_seg, seg_length) # num_seg x seg_length
 
     # ------------------------------------------------------------------------------
     # Within-segment statistics
     # ------------------------------------------------------------------------------
-    segMean = np.mean(z, axis=1)
-    segVar = np.var(z, axis=1, ddof=1)
+    seg_mean = np.mean(z, axis=1)
+    seg_var = np.var(z, axis=1, ddof=1)
     # Standardized (not moments-style raw/std) skewness and kurtosis: keeps the
     # shape-trend signal from being conflated with the (separately tracked)
-    # scale-trend signal in segVar, since moments normalizes by std^1 regardless of
+    # scale-trend signal in seg_var, since moments normalizes by std^1 regardless of
     # moment order rather than std^3/std^4.
-    segSkew = skew(z, axis=1)
-    segKurt = kurtosis(z, axis=1, fisher=False)
-    segAC1 = np.zeros(num_seg)
-    segAsymAC1 = np.zeros(num_seg)
+    seg_skew = skew(z, axis=1)
+    seg_kurt = kurtosis(z, axis=1, fisher=False)
+    seg_ac1 = np.zeros(num_seg)
+    seg_asym_ac1 = np.zeros(num_seg)
     for i in range(num_seg):
-        segAC1[i] = autocorr(z[i, :], 1, 'Fourier')[0]
+        seg_ac1[i] = autocorr(z[i, :], 1, 'Fourier')[0]
         zseg = (z[i, :] - np.mean(z[i, :])) / np.std(z[i, :], ddof=1) # z-scored *within* this segment
-        segAsymAC1[i] = np.mean(zseg[:-1] * zseg[1:] * (zseg[1:] - zseg[:-1]))
+        seg_asym_ac1[i] = np.mean(zseg[:-1] * zseg[1:] * (zseg[1:] - zseg[:-1]))
 
     # ------------------------------------------------------------------------------
     # Kendall's tau and Pearson's r (each with p-value) against segment index
[…]
