+++ pyhctsa/operations/spectral.py
@@ -6,7 +6,7 @@ from numpy.typing import ArrayLike
 import scipy.fft
 import scipy.signal
 
-from ..toolboxes.matlab.matlab_fit import lsqcurvefit_trr, goodness_of_fit, robustfit
+from ..toolboxes.matlab.matlab_fit import lsqcurvefit_trr, goodness_of_fit, polyfit, robustfit
 
 from ..operations.correlation import autocorr, first_crossing
 from ..operations.distribution import moments
@@ -766,3 +766,120 @@ def _fit_gaussian(log_f: ArrayLike, resid: ArrayLike, i_pk: int,
         'width': w,
         'pred': h * np.exp(-(log_f - m) ** 2 / (2 * w ** 2)),
     }
+
+def cepstrum(y: ArrayLike, max_period: int = 100, min_period: int = 4) -> dict:
+    """
+    Cepstral statistics: harmonic (comb) structure of the power spectrum.
+
+    Computes the real cepstrum, the inverse Fourier transform of the log magnitude
+    spectrum, and summarizes the structure of its dominant peak.
+
+    Parameters
+    ----------
+    y : array-like
+        The input time series.
+    max_period : int, optional
+        The longest fundamental period (in samples) to search for. Default is
+        100.
+    min_period : int, optional
+        The shortest fundamental period (in samples) to search for. Default is 4.
+
+    Returns
+    -------
+    dict
+        - ``period``: the estimated fundamental period (quefrency of the dominant
+          cepstral peak), in samples.
+        - ``peak``: the height of that peak.
+        - ``meanCeps``, ``stdCeps``: the mean and standard deviation of the cepstrum
+          over the search range.
+        - ``peakRatio``: the peak height in units of the standard deviation of the
+          cepstrum over the search range.
+        - ``CPP``: the cepstral peak prominence, the standard robust measure, being
+          the peak height above a linear regression fit through the cepstrum across
+          the search range (this normalizes away the overall cepstral trend, so it
+          does not simply track the spectrum's dynamic range).
+        - ``rahmonicRatio``: comparing the cepstrum at twice the peak quefrency to
+          the peak itself (a genuine harmonic comb repeats at multiples of the
+          fundamental period, so a real periodicity shows a secondary 'rahmonic'
+          peak, whereas an isolated fluke does not).
+
+        Returns NaN if the time series is too short for the requested search range,
+        or is constant.
+    """
+    y = np.asarray(y, dtype=float).ravel()
+
+    max_period, min_period = int(max_period), int(min_period)
+    if min_period < 2:
+        raise ValueError(f"min_period = {min_period} is below the Nyquist limit "
+                         f"(a period needs >= 2 samples)")
+    if max_period <= min_period:
+        raise ValueError(f"max_period ({max_period}) must exceed min_period ({min_period})")
+
+    N = len(y)
+    minCycles = 4 # need several cycles of the longest period searched for a meaningful estimate
+    if N < minCycles * max_period:
+        warnings.warn(f"Time series (N = {N}) too short to search for periods up to "
+                      f"{max_period} samples (need >= {minCycles * max_period})")
+        return np.nan
+
+    if np.all(y == y[0]): # constant series has an all-zero spectrum -> log(0)
+        warnings.warn("Constant time series has no spectral (or cepstral) structure")
+        return np.nan
+
[…]
