### 2026-08-27 1ccac40
add phase amp coupling
